### 2025-12-01 #732 @cwy20030
ActiGlobe: Wearable Recording Processor for Time Shift Adjustment and Data Analysis
Submitting Author Name: C. William Yao
Submitting Author Github Handle: @cwy20030
Other Package Authors Github handles: (comma separated, delete if none) @GVARESCO, @FBSportScience, @gsimonelliCA
Repository:  https://github.com/cwy20030/ActiGlobe
Version submitted: 0.2.1
Submission type: Standard

Version accepted: TBD

---

```
Package: ActiGlobe
Title: Wearable Recording Processor for Time Shift Adjustment and Data Analysis
Version: 0.2.1
Authors@R: c(person("C. William", "Yao", email = "chun.william.yao.cnmtl@ssss.gouv.qc.ca", role = c("aut","cre","cph"), comment = c(ORCID="0000-0002-7234-7375")),
             person("Giorgio", "Varesco", email = "giorgio.varesco@umontreal.ca", role = c("aut","ctb"), comment = c(ORCID="0000-0001-9385-6972")),
			 person("François", "Bieuzen", email = "fbieuzen@insquebec.org", role = c("aut","ctb"), comment = c(ORCID="0000-0002-9690-9168")),
             person("Guido", "Simonelli", email = "guido.simonelli@umontreal.ca", role = c("aut","ctb","fnd"), comment = c(ORCID="0000-0002-5400-319X")))
Maintainer: C. William Yao <chun.william.yao.cnmtl@ssss.gouv.qc.ca>
Description: Cross-continental travel and local practices of daylight saving time can induce a time shift. When left unadjusted, it can lead to biases in interpreting wearable measures. This research tool streamlines data processing and analysis (e.g., cosinor analysis, circularized kernel density estimation) for longitudinal recordings. Together with additional features such as exporting daily wearable measures and creating visual reports, it aims to facilitate research on longitudinal wearable measurements. To learn its application, please refer to the provided vignettes for a brief tutorial. This tool is developed based on the methodology detailed in: Varesco, G., Yao, C. W., Dubé, E., Simonelli, G., Bieuzen, F., (2024) <doi:10.1113/ep092195>.
License: AGPL (>= 3)
    ggrepel,
    grDevices,
    gridExtra,
    hms,
    sandwich,
    viridis,
    parallel,
    scales
    ggthemes,
    zeallot
NeedsCompilation: no

```

ActiGlobe automates the data harmonization workflow for adjusting time shift during pre-processing (via `BriefSum` and `TAdjust`). It also supports parameter extraction for further rest-activity-based circadian analysis (e.g., `CosinorM`) and reproducible documentation (e.g., `write.act`) after segmenting the continuous recording by day (i.e., `Act2Daily`).

ActiGlobe aims to facilitate researchers studying rest-activity-based circadian rhythms and human behavioral pattern based on longitudinal wearable measurements.

ActiGlobe is a strong candidate for rOpenSci review because it delivers an end‑to‑end, reproducible actigraphy workflow that combines robust preprocessing (including explicit DST and travel handling), multiple analytical strategies (parametric and KDE cosinor, bootstrapped uncertainty), and consolidated, FAIR‑aware reporting (per‑subject PDFs plus CSV outputs). Compared with other R packages in the domain, ActiGlobe’s distinguishing value is the combination of (a) explicit, auditable time‑change correction, (b) circadian modeling tuned for irregular/real‑world recordings, and (c) built‑in consolidated reporting and metadata export that support FAIR principles and reproducible pipelines.

Package (Language)                | Pre processing | Adjust time change     | Circadian analyses | Consolidated report | FAIR support | Notes
--------------------------------- | -------------- | ---------------------- | ------------------ | ------------------- | ------------ | ----------------------------------------------------------------------------------------------
ActiGlobe (R)                     | ✓              | ✓                      | ✓                  | ✓                   | Full         | Produces per subject PDFs and CSV/HTML outputs; anonymization and reproducibility align with FAIR principles.
acc (R)                           | ✓              | ✗                      | ✓                  | ✗                   | Partial      | CRAN package for accelerometer data: preprocessing, visualization, and analysis pipelines.
accelerometry (R)                 | ✓              | ✗                      | ✓                  | ✗                   | Partial      | Non wear/bout detection outputs; structured but metadata incomplete.
ActFrag (R)                       | ✓              | ✗                      | ✓ (fragmentation)  | ✗                   | Partial      | Fragmentation metrics; structured tables but limited metadata.
Actigraphy (R)                    | ✓              | ✗                      | ✓                  | ✗                   | Partial      | FDA/circadian modeling; structured outputs but metadata schema unclear.
actigraph.sleepr (R)              | ✓              | ✗                      | ✓                  | ✗                   | Partial      | Tidy sleep scoring tables; lacks consolidated reporting.
actverse (R)                      | ✓              | ✗/✓ (tz parse)         | ✓                  | ✗                   | Partial      | Tidyverse pipelines; partial tz handling; no consolidated reports.
agcounts (R)                      | ✓              | ✗                      | ✗/✓ (counts)       | ✗                   | Partial      | ActiGraph style counts; standardized but metadata minimal.
CircaCP (R)                       | ✓              | ✗                      | ✓                  | ✗                   | Partial      | Cosinor/change point outputs; tabular FAIR structure, but limited metadata.
GGIR (R)                          | ✓              | ✗/✓ (drift/logs)       | ✓                  | ✓                   | Full         | g.report produces HTML/CSV with metadata; closest to full FAIR compliance.
nparACT (R)                       | ✓              | ✗                      | ✓                  | ✗                   | Partial      | Dedicated nonparametric circadian metrics; standardized tables. No embedded reporting.

Yes.
ActiGlobe includes built‑in anonymization options, minimal‑export defaults, and provenance logging to support ethical handling of human subjects data. The package documentation and vignettes demonstrate how to produce de‑identified, FAIR‑ready outputs and remind users to obtain appropriate consent and institutional approvals before sharing data.

rOpenSci | Software Peer Review
