### 2025-12-02 d55fdf7
rOpenSci Update
