+++ R/GuessTZ.R
@@ -24,93 +24,97 @@
 #' @param All Logical, if TRUE, as default, it will provide all possible TZ codes. If FALSE, it will retrieve the first one.
 #' @noRd
 
-GuessTZ <- function(aOF, DT = NULL, iTZ = NULL, All = TRUE) {
+GuessTZ <- function (aOF, DT = NULL, iTZ = NULL, All = TRUE) {
 
-  # Establish initial time zone...
-  TZ1 = ifelse(iTZ == "local", Sys.timezone(), iTZ)
+    # Establish initial time zone...
+    TZ1 <- ifelse (iTZ == "local", Sys.timezone (), iTZ)
 
-  if (is.null(iTZ)) TZ1 = NULL
-
-
-
-  # Extract all known time zones
-  oTZs <- OlsonNames()
-
-  ## Process DT
-  if (is.null(DT)) {
-
-    # Determine if DST exists using time offset on January 1st of 2021
-    DT <- as.POSIXct("2021-01-01", tz = "UTC")
-
-  }
-
-  if (!length(DT) == 1)  DT = DT[[1]]
+    if (is.null (iTZ)) TZ1 <- NULL
 
 
+    # Extract all known time zones
+    oTZs <- OlsonNames ()
 
-  # Extract
-  ## Check number of cores available...
-  NCore = parallel::detectCores()
+    ## Process DT
+    if (is.null (DT)) {
 
-  if (NCore > 6) {
-    # Step 1: Create a cluster
-    cl <- parallel::makeCluster(4)  # Use all but one core
+        # Determine if DST exists using time offset on January 1st of 2021
+        DT <- as.POSIXct ("2021-01-01", tz = "UTC")
 
-    # Step 2: Export variables and functions to cluster
-    parallel::clusterExport(cl, varlist = c("DT"), envir = environment())
+    }
 
-    # Step 3: Run the parallelized task
-    Toffs <- parallel::parLapply(cl, oTZs, function(tz) {
-      format(as.POSIXct(DT, tz = tz), "%z")
-    })
-
-    # Step 4: Clean up
-    parallel::stopCluster(cl)
-
-    # Optional: Convert result to character vector (like vapply would produce)
-    Toffs <- unlist(Toffs)
-
-  } else {
-
-    Toffs <- vapply(oTZs,
-                    function(tz) format(as.POSIXct(DT, tz = tz), "%z"),
-                    character(1))
-
-  }
+    if (!length (DT) == 1) DT <- DT [[1]]
 
 
-  #### Step 1 Guess all possible TZ indicators
-  pTZs <- sapply(aOF,
-                 function(x) oTZs[Toffs %in% x])
+    # Extract
+    ## Check number of cores available...
+    NCore <- parallel::detectCores ()
 
-  #### Step 2 Check if the initial time zone is included
-  if (!is.null(TZ1))
-    if (length(aOF) == 1) {
+    if (NCore > 6) {
+        # Step 1: Create a cluster
+        cl <- parallel::makeCluster (4) # Use all but one core
 
-      pTZs <-ifelse(TZ1 %in% pTZs, TZ1, pTZs)
+        # Step 2: Export variables and functions to cluster
+        parallel::clusterExport (cl, varlist = c ("DT"), envir = environment ())
+
+        # Step 3: Run the parallelized task
+        Toffs <- parallel::parLapply (cl, oTZs, function (tz) {
+            format (as.POSIXct (DT, tz = tz), "%z")
+        })
+
+        # Step 4: Clean up
+        parallel::stopCluster (cl)
+
+        # Optional: Convert result to character vector (like vapply would produce)
+        Toffs <- unlist (Toffs)
 
     } else {
 
-      pTZs <- sapply(pTZs,
[…]
