+++ tests/testthat/test-R2P.R
@@ -1,3 +1,3 @@
-test_that("skip R2P", {
-  skip('skip')
+test_that ("skip R2P", {
+    skip ("skip")
 })
