+++ tests/testthat/test-TAdjust.R
@@ -1,3 +1,3 @@
-test_that("skip TAdjust", {
-  skip('skip')
+test_that ("skip TAdjust", {
+    skip ("skip")
 })
