+++ tests/testthat/test-Num2UTC.R
@@ -1,3 +1,3 @@
-test_that("skip Num2UTC", {
-  skip('skip')
+test_that ("skip Num2UTC", {
+    skip ("skip")
 })
