+++ tests/testthat/test-CosinorM.KDE.R
@@ -1,3 +1,3 @@
-test_that("skip CosinorM.KDE", {
-  skip('skip')
+test_that ("skip CosinorM.KDE", {
+    skip ("skip")
 })
