+++ R/ggCosinorM.R
@@ -34,32 +34,40 @@
 #' @param legend.position Position of the legend on the plot; default is `"right"`. Other options include `"top"`, `"bottom"`, `"left"`, or a numeric vector of length two specifying x and y coordinates.
 #' @param ... Additional arguments (currently ignored) kept for future update
 #'
+#' @return A `ggplot` object representing the cosinor model fit visualization.
+#'
 #' @examples
 #' \dontrun{
 #' # Import data
-#' data(FlyEast)
+#' data (FlyEast)
 #'
-#' BdfList =
-#' BriefSum(df = FlyEast ,
-#'          SR = 1/60,
-#'          Start = "2017-10-24 13:45:00")
+#' BdfList <-
+#'     BriefSum (
+#'         df = FlyEast,
+#'         SR = 1 / 60,
+#'         Start = "2017-10-24 13:45:00"
+#'     )
 #'
 #' # Let's extract actigraphy data from a single day
 #' df <- BdfList$df
-#' df <- subset(df, df$Date == "2017-10-28")
+#' df <- subset (df, df$Date == "2017-10-28")
 #'
-#' fit <- CosinorM(time = df$Time,
-#'                 activity = df$Activity,
-#'                 tau = 24,
-#'                 method = "OLS")
+#' fit <- CosinorM (
+#'     time = df$Time,
+#'     activity = df$Activity,
+#'     tau = 24,
+#'     method = "OLS"
+#' )
 #'
 #'
-#' p <- ggCosinorM(object = fit,
-#'                 labels = TRUE,
-#'                 ci = TRUE,
-#'                 ci_level = 0.95,
-#'                 title_extra = "2017-10-24")
-#' print(p)
+#' p <- ggCosinorM (
+#'     object = fit,
+#'     labels = TRUE,
+#'     ci = TRUE,
+#'     ci_level = 0.95,
+#'     title_extra = "2017-10-24"
+#' )
+#' print (p)
 #' }
 #'
 #' @seealso
@@ -67,348 +75,368 @@
 #' @keywords plot ggplot graph cosinor
 #' @export
 
-ggCosinorM <- function(object, labels = TRUE, ci = TRUE, ci_level = 0.95,
-                       n = 400, point_size = 0.5, title_extra = NULL, legend.position = "right",
-                       ...) {
+ggCosinorM <- function (object, labels = TRUE, ci = TRUE, ci_level = 0.95,
+                        n = 400, point_size = 0.5, title_extra = NULL, legend.position = "right",
+                        ...) {
 
-  # Accept both parametric CosinorM and KDE-based CosinorM.KDE
-  if (!inherits(object, c("CosinorM", "CosinorM.KDE"))) {
-    stop("ggCosinorM: object must be a CosinorM or CosinorM.KDE fit.", call. = FALSE)
-  }
-
-
-  # Check Essential Parameters ----------------------
-  day <- 24
+    # Accept both parametric CosinorM and KDE-based CosinorM.KDE
+    if (!inherits (object, c ("CosinorM", "CosinorM.KDE"))) {
+        stop ("ggCosinorM: object must be a CosinorM or CosinorM.KDE fit.", call. = FALSE)
+    }
 
 
-  tau <- if (!is.null(object$tau)) object$tau else 24
-  nT  <- length(tau)
+    # Check Essential Parameters ----------------------
+    day <- 24
 
-  # Observed data --------------
-  if (!is.null(object$model) && !is.null(object$model$time)) {
-    t_obs <- object$model$time
-    y_obs <- if (!is.null(object$model$activity)) object$model$activity else
-      stats::model.response(stats::model.frame(object))
-  } else {
[…]
