### 2025-12-02 32331ad
Update issue templates
