+++ .github/ISSUE_TEMPLATE/new-idea.md
@@ -1,0 +1,40 @@
+---
+name: New Idea
+about: Propose and describe a new idea that may be implemented into ActiGlobe
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+# Proposal: [Title of Your Idea]
+
+## Summary
+Provide a concise overview of the idea in 2–3 sentences.  
+Explain what the idea is and why it matters for ActiGlobe.
+
+## Background
+- What problem or gap does this idea address?
+- Why is it relevant to ActiGlobe’s mission (e.g., reproducibility, accessibility, analytic transparency)?
+- Any prior work or related approaches?
+
+## Proposed Implementation
+- **Core Concept:** Describe the functionality or feature in detail.
+- **Technical Approach:** Outline how it could be implemented (e.g., new function, workflow, visualization).
+- **Dependencies:** Note any packages, datasets, or infrastructure required.
+
+## Benefits
+- How will this improve ActiGlobe for researchers and collaborators?
+- What measurable outcomes (e.g., faster benchmarking, clearer documentation, more robust pipelines) are expected?
+
+## Risks & Considerations
+- Potential challenges (e.g., performance, compatibility, maintenance).
+- How these risks might be mitigated.
+
+## Example Use Case
+Provide a short example (code snippet or workflow) showing how a user would interact with the new feature.
+
+```r
+# Example usage of the proposed feature
+ActiGlobe::NewFunction(data, option = TRUE)
+```
