### 2025-12-03 43e5eba
Amend for rOpenSource Submission
Augment documentation
Modify testsuit
