+++ R/globals.R
@@ -1,0 +1,2 @@
+
+utils::globalVariables(c("IANA"))
