+++ tests/testthat/test-ggActiGlobe.R
@@ -1,3 +1,3 @@
-test_that("skip ggActiGlobe", {
-  skip('skip')
+test_that ("skip ggActiGlobe", {
+    skip ("skip")
 })
