+++ .github/ISSUE_TEMPLATE/documentation-issue.md
@@ -1,0 +1,31 @@
+---
+name: Documentation Issue
+about: To report typo or mistake, or to offer suggestion on documentation
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+# Documentation Issue Report
+
+## Type of Issue
+- [ ] Typo
+- [ ] Incorrect information
+- [ ] Missing information
+- [ ] Suggestion for improvement
+
+## Location
+- File/Section: [e.g., README.md, Installation Guide]
+- Line/Paragraph: [optional, if known]
+
+## Suggested Correction or Improvement
+Explain how this should be corrected or improved.  
+(e.g., "Change `teh` to `the`", "Clarify installation steps for Windows users")
+
+## Impact (optional when reporting typo)
+Why does this matter?  
+(e.g., "Confusing for new users", "Could lead to incorrect usage")
+
+## Additional Context
+Add any other context, screenshots, or references that help explain the issue.
