### 2025-12-03 4c6c4e6
Update issue templates
