+++ R/CosinorM.R
@@ -67,12 +67,13 @@
 #'   \item if the model is not a mono-phase (i.e., more than one \eqn{\tau }), there will be \eqn{\frac{24 (hours)} {\tau} } number of peaks within a day.
 #' }
 #'
-#'
 #' \strong{Amplitude \eqn{(A) } estimation}:
 #' Amplitude is calculated from fitted sine and cosine coefficients as:
 #' \deqn{A= \sqrt {(\beta^2 + \gamma^2)}}
 #'
+#'
 #' @import stats sandwich
+#'
 #'
 #' @param time Numeric vector of time coordinates for each data point
 #' @param activity Numeric vector of activity counts from an actigraphy device
@@ -93,6 +94,7 @@
 #'   \item "FALSE": All essential parameters would be produced.  (default)
 #'   \item "TRUE": Only cosinor coefficients are returned. This is suited for post-hoc processes, such as computing confidence interval via nonparametric bootstrap
 #' }
+#'
 #'
 #' @returns
 #' A list of class c("CosinorM", "lm") containing:
@@ -116,6 +118,7 @@
 #' }
 #' Inherits all components from \code{\link[stats]{lm}}.
 #'
+#'
 #' @references
 #' Chambers, J. M. (1992) Linear models. Chapter 4 of Statistical Models in S eds J. M. Chambers and T. J. Hastie, Wadsworth & Brooks/Cole.
 #'
@@ -125,6 +128,7 @@
 #'
 #' @seealso
 #' \code{\link[stats]{lm}}  \code{\link{CosinorM.KDE}}
+#'
 #'
 #' @examples
 #' require (stats)
@@ -162,13 +166,15 @@
 #' # plot Cosinor in hours
 #' plot (fit$time, fit$fitted.values, type = "l", xlab = "Hour", ylab = "24-Hour Cosinor Model")
 #' }
+#'
+#'
 #' @keywords cosinor
 #' @export
 
 
 CosinorM <- function (time, activity, tau, method = "OLS", arctan2 = TRUE, type = "HC3", dilute = FALSE) {
 
-    # Check the variable class
+    # Check the variable class -----------------------------
     if (!inherits (activity, "numeric")) activity <- as.numeric (as.character (activity))
     if (all (activity == 0)) stop ("all activity values are zero")
     if (any (!is.finite (activity))) stop ("activity contains NA/NaN/Inf")
@@ -176,7 +182,7 @@ CosinorM <- function (time, activity, tau, method = "OLS", arctan2 = TRUE, type 
     if (any (time > 24 | time < 0)) stop ("Currently, the model cannot fit actigraphy recordings lasting longer than a day.
                                        Please, rescale the time coordinate to between 0 and 24.
                                        Note that it is crucial to have the proper time coordinate since the model relies on it.")
-    ## Get Essential Info -----------------
+    # Get Essential Info -----------------
     nT <- length (tau) ### Number of assumed rhythms
     vars <- as.vector (outer (c ("C", "S"), 1:nT, paste0)) ### C = x and S = z in the linear equation
 
@@ -188,7 +194,7 @@ CosinorM <- function (time, activity, tau, method = "OLS", arctan2 = TRUE, type 
     Epc <- 1 / min (dt)
 
 
-    # Build cosine and sine columns for each period
+    # Build cosine and sine columns for each period -------------------
     ## Create a model data.frame
     Mdf <- data.frame (matrix (nrow = length (time), ncol = nT * 2))
     names (Mdf) <- vars
@@ -205,7 +211,7 @@ CosinorM <- function (time, activity, tau, method = "OLS", arctan2 = TRUE, type 
[…]
