+++ R/Rad2Hr.R
@@ -17,13 +17,16 @@
 #
 #
 #
-#' @title Rad2Hr
+#' @title Radian to Hour
+#'
 #' @description
 #' Convert `acrophase` in radian to clock time based on tau.
 #'
 #' @param x Numeric vector of phases in radian.
 #' @param tau Numeric scalar. Period length in the same time units you want returned (for hours use 24). Must be a single numeric value > 0 and <= 24.
+#'
 #' @return Numeric vector of times in [0, tau). NA values propagate.
+#'
 #' @examples
 #' Rad2Hr (pi / 2, tau = 24)
 #'
@@ -34,10 +37,10 @@
 
 Rad2Hr <- function (x, tau) {
 
-    # coerce x to numeric while preserving NA
+    # coerce x to numeric while preserving NA -------------
     if (!is.numeric (x)) x <- as.numeric (as.character (x))
 
-    # validate tau
+    # validate tau --------------
     if (!is.numeric (tau) || length (tau) != 1 || is.na (tau) || !is.finite (tau)) {
         stop ("tau must be a single finite numeric value.")
     }
@@ -45,7 +48,7 @@ Rad2Hr <- function (x, tau) {
         stop ("tau must be greater than 0 and less than or equal to 24.")
     }
 
-    # conversion and wrap to [0, tau)
+    # conversion and wrap to [0, tau) ---------------
     acrophase_time <- (x * tau) / (2 * pi)
     out <- acrophase_time %% tau
 
