+++ R/globals.R
@@ -1,2 +1,2 @@
 # Set Global Functions for Internal Uses of Data
-utils::globalVariables(c("IANA"))
+utils::globalVariables(c("IANAi"))
