### 2025-12-04 e1186f1
Refactor IANA timezone handling and update tests
Replaced usage of global IANAi object with new mIANA() function for timezone data in Act2Daily and TAdjust. Removed globals.R and sysdata.rda, and added mIANA.R for internal IANA table generation. Updated test cases to explicitly specify TZ argument for consistency.
