+++ R/mIANA.R
@@ -1,0 +1,91 @@
+#  File ActiGlobe/R/sIANA.R
+#
+#  Copyright (C) 2025  C. William Yao, PhD
+#
+#  This program is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU Affero General Public License as
+#  published by the Free Software Foundation, either version 3 of the
+#  License, or any later version.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU Affero General Public License for more details.
+#
+#  You should have received a copy of the GNU Affero General Public License
+#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+#' @title Mini IANA Table for Internal Uses
+#'
+#' @description
+#' A miniturized IANA table generator that only extract the first timezone based
+#' on the abbreviated time zone code. An internal function to resolve the internal
+#' data access issue.
+#'
+#' @import utils
+#'
+#' @param Write A binary code to indicate whether to write a .csv file containing the template needed for the travel log. (default = FLASE)
+#' When set to FALSE, the template will be returned as an object.
+#' When set to TRUE, user must provide the target directory where they wish to store the template.
+#' @param Dir The directory where the travel log template to be exported <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">
+#'
+#' @return a miniturized IANA table template as a data.frame or written as a CSV file
+#'
+#' @examples
+#' \dontrun{
+#'
+#' sIANA <- mIANA(Wirte = FALSE)
+#'
+#' print(sIANA)
+#'
+#' }
+#'
+#' @keywords Travel Log Template Timezone Shift
+#' @Rd
+
+mIANA <- function(Write = FALSE, Dir = NULL){
+
+
[…]
