+++ R/globals.R
@@ -1,2 +1,0 @@
-# Set Global Functions for Internal Uses of Data
-utils::globalVariables(c("IANAi"))
