+++ R/BriefSum.R
@@ -88,6 +88,7 @@
 #' ##          SR = 1/60,
 #' ##          Start = "2017-10-24 13:45:00")
 #' }
+#'
 #' @keywords summary actigraphy
 #' @export
 
