+++ R/write.cosinor.R
@@ -24,6 +24,7 @@
 #' A multi-wrapper function which generate data trend using \code{\link[ggplot2]{geom_smooth}} and model plot using \code{\link{CosinorM}} or \code{\link{CosinorM.KDE}}. All daily results could be found in the exported report. A summary file in .csv format is also generated to update the original `Bdf` with cosinor model coefficients for each day.
 #'
 #' @import gridExtra ggplot2 grDevices utils stats
+#'
 #' @param Dir The directory where the recordings to be exported <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">
 #' @param ID The subject's ID which would be used to create a folder.
 #' @param DailyAct A data.frame list generated by `Act2Daily()`
@@ -47,6 +48,8 @@
 #' \item PDF  A multi-page PDF written to `pdfDir` (one file per `ID`). Each page corresponds to a recording day and contains two panels: the top panel shows the raw activity (points) and a smoothed trend; the bottom panel shows the model fit for the same day
 #' \item `Summary.csv` An updated `Bdf` documentation with one row per recording day. Model-based circadian parameters are appened at the end of the file for each day.
 #' }
+#'
+#' @seealso \code{\link{Act2Daily}} \code{\link{CosinorM}} \code{\link{CosinorM.KDE}}
 #'
 #' @examples
 #' \dontrun{
@@ -101,7 +104,7 @@
 #' ## paste it as an url in a web browser or the file explorer.
 #' getwd ()
 #' }
-#' @seealso \code{\link{Act2Daily}} \code{\link{CosinorM}} \code{\link{CosinorM.KDE}}
+#'
 #' @keywords write export actigraphy cosinor trend
 #' @export
 
