+++ R/DST2GL.R
@@ -40,18 +40,21 @@
 #'
 #' @examples
 #' # Example around a typical spring-forward transition (e.g. US second Sunday in March)
-#' dates <- as.Date (c ("2021-03-13", "2021-03-14", "2021-03-15"))
+#' dates <- as.Date ("2021-03-14")
 #'
 #' # On 2021-03-14 clocks jumped forward: day is 23h so output = -1
 #' DST2GL (dates)
 #'
-#' \donttest{
+#'
 #' # Example around a fall-back transition (e.g. first Sunday in November)
-#' dates <- as.Date (c ("2021-10-31", "2021-11-01", "2021-11-02"))
+#' dates <- as.Date ("2021-11-07")
 #'
 #' # On 2021-11-07 clocks fall back: day is 25h so output = +1
 #' DST2GL (dates)
-#' }
+#'
+#' # Multiple Dates
+#' sapply(c ("2021-03-13", "2021-03-14", "2021-03-15"), DST2GL)
+#'
 #' @export
 
 
