+++ R/Act2Daily.R
@@ -17,6 +17,7 @@
 #
 #
 #' @title Split and Annotate Actigraphy Data into Daily Records
+#'
 #' @description
 #' `Act2Daily` is used to split the actigraphy recording by the recording dates stored in Bdf (a BriefSum object).
 #'
@@ -26,6 +27,7 @@
 #' when you move into a time zone with a smaller UTC offset, a brief gap appears
 #' on "Travel Day", with labeling - "Unallocated", which represents "time lose".
 #'
+#'
 #' @details Transforms a continuous actigraphy time-series into a day-by-day list
 #'  of aligned epoch and produces a fully annotated data frame. Based on the
 #'  summary table generated by either \code{\link{BriefSum}} or \code{\link{TAdjust}}, it re-organize and
@@ -34,6 +36,7 @@
 #'  adjustments. All labels contained in the summary table (i.e., Bdf) will be
 #'  assigned to the corresponding data points. This workflow ensures that each day's recording is bounded by its
 #'  calendar date while accounts for travel--induced overlaps, and clearly documents any incomplete or unallocated points.
+#'
 #'
 #' @param df A data.frame of raw actigraphy recording. Both time and activity count
 #' should be included in the `df`. See `VAct` and `VTm` for further detail.
@@ -51,7 +54,11 @@
 #'   warning is issued). If FALSE, travel days and the day before/after are excluded.
 #'   Default = TRUE.
 #'
+#'
 #' @returns A data list that includes the adjusted recordings both in long format and as a nested list by day.
+#'
+#'
+#' @seealso \code{\link{TAdjust}} \code{\link{write.act}}
 #'
 #' @examples
 #' \dontrun{
@@ -69,14 +76,14 @@
 #'     )
 #'
 #'
-#' # Let's extract the quick summary of the recording
+#' # Lets extract the quick summary of the recording
 #' Bdf <- BdfList$Bdf
 #'
 #' ### Note that since the original data was affected by travel-induced time
 #' ### shift, the recordings would not be properly segmented from 2017-11-02.
 #'
 #' ## To avoid time shift due to travelling, we will keep only the first 8 days.
-#' Bdf <- Bdf [1:8, ]
+#' Bdf <- Bdf[1:8, ]
 #' df <- BdfList$df
 #'
 #' # Segment Data by Day
@@ -93,17 +100,26 @@
 #' str (dfList) ### Look at the output structure
 #' View (dfList)
 #' }
-#' @seealso \code{\link{TAdjust}} \code{\link{write.act}}
+#'
 #' @keywords Daily Actigraphy Segment
 #' @export
 
 Act2Daily <- function (df, Bdf, TUnit = "hour", VAct = NULL, VTm = NULL,
                        Incomplete = FALSE, Travel = TRUE) {
 
-    ## Extract essential per-day metadata from Bdf -------------------
+    # Extract essential per-day metadata from Bdf -------------------
+    sIANA <- mIANA()
+    iTZ <- sIANA$Timezone_IANA
+    STD <- sIANA$TZ_Code
+
 
     DT <- Bdf$Date # Calendar date for each recording day
-    aTZ <- Bdf$TZ_code # Time zone identifier per day
+
+    aTZ <- sapply (Bdf$TZ_code, function (x) { # Time zone identifier per day
+       iTZ[STD %in% x][1]
+    })
+
+
[…]
