+++ R/R2P.R
@@ -17,13 +17,42 @@
 #
 #
 #
-#' @title
-#' Convert Travelling Log to Parameters for TAdjust
+#' @title Convert Travelling Log to Parameters for TAdjust
+#'
 #' @param Bdf A \code{\link{BriefSum}} object containing per-day metadata for the recording. Note, if jet lag occurred during the recording, please, update the metadata using \code{\link{TAdjust}} before passing to this function.
 #' @param D The date travelling by plane.
 #' @param U The UTC offset values <e.g., "UTC+09:30" or "UTC-07:00">
-#' @noRd
 #'
+#' @returns
+#' A data frame containing the following columns:
+#' \itemize{
+#'   \item \code{Date} — The recording dates.
+#'   \item \code{Recording_Period} — An integer index indicating the travel period
+#'         (e.g., before/after flight segments).
+#'   \item \code{UTC} — The updated UTC offset string for each date.
+#'   \item \code{Hour_to_Adjust} — Numeric values representing the difference in hours
+#'         between the old and new UTC offsets, used for time adjustment.
+#' }
+#'
+#' @examples
+#' data (FlyEast)
+#'
+#'
+#' # Create quick summary of the recording with adjustment for daylight saving.
+#' BdfList <-
+#'     BriefSum (
+#'         df = FlyEast,
+#'         SR = 1 / 60,
+#'         Start = "2017-10-24 13:45:00"
+#'     )
+#'
+#' # Let's extract actigraphy data from a single day
+#' Bdf <- BdfList$Bdf
+#'
+#' data (TLog)
+#' R2P (Bdf, D = TLog$Date, U = TLog$UTC_Offset)
+#'
+#' @noRd
 
 
 R2P <- function (Bdf, D, U) {
