### 2025-12-04 3492889
Add TZ parameter to BriefSum and TAdjust test calls
Updated test cases in test-Act2Daily.R, test-CosinorM.R, and test-TAdjust.R to include the 'TZ' parameter set to 'America/Toronto' when calling BriefSum and TAdjust. This ensures tests are run with the correct timezone context.
