+++ tests/testthat/test-TAdjust.R
@@ -5,7 +5,8 @@ test_that ("TAdjust Test Completed", {
   BdfList <- BriefSum(
     df = FlyEast,
     SR = 1 / 60,
-    Start = "2017-10-24 13:45:00"
+    Start = "2017-10-24 13:45:00",
+    TZ = "America/Toronto"
   )
 
   # Extract actigraphy data for a single day
@@ -30,7 +31,9 @@ test_that ("TAdjust Test Completed", {
   )
 
 
-  Nadj <- TAdjust (Bdf = Bdf, TLog = tz_db)
+  Nadj <- TAdjust (Bdf = Bdf,
+                   TLog = tz_db,
+                   TZ = "America/Toronto")
 
 
 
