+++ tests/testthat/test-CosinorM.R
@@ -3,7 +3,9 @@ test_that("CosinorM works on FlyEast actigraphy data", {
   BdfList <- BriefSum(
     df = FlyEast,
     SR = 1 / 60,
-    Start = "2017-10-24 13:45:00"
+    Start = "2017-10-24 13:45:00",
+    TZ = "America/Toronto"
+	
   )
 
   # Extract actigraphy data for a single day
