### 2025-12-04 38e7708
Remove unused TZ argument and add debug prints in tests
Eliminated the unused 'TZ' argument from BriefSum in test-CosinorM.R. Added print statements for Bdf$TZ_code and tz_db in test-TAdjust.R to aid debugging.
