+++ tests/testthat/test-UTCwDST.R
@@ -1,11 +1,12 @@
 test_that("UTCwDST correctly detects DST presence", {
-  res <- sapply(c(1, 8), UTCwDST)
+
+  res <- UTCwDST(UTC = c(1, 5))
 
   # Expect a logical vector of length 2
-  expect_type(res, "logical")
+  expect_true(is.logical(res) || is.list(res))
   expect_length(res, 2)
 
   # Check specific values
-  expect_equal(res[["1"]], TRUE)
-  expect_equal(res[["8"]], FALSE)
+  expect_equal(all(res[["1"]]), TRUE)
+  expect_equal(all(res[["5"]]), FALSE)
 })
