### 2025-12-04 c6c9f7c
Update test-Act2Daily.R
