+++ _pkgdown.yml
@@ -1,4 +1,5 @@
-url: https://github.com/cwy20030/ActiGlobe
+url: https://cwy20030.github.io/ActiGlobe/
 template:
   bootstrap: 5
   math-rendering: mathjax
+
