+++ R/BriefSum.R
@@ -93,20 +93,19 @@
 #' @export
 
 BriefSum <- function (df, SR, Start, TZ = "local") {
-
     # Prepare Basic Variables ------------------
-    MN <- hms::as_hms ("00:00:00")
-    MN2 <- hms::as_hms ("24:00:00")
+    # MN <- hms::as_hms("00:00:00")
+    # MN2 <- hms::as_hms("24:00:00")
     TZ <- ifelse (TZ == "local", Sys.timezone (), TZ)
     DP <- nrow (df) # Number of Data Points
     Epc <- 1 / SR # Compute epoch length
-    TT <- DP * Epc # Total time of recordings in seconds
+    # TT <- DP * Epc # Total time of recordings in seconds
 
     nDPHr <- 3600 / Epc # Compute numbers of data points per hour
-    nDPMn <- 60 / Epc # Compute numbers of data points per minute
-    nDPSc <- 1 / Epc # Compute numbers of data points per second
+    # nDPMn <- 60 / Epc # Compute numbers of data points per minute
+    # nDPSc <- 1 / Epc # Compute numbers of data points per second
 
-    FDP <- SR * 3600 * 24 # Total data points per 24 hours
+    # FDP <- SR * 3600 * 24 # Total data points per 24 hours
 
     # Define the last time point of the day ------------------
     # Subtractor = hms::as_hms(as.difftime(Epc, units = "secs"))
@@ -179,7 +178,7 @@ BriefSum <- function (df, SR, Start, TZ = "local") {
     df$Time <- Ts ### Time in string
     df$UTC <- UTCs ### UTC in string
     df$DaylightSaving <- DSTs ### Daylight saving indicator in logical
-    df$nPoint <- 1:nrow (df) ### cumulative data point index in numeric.
+    df$nPoint <- seq_len (length.out = nrow (df)) ### cumulative data point index in numeric.
 
 
     # Initialize Report ------------
@@ -208,6 +207,4 @@ BriefSum <- function (df, SR, Start, TZ = "local") {
     class (Summary) <- c ("ActiGlobe", "data.frame")
 
     return (list ("Bdf" = Summary, "df" = df))
-
-
 }
