### 2025-12-04 7043513
Fix R-CMD-check badge link in README
Updated R-CMD-check badge link to point to the Test branch.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
