### 2025-12-04 371e787
Fix R-CMD-check badge link in README
Updated R-CMD-check badge to reflect the main branch.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
