### 2025-12-04 f79acb0
Fix badge link formatting in README.md
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
