### 2025-12-04 dbec447
Fix badge link in README.md
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
