### 2025-12-04 51e5d82
Correct badge link for R-CMD-check
Fixed badge link formatting in README.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
