### 2025-12-04 67040c8
Update NEWS.md
