### 2025-12-04 be48afd
Text update
