+++ NEWS.md
@@ -1,19 +1,19 @@
 Revision history for the R/ActiGlobe package
 
-## Next Version Update: 0.x.y
+## Next Version: 0.x.y
 - New master and class functions added: 
 
 - New internal function added:
 
 - Version update and debugged:
 
-## Version Update: 0.2.1, 2025-11-30
+## Version: 0.2.1, 2025-11-30
 
 - New master and class functions added: NA
 
 - New internal function added: Prob.Inact, mIANA
 
-- Version update and debugged: ggCosinorM, TAdjust, Act2Daily, GuessTZ
+- Version update and debugged: ggCosinorM, TAdjust, Act2Daily, GuessTZ, UTCwDST
 
 - Brief Detail: 
     - piecewise binomial estimation of inactive period using Prob.Inact
@@ -21,16 +21,15 @@ Revision history for the R/ActiGlobe package
     - augment ggCosinorM legend
     - modify internal uses of IANA in Act2Daily
     - enable switches for paralelle process for GuessTZ 
+    - add fallback in UTCwDST: map numeric UTC offsets to IANA Etc/GMT zones when OlsonNames lookup fails
 
-Anxel 
-
-## Version Update: 0.2.0, 2025-10-31
+## Version: 0.2.0, 2025-10-31
 
 - New master and class functions added: boot.seci
 
 - Version update and debugged: Description, README, CosinorM.KDE, ggCosinorM
 
-## Version Update: 0.1.9, 2025-10-27
+## Version: 0.1.9, 2025-10-27
 
 - New master and class functions added: NA
 
