+++ DESCRIPTION
@@ -13,7 +13,7 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.3.3
 Depends: 
-    R (>= 4.0.0)
+    R (>= 4.2.0)
 Date: 2025-10-14
 Imports: 
     ggplot2,
