### 2025-12-05 bdf9547
Update NEWS.md
