+++ .github/workflows/R-CMD-check.yaml
@@ -23,9 +23,15 @@ jobs:
       fail-fast: false
       matrix:
         config:
-          - {os: macos-latest,   r: 'release'}
+          - {os: macos-latest,  r: 'devel', http-user-agent: 'release'}
+          - {os: macos-latest,  r: 'release'}
+          - {os: macos-latest,  r: 'oldrel-1'}
+          - {os: macos-latest,  r: 'oldrel-2'}
 
-          - {os: windows-latest, r: 'release'}
+          - {os: windows-latest,  r: 'devel', http-user-agent: 'release'}
+          - {os: windows-latest,  r: 'release'}
+          - {os: windows-latest,  r: 'oldrel-1'}
+          - {os: windows-latest,  r: 'oldrel-2'}
 
           - {os: ubuntu-latest,  r: 'devel', http-user-agent: 'release'}
           - {os: ubuntu-latest,  r: 'release'}
