+++ man/Act2Daily.Rd
@@ -66,16 +66,16 @@ which represents "time lose".
 \dontrun{
 
 # Import data
-data(FlyEast)
+data (FlyEast)
 
 
 # Create quick summary of the recording with adjustment for daylight saving.
 BdfList <-
-  BriefSum(
-    df = FlyEast,
-    SR = 1 / 60,
-    Start = "2017-10-24 13:45:00"
-  )
+    BriefSum (
+        df = FlyEast,
+        SR = 1 / 60,
+        Start = "2017-10-24 13:45:00"
+    )
 
 
 # Lets extract the quick summary of the recording
@@ -85,22 +85,22 @@ Bdf <- BdfList$Bdf
 ### shift, the recordings would not be properly segmented from 2017-11-02.
 
 ## To avoid time shift due to travelling, we will keep only the first 8 days.
-Bdf <- Bdf[1:8, ]
+Bdf <- Bdf [1:8, ]
 df <- BdfList$df
 
 # Segment Data by Day
 dfList <-
-  Act2Daily(
-    df = df,
-    Bdf = Bdf,
-    VAct = "Activity",
-    VTm = "Time",
-    Incomplete = TRUE,
-    Travel = TRUE
-  )
+    Act2Daily (
+        df = df,
+        Bdf = Bdf,
+        VAct = "Activity",
+        VTm = "Time",
+        Incomplete = TRUE,
+        Travel = TRUE
+    )
 
-str(dfList) ### Look at the output structure
-View(dfList)
+str (dfList) ### Look at the output structure
+View (dfList)
 }
 
 }
