+++ README.md
@@ -5,8 +5,8 @@
 
 <!-- badges: start -->
 
-![version](https://img.shields.io/badge/version-0.2.0-green)
-[![R-CMD-check](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml/badge.svg?branch=Test)](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml)
+![version](https://img.shields.io/badge/version-0.1.9-green)
+[![R-CMD-check](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml)
 <!-- badges: end -->
 
 ActiGlobe is designed to make it easy to pre-process longitudinal
@@ -15,7 +15,8 @@ shift due to daylight saving changes and or cross-continental travel. It
 streamlines the process from end-to-end, to simplify the process of
 pre-processing, analyzing, and exporting daily actigraphy data and
 visual reports, making it an essential tool for researchers and
-professionals in the field.
+professionals in the field. To learn how to use it, check the vignettes
+for a quick tutorial.
 
 ## Installation
 
@@ -25,20 +26,21 @@ The pre-released version of ActiGlobe can installed from
 #### ActiGlobe-dev
 
 ``` r
-# If devtools is not available locally, please download it by removing the number symbol before the 'install.packages' code. 
+# If devtools is not available locally, please download it by removing the number symbol before the 'install.packages' code.
 # install.packages("devtools")
-devtools::install_github("cwy20030/ActiGlobe")
+devtools::install_github ("cwy20030/ActiGlobe")
 
 
 # To properly install tutorial, please, use the following code
-devtools::install_github("cwy20030/ActiGlobe",
-                         build_vignettes = TRUE)
+devtools::install_github ("cwy20030/ActiGlobe",
+    build_vignettes = TRUE
+)
 ```
 
 Coming soon… \#### ActiGlobe-release
 
 ``` r
-install.packages("ActiGlobe")
+install.packages ("ActiGlobe")
 ```
 
 <br>
@@ -48,20 +50,20 @@ install.packages("ActiGlobe")
 ### For the pre-release package ‘ActiGlobe’ in publication:
 
 ``` r
-citation("ActiGlobe")
+citation ("ActiGlobe")
 #> To cite package 'ActiGlobe' in publications use:
 #> 
-#>   Yao C, Varesco G, Simonelli G (2025). _ActiGlobe: Wearable Recording
-#>   Processor for Cross-Continental Travel_. R package version 0.1.9,
-#>   <https://github.com/cwy20030/ActiGlobe>.
+#>   Yao C, Varesco G, Bieuzen F, Simonelli G (2025). _ActiGlobe: Wearable
+#>   Recording Processor for Time Shift Adjustment and Data Analysis_. R
+#>   package version 0.2.0, <https://github.com/cwy20030/ActiGlobe>.
 #> 
 #> A BibTeX entry for LaTeX users is
 #> 
 #>   @Manual{,
-#>     title = {ActiGlobe: Wearable Recording Processor for Cross-Continental Travel},
-#>     author = {C. William Yao and Giorgio Varesco and Guido Simonelli},
+#>     title = {ActiGlobe: Wearable Recording Processor for Time Shift Adjustment and Data Analysis},
+#>     author = {C. William Yao and Giorgio Varesco and François Bieuzen and Guido Simonelli},
 #>     year = {2025},
-#>     note = {R package version 0.1.9},
[…]

[…]
