+++ tests/testthat/test-data.R
@@ -1,0 +1,3 @@
+test_that("skip data", {
+  skip('skip')
+})
