+++ tests/testthat/test-C2T.R
@@ -10,12 +10,12 @@ test_that ("C2T discrete time conversion works", {
 })
 
 test_that ("C2T continuous time conversion works", {
-   BdfList <-
-       BriefSum (
-           df = FlyEast,
-           SR = 1 / 60,
-           Start = "2017-10-24 13:45:00"
-       )
+  BdfList <-
+    BriefSum (
+      df = FlyEast,
+      SR = 1 / 60,
+      Start = "2017-10-24 13:45:00"
+    )
 
 
   df <- BdfList$df
@@ -24,13 +24,13 @@ test_that ("C2T continuous time conversion works", {
 
   # Let's extract actigraphy data from a single day
   ## Extract Second day -------------
-  fnDP <- Bdf$nDataPoints[1]
+  fnDP <- Bdf$nDataPoints [1]
 
   fDP <- fnDP + 1 ### Midnight of the second day
-  eDP <- sum(Bdf$nDataPoints[c(1,2)])
+  eDP <- sum (Bdf$nDataPoints [c (1,2)])
 
 
-  df <- df[fDP:eDP,]
+  df <- df [fDP:eDP,]
 
 
   res <- C2T (Time = df$DateTime,
@@ -39,7 +39,7 @@ test_that ("C2T continuous time conversion works", {
 
   ### expected sequence --------------
   expected <- seq(0, 24, by = 1/60)  # from 0 to 0.1666... in steps of 0.0166...
-  expected <- expected[-length(expected)]
+  expected <- expected [-length (expected)]
 
   expect_equal (res, expected)
 
