+++ DESCRIPTION
@@ -34,7 +34,8 @@ Suggests:
     knitr,
     rmarkdown,
     testthat (>= 3.0.0),
-    zeallot
+    zeallot,
+	mockery
 URL: https://github.com/cwy20030/ActiGlobe, https://cwy20030.github.io/ActiGlobe/
 BugReports: https://github.com/cwy20030/ActiGlobe/issues
 ByteCompile: true
