+++ tests/testthat/test-R2P.R
@@ -22,16 +22,16 @@ Out <-
 
 
   #### Structure checks -----------
-  expect_s3_class (Out, "data.frame")
-  expect_equal (colnames (Out), c ("Date","Recording_Period","UTC","Hour_to_Adjust"))
-  expect_equal (nrow(Out), 35)
+expect_s3_class (Out, "data.frame")
+expect_equal (colnames (Out), c ("Date","Recording_Period","UTC","Hour_to_Adjust"))
+expect_equal (nrow(Out), 35)
 
-  #### Column type checks --------------------
-  expect_type (Out$Recording_Period, "integer")
-  expect_type (Out$UTC, "character")
-  expect_type (Out$Hour_to_Adjust, "double")
+#### Column type checks --------------------
+expect_type (Out$Recording_Period, "integer")
+expect_type (Out$UTC, "character")
+expect_type (Out$Hour_to_Adjust, "double")
 
-  #### Value checks ------------------
-  expect_equal (unique (Out$Recording_Period), seq_len (4))
-  expect_equal (unique (Out$Hour_to_Adjust), c (0,12,13))
+#### Value checks ------------------
+expect_equal (unique (Out$Recording_Period), seq_len (4))
+expect_equal (unique (Out$Hour_to_Adjust), c (0,12,13))
 })
