+++ tests/testthat/test-CosinorM.KDE.R
@@ -3,8 +3,7 @@ test_that ("CosinorM.KDE works on FlyEast actigraphy data", {
     BdfList <- BriefSum (
         df = FlyEast,
         SR = 1 / 60,
-        Start = "2017-10-24 13:45:00",
-        TZ = "America/New_York"
+        Start = "2017-10-24 13:45:00"
     )
 
     # Extract actigraphy data for a single day
