+++ .github/workflows/R-CMD-check.yaml
@@ -33,7 +33,6 @@ jobs:
           - {os: windows-latest,  r: 'oldrel-1'}
           - {os: windows-latest,  r: 'oldrel-2'}
 
-          - {os: ubuntu-latest,  r: 'devel', http-user-agent: 'release'}
           - {os: ubuntu-latest,  r: 'release'}
           - {os: ubuntu-latest,  r: 'oldrel-1'}
           - {os: ubuntu-latest,  r: 'oldrel-2'}
