+++ NEWS.md
@@ -15,8 +15,6 @@ Revision history for the R/ActiGlobe package
 
 - Version update and debugged: ggCosinorM, TAdjust, Act2Daily, GuessTZ, UTCwDST
 
-- Dependency update: R >= 4.2.0
-
 - Brief Detail: 
     - piecewise binomial estimation of inactive period using Prob.Inact
     - enable internal call function for data through mIANA setting
@@ -24,6 +22,7 @@ Revision history for the R/ActiGlobe package
     - modify internal uses of IANA in Act2Daily
     - enable switches for paralelle process for GuessTZ 
     - add fallback in UTCwDST: map numeric UTC offsets to IANA Etc/GMT zones when OlsonNames lookup fails
+    - augment test suits for all internal/helper functions
 
 ## Version: 0.2.0, 2025-10-31
 

[…]
