### 2025-12-05 #732 @cwy20030
Hi @jhollist,

Thanks for the feedback.

We failed to remove the R2P.Rd file in the previous edit, which was restructured to become an internal function. In regard to test suite coverage of the package. We mistakenly thought that the coverage included only the user-facing functions, which would have been satisfied after the previous edit **81% (18/22)**. We believed that the issue arose from our package structure. Specifically, we require all developers to save frequently used internal/helper functions created in their own `.R` files, with a @noRd tag. It is likely that the review bot computed the ratio, which included these internal functions. We have now added new test suites to include these individually saved functions. Our estimated coverage ratio is **86.2% (25/29)**. We hope that this will resolve all the issues flagged by the review bot.

[![R-CMD-check](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml) [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)

For further details regarding the edit, please see the documentation below.

✖️ The following function has no documented return value: [R2P] <br>
✖️ These functions do not have examples: [R2P].
We apologize for the R2P issue. It was restructured as an internal function. We had updated the documentation for R2P, but forgot to remove it from the `man` folder on the GitHub page. Please see R2P in the R folder.  A brief snip of the now-deleted `.Rd` file is listed below. <br>
══════════════════════════════════════════════════════════════════════ <br>
#' @title Convert Travelling Log to Parameters for TAdjust<br>
.<br>
.<br>
.<br>
#' @returns <br>
#' A data frame containing the following columns: <br>
#' \itemize{ <br>
#'   \item \code{Date} — The recording dates. <br>
#'   \item \code{Recording_Period} — An integer index indicating the travel period <br>
#'         (e.g., before/after flight segments). <br>
#'   \item \code{UTC} — The updated UTC offset string for each date. <br>
#'   \item \code{Hour_to_Adjust} — Numeric values representing the difference in hours <br>
#'         between the old and new UTC offsets, used for time adjustment. <br>
#' } <br>
#' <br>
#' @examples<br>
#' data (FlyEast)<br>
#'<br>
#'<br>
#' # Create quick summary of the recording with adjustment for daylight saving.<br>
#' BdfList <-<br>
#'     BriefSum (<br>
#'         df = FlyEast, <br>
#'         SR = 1 / 60, <br>
#'         Start = "2017-10-24 13:45:00" <br>
#'     ) <br>
#' <br>
#' # Let's extract actigraphy data from a single day <br>
#' Bdf <- BdfList$Bdf <br>
#' <br>
#' data (TLog) <br>
#' R2P (Bdf, D = TLog$Date, U = TLog$UTC_Offset) <br>
══════════════════════════════════════════════════════════════════════
✖️ Package coverage is 51.5% (should be at least 75%).
We had already augmented the `testthat` tests in the previous fix. Most functions (18/22) contain at least one test suite via `testthat`. The remaining four user-facing functions are graphic and documentation generators, which were not tested in CI.  In this edit, we have added new individual test suites for all seven internal or helper functions that are not accessible to users. Of note, messages were generated in the test report for both Demand() and TAdjust() by design.

══════════════════════════════════════════════════════════════════════
==> devtools::test()

ℹ Testing ActiGlobe
✔ | F W  S  OK | Context
✔ |          3 | Act2Daily
✔ |          8 | boot.seci [48.4s]
✔ |          4 | BriefSum
✔ |          2 | C2T
✔ |          5 | CosinorM.KDE [1.2s]
✔ |         10 | CosinorM
✔ |          8 | data
✔ |          4 | Date2TotalT
✔ |          3 | DateFormat
⠏ |          0 | Demand
1: Option A
2: Option B
3: Other
✔ |          1 | Demand
✔ |          2 | DST
✔ |          2 | DST2GL
✔ |      1   0 | ggActiGlobe
✔ |      1   0 | ggCosinorM
✔ |          7 | GuessTZ [1.8s]
✔ |          6 | mIANA
✔ |          3 | Num2UTC
✔ |          4 | Prob.Inact [1.1s]
✔ |          8 | R2P
✔ |          4 | Rad2Hr
✔ |          4 | se.CosinorM
⠏ |          0 | TAdjust
Time change due to daylight saving occured at the local time zone.
✔ |          1 | TAdjust [37.4s]
✔ |          3 | TimeFormat
✔ |          9 | TravelLog
✔ |          3 | TZ2UTC
✔ |          4 | UTC2Num
✔ |          4 | UTCwDST [1.1s]
✔ |      1   0 | write.act
✔ |      1   0 | write.cosinor

══ Results ══════════════════════════════════════════════════════════════
Duration: 98.6 s

── Skipped tests (4) ────────────────────────────────────────────────────────
• skip (4): test-ggActiGlobe.R:2:5, test-ggCosinorM.R:2:5,
  test-write.act.R:2:5, test-write.cosinor.R:2:5

[ FAIL 0 | WARN 0 | SKIP 4 | PASS 112 ]
══════════════════════════════════════════════════════════════════════
