+++ reference/CosinorM.KDE.md
@@ -1,268 +1,0 @@
-# KDE-based circadian cosinor summary
-
-Fit a Gaussian kernel density estimate (KDE) on circular time (hours of
-day) weighted by activity and extract first-harmonic cosinor summaries.
-
-## Usage
-
-``` r
-CosinorM.KDE(
-  time,
-  activity,
-  bw = 0.8,
-  grid = 360L,
-  arctan2 = TRUE,
-  dilute = FALSE
-)
-```
-
-## Arguments
-
-- time:
-
-  Numeric vector of time coordinates for each data point. Values are
-  interpreted modulo 24 and must lie in \[0, 24). If not numeric, an
-  internal function will be used to convert it to numeric values in unit
-  of hour.
-
-- activity:
-
-  Numeric vector of activity counts from an actigraphy device in
-  correspondance to each time point. Non-numeric inputs will be coerced.
-  All-zero or non-finite activity values produce an error.
-
-- bw:
-
-  Numeric scalar. Kernel standard deviation (SD) controlling smoothing.
-  bw is interpreted in the same units as the angular scale after
-  conversion: the implementation converts \`bw\` from hours to radians
-  internally using \\bw\_{rad} = bw \* 2\pi / \tau\\. Default: 0.8.
-  Note, small `bw` (approaching 0) can cause numerical instability,
-  while large `bw` (approaching 1.5) leads a near-uniform kernel and
-  reduces temporal resolution.
-
-- grid:
-
-  Integer number of evaluation points used for the dense grid on \\\[0,
-  2 \* \pi)\\. The grid excludes the duplicate 2\*pi endpoint. Default:
-  360.
-
-- arctan2:
-
-  Logical; if TRUE (default) acrophase is computed with
-  `atan2(gamma, beta)`, resulting in the quadrant interval between
-  \\-\pi\\ and \\\pi\\. Whereas, when set to FALSE, the legacy
-  arctangent quadrant is mapped. The resulting interval lies between
-  \\-\frac{\pi}{2}\\ and \\\frac{\pi}{2}\\.
-
-- dilute:
-
-  Logical; if FALSE (default), all essential parameters would be
-  produced. When set to TRUE, only cosinor coefficients are returned.
-  This is suited for post-hoc processes, such as computing confidence
-  interval via nonparametric bootstrap
-
-## Value
-
-A list of class `c("CosinorM.KDE")` with elements:
-
-- parm: Parameters specified in the model (list with `time`, `activity`,
-  `bw`, `grid`)
-
-- tau: Period in hours (default to 24hour)
-
-- kdf: A data.frame with the following columns:
-
-  - theta: Angular positions (radians) at the observation angles
-
-  - density: Estimated PDF values at the observation angles (integrates
-    to 1 over \\\[0,2\pi)\\ when scaled)
-
-  - trapizoid.weight: Numeric vector of trapezoid integration weights at
-    observation angles
-
-  - kernel.weight: Denominator from kernel convolution (kernel mass at
-    each observation angle)
-
-  - fitted.values: Normalized fitted KDE (activity-scale) at observation
-    angles
-
-  - fitted.var: Variance estimate of the fitted values (for SE)
-
-  - fitted.se: Standard error of the fitted values
-
-  - hour: Corresponding clock time in hours
-
-- coef.cosinor: Named numeric vector with entries:
-
-  - MESOR: the mean of activity density over 24 hours
-
[…]
