+++ reference/UTC2Num.md
@@ -1,0 +1,31 @@
+# Convert Standard UTC Offset to Numbers
+
+Convert Standard UTC Offset to Numbers
+
+## Usage
+
+``` r
+UTC2Num(x)
+```
+
+## Arguments
+
+- x:
+
+  UTC Offsets \<e.g., "UTC+09:30" or "UTC-07:00"\>
+
+## Value
+
+Numeric values of UTC offsets
+
+## Examples
+
+``` r
+# Convert UTC to numeric values
+x <- c ("UTC+09:30", "UTC-07:00")
+
+x1 <- UTC2Num (x)
+
+print (x)
+#> [1] "UTC+09:30" "UTC-07:00"
+```

[…]
