+++ news/index.md
@@ -1,5 +1,39 @@
 # Changelog
 
+## Version 0.2.1, 2025-11-30
+
+- New master and class functions added: NA
+
+- New internal function added: Prob.Inact, mIANA
+
+- Version update and debugged: ggCosinorM, TAdjust, Act2Daily, GuessTZ,
+  UTCwDST
+
+- Brief Detail:
+
+  - piecewise binomial estimation of inactive period using Prob.Inact
+  - enable internal call function for data through mIANA setting
+  - augment ggCosinorM legend
+  - modify internal uses of IANA in Act2Daily
+  - enable switches for paralelle process for GuessTZ
+  - add fallback in UTCwDST: map numeric UTC offsets to IANA Etc/GMT
+    zones when OlsonNames lookup fails
+  - augment test suits for all internal/helper functions
+
+## Version 0.2.0, 2025-10-31
+
+- New master and class functions added: boot.seci
+
+- Version update and debugged: Description, README, CosinorM.KDE,
+  ggCosinorM
+
+## Version 0.1.9, 2025-10-27
+
+- New master and class functions added: NA
+
+- Version update and debugged: ggActiGlobe, ggCosinorM, vignettes,
+  TAdjust
+
 ## Version 0.1.8, 2025-09-23
 
 - New master and class functions added: NA
