+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   Circadian-Analysis: Circadian-Analysis.html
   Graphic-Report: Graphic-Report.html
   Time-Shift: Time-Shift.html
-last_built: 2025-12-05T23:21Z
+last_built: 2025-12-05T23:43Z
 urls:
   reference: https://cwy20030.github.io/ActiGlobe/reference
   article: https://cwy20030.github.io/ActiGlobe/articles
