### 2025-12-05 66b4ac3
Fix formatting in NEWS.md version history
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
