### 2025-12-08 80d2a25
Add overview images to README
Added images to enhance the overview of raw and adjusted recordings.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
