### 2025-12-08 6e81aaa
Fix README formatting and add overview figures
Corrected formatting and added figures in the README.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
