### 2025-12-08 414abf7
update vignettes
