+++ vignettes/Time-Shift.Rmd
@@ -74,13 +74,13 @@ For more information, type `?FlyEast` or `help(FlyEast)` in your R console.
 
 < Tip for Beginner to R> 
 1. Type: View(FlyEast) in the R console will allow users to review the data structure of FlyEast.
-2. Use <b>Tab</b> button to help finish the code. Try typing: "Fly" in the Console (i.e., the window where we can see a `>`) then press the button <b>Tab</b> on the keyboard.
+2. Use <b>Tab</b> button to help finish the code. Try typing: "Fly" in the Console (i.e., the window where we can see a `>`), then press the button <b>Tab</b> on the keyboard.
 
 
 ## Step 0. Create a brief summary of the recordings
-In general, when pre-processing or harmonizing large volumes of data, it's good practice to first generate a structured summary of the dataset at hand.
+In general, when pre-processing or harmonizing large volumes of data, it's good practice to first generate a structured summary of the dataset.
 
-Here, we use `BriefSum()` to generate an initial summary of the actigraphy data recorded in FlyEast. Based on the initial recording time, it runs through the raw recording file to provide a quick overview with simple summaries for each recording day and a newly enriched actigraphy data. This pre-processing step also automatically adds time-stamps to the original longitudinal recording. If the original location inherits the practice of daylight saving, `BriefSum()` would also automatically handle any gain or loss in time. By default, `BriefSum()` would allow us to store both the summary and the newly enriched recording data in a data.list.
+Here, we use `BriefSum()` to generate an initial summary of the actigraphy data recorded in FlyEast. Based on the initial recording time, it runs through the raw recording file to provide a quick overview with simple summaries for each recording day and a newly enriched actigraphy data. This pre-processing step also automatically adds timestamps to the original longitudinal recording. If the original location inherits the practice of daylight-saving transitioning, `BriefSum()` would also automatically handle any gain or loss in time. By default, `BriefSum()` would allow us to store both the summary and the newly enriched recording data in a `data.list`.
 
 ```{r BriefSum, message=FALSE, warning=FALSE}
 BdfList <-
@@ -105,7 +105,7 @@ c(Bdf, df) %<-%
   )
 ```
 
-In the brief summary of daily recording, we would have thirteen parameters including the information on the beginning and the end of each daily recording period, labels warning the total time of a recording is less than 24 hours and the presence of daylight saving time (based on the initial time zone).
+In the brief summary of daily recording, we would have thirteen parameters, including the information on the beginning and the end of each daily recording period, labels warning the total time of a recording is less than 24 hours and the presence of daylight-saving time (based on the initial time zone).
 
 ```{r message=FALSE, warning=FALSE, results='hide'}
 Bdf <- BdfList$Bdf
@@ -135,13 +135,13 @@ head(df) ### This should give us the same first few lines of FlyEast dataset wit
 ```{r message=FALSE, warning=FALSE,echo=FALSE, out.width="30%", paged.print=TRUE,tab.cap = "Table 3. BriefSum Enriched Longitudinal Recording"}
 knitr::kable(head(df), format = "html", table.attr = "style='width:800px;'")
 ```
-For longitudinal recordings that did not involve long-distance traveling, we may simply stop here and move onto the estimating cosinor parameters for further analysis. But given that `ActiGlobe` was designed specifically to harmonize and correct recordings affected by changes in the time zone. We have to go a few steps further to clean our data.
+For longitudinal recordings that did not involve long-distance travel, we may simply stop here and proceed to estimate cosinor parameters for further analysis. But given that `ActiGlobe` was designed specifically to harmonize and correct recordings affected by time zone changes. We have to take a few more steps to clean up the data.
 
 ## Step 1: Look into a travel diary
 
[…]

[…]
