+++ vignettes/Circadian-Analysis.Rmd
@@ -22,7 +22,7 @@ library (ActiGlobe)
 
 
 ## Cosinor Analysis
-After we segmented the data by date, we can now model the circadian rhythm using the activity scores. For this, we will use the pre-travel recording on the 27<sup>th</sup> of October. Since the recording was not affected by the time change. We can simply store the unaffected segments in a new variable called \code{df}, which stands for data frame.
+After we segmented the data by date, we can now model the circadian rhythm using the activity scores. For this, we will use the pre-travel recording on the 27<sup>th</sup> of October. Since the recording was not affected by the time change. We can simply store the unaffected segments in a new variable called `df`, which stands for data frame.
 ```{r Import 20171027, message=FALSE, warning=FALSE, results='asis'}
 # Import data
 data ("FlyEast")
@@ -72,12 +72,12 @@ ggCosinorM (fit.ols)
 
 
 #### Note 
-If we process \code{\link{FlyEast}} using the \code{\link[cosinor]{cosinor.lm}}, we will notice that the results look a bit different. This is because ActiGlobe treats the first time point of the day as \eqn{\text{00:00:00}}, whereas this is sometimes treated as \eqn{\text{00:00:00}} plus one unit of the device sampling rate due to the use of the sequence generator. 
+If we process `FlyEast` using the `cosinor::cosinor.lm`, we will notice that the results look a bit different. This is because ActiGlobe treats the first time point of the day as `00:00:00`, whereas this is sometimes treated as `00:00:00` plus one unit of the device sampling rate due to the use of the sequence generator. 
 
 
 
 #### Ecnometry-modified Cosinor Model with Feasible General Least Sqaure (FGLS)
-Besides, the traditional OLS-based cosinor model, \code{\link{CosinorM}} also allows users to estimate the model using the FGLS-based method. In ActiGlobe, the FGLS was implemented via a weighted residual least square approximation. This two-step process often produces a better representation of the "border" between sleep and wakefulness than traditional OLS. Nevertheless, neither method can realistically capture our activity patterns, given that they are simply single-phase cosinor models.
+Besides, the traditional OLS-based cosinor model,`CosinorM` also allows users to estimate the model using the FGLS-based method. In ActiGlobe, the FGLS was implemented via a weighted residual least square approximation. This two-step process often produces a better representation of the "border" between sleep and wakefulness than traditional OLS. Nevertheless, neither method can realistically capture our activity patterns, given that they are simply single-phase cosinor models.
 ```{r Model 1 FGLS, message=FALSE, warning=FALSE, fig.height=5, fig.width=7, fig.alt=  "Figure 2: Fitting of FGLS-based Cosinor Model"}
 fit.fgls <-
     CosinorM (
@@ -98,7 +98,7 @@ ggCosinorM (fit.fgls)
 
 ### Dual-phase Model (i.e., 12 + 24hour rhythms)
 
-While certain circadian measures, such as melatonin, often follow a roughly 24-hour single-phase, human activity pattern is generally fragmented. To capture the non-linearity in our activity pattern, it is possible to add more phase components when using \code{\link{CosinorM}}. Here, we provided two examples of cosinor models with two phase components, each fitted using OLS or FGLS-based methods. 
+While certain circadian measures, such as melatonin, often follow a roughly 24-hour single-phase, human activity pattern is generally fragmented. To capture the non-linearity in our activity pattern, it is possible to add more phase components when using `CosinorM`. Here, we provided two examples of cosinor models with two phase components, each fitted using OLS or FGLS-based methods. 
 
[…]
