+++ articles/Time-Shift.md
@@ -47,14 +47,14 @@ in your R console.
 \< Tip for Beginner to R\> 1. Type: View(FlyEast) in the R console will
 allow users to review the data structure of FlyEast. 2. Use **Tab**
 button to help finish the code. Try typing: “Fly” in the Console (i.e.,
-the window where we can see a `>`) then press the button **Tab** on the
+the window where we can see a `>`), then press the button **Tab** on the
 keyboard.
 
 ## Step 0. Create a brief summary of the recordings
 
 In general, when pre-processing or harmonizing large volumes of data,
-it’s good practice to first generate a structured summary of the dataset
-at hand.
+it’s good practice to first generate a structured summary of the
+dataset.
 
 Here, we use
 [`BriefSum()`](https://cwy20030.github.io/ActiGlobe/reference/BriefSum.md)
@@ -62,14 +62,14 @@ to generate an initial summary of the actigraphy data recorded in
 FlyEast. Based on the initial recording time, it runs through the raw
 recording file to provide a quick overview with simple summaries for
 each recording day and a newly enriched actigraphy data. This
-pre-processing step also automatically adds time-stamps to the original
+pre-processing step also automatically adds timestamps to the original
 longitudinal recording. If the original location inherits the practice
-of daylight saving,
+of daylight-saving transitioning,
 [`BriefSum()`](https://cwy20030.github.io/ActiGlobe/reference/BriefSum.md)
 would also automatically handle any gain or loss in time. By default,
 [`BriefSum()`](https://cwy20030.github.io/ActiGlobe/reference/BriefSum.md)
 would allow us to store both the summary and the newly enriched
-recording data in a data.list.
+recording data in a `data.list`.
 
 ``` r
 BdfList <-
@@ -99,9 +99,9 @@ c(Bdf, df) %<-%
 ```
 
 In the brief summary of daily recording, we would have thirteen
-parameters including the information on the beginning and the end of
+parameters, including the information on the beginning and the end of
 each daily recording period, labels warning the total time of a
-recording is less than 24 hours and the presence of daylight saving time
+recording is less than 24 hours and the presence of daylight-saving time
 (based on the initial time zone).
 
 ``` r
@@ -109,7 +109,9 @@ Bdf <- BdfList$Bdf
 head(Bdf)
 ```
 
-##### Table 2 The Header of Bdf ![Table 2 The Header of Bdf](images/Bdf_head.png)
+![Table 2 The Header of Bdf](images/Bdf_head.png)
+
+Table 2 The Header of Bdf
 
 In the enriched data - `df`, we have both the original data stored in
 `FlyEast` and some new information, such as the time stamp of each data
@@ -132,25 +134,25 @@ head(df) ### This should give us the same first few lines of FlyEast dataset wit
 
 Table 3. BriefSum Enriched Longitudinal Recording
 
-For longitudinal recordings that did not involve long-distance
-traveling, we may simply stop here and move onto the estimating cosinor
-parameters for further analysis. But given that `ActiGlobe` was designed
-specifically to harmonize and correct recordings affected by changes in
-the time zone. We have to go a few steps further to clean our data.
+For longitudinal recordings that did not involve long-distance travel,
+we may simply stop here and proceed to estimate cosinor parameters for
[…]
