+++ articles/Circadian-Analysis.md
@@ -2,11 +2,11 @@
 
 ## Cosinor Analysis
 
-After we segmented the data by dates, we can now move on to model the
-circadian rhythm based on the activity scores. For this, we are going to
-use the pre-travel recording on the 27^(th) of October. Since the
-recording was not affected by time change. We can simply export it to a
-new variable, called , which is short for data frame.
+After we segmented the data by date, we can now model the circadian
+rhythm using the activity scores. For this, we will use the pre-travel
+recording on the 27^(th) of October. Since the recording was not
+affected by the time change. We can simply store the unaffected segments
+in a new variable called `df`, which stands for data frame.
 
 ``` r
 # Import data
@@ -52,20 +52,23 @@ Model](Circadian-Analysis_files/figure-html/Model%201%20OLS-1.png)
 
 #### Note
 
-If we process using the , we will notice that the results look a bit
-different. This is because ActiGlobe treats the first time point of the
-day as , whereas it is often treated as the first time point due to the
-specification in the sequence generator.
+If we process `FlyEast` using the
+[`cosinor::cosinor.lm`](https://rdrr.io/pkg/cosinor/man/cosinor.lm.html),
+we will notice that the results look a bit different. This is because
+ActiGlobe treats the first time point of the day as `00:00:00`, whereas
+this is sometimes treated as `00:00:00` plus one unit of the device
+sampling rate due to the use of the sequence generator.
 
 #### Ecnometry-modified Cosinor Model with Feasible General Least Sqaure (FGLS)
 
-Besides, the traditioanl OLS-based cosinor model, also allows user to
-estimate model using FGLS-based method. In ActiGlobe, the FGLS was
-implemented via a weighted residual least square approximation. This
-two-step process often produces a better representation of the “boarder”
-between sleep and awake activity than otherwise from the traditional
-OLS. Nevertheless, neither methods can offer realistic few of our
-activity pattern given they are simply single-phase cosinor models.
+Besides, the traditional OLS-based cosinor model,`CosinorM` also allows
+users to estimate the model using the FGLS-based method. In ActiGlobe,
+the FGLS was implemented via a weighted residual least square
+approximation. This two-step process often produces a better
+representation of the “border” between sleep and wakefulness than
+traditional OLS. Nevertheless, neither method can realistically capture
+our activity patterns, given that they are simply single-phase cosinor
+models.
 
 ``` r
 fit.fgls <-
@@ -93,7 +96,7 @@ Model](Circadian-Analysis_files/figure-html/Model%201%20FGLS-1.png)
 While certain circadian measures, such as melatonin, often follow a
 roughly 24-hour single-phase, human activity pattern is generally
 fragmented. To capture the non-linearity in our activity pattern, it is
-possible to add additional phase components when using . Here, we
+possible to add more phase components when using `CosinorM`. Here, we
 provided two examples of cosinor models with two phase components, each
 fitted using OLS or FGLS-based methods.
 
@@ -159,11 +162,11 @@ ggCosinorM (fit.fgls2)
 Model](Circadian-Analysis_files/figure-html/Model%202%20FGLS-1.png)
 
 It is worth noting that we cannot individually interpret each cosinor
-parameters alone when fitting a multicomponent model, This is because
-the rest-activity patterns estimated by general linear regression is the
-cumulative sum of all the fitted cosinor components (i.e. both 12 and
-24hour waves). As such, it is better to use the parameters extracted
[…]
