### 2025-12-09 #732 @jhollist
@cwy20030 A couple of ideas for you.

For the ggplot output, you can test the objects themselves.  A nice explanation of this is available at https://www.jumpingrivers.com/blog/r-testthat/#testing-plots

Also, I don't think you should have any issues with the temporary files.  In your tests, you can save the the temp file name and use that in your tests.  Something like:

```
temp_file <- tempfile()
write.csv(the_data, temp_file)
test_the_data <- read.csv(temp_file)
# your tests here
```

Apologies if I am not understanding the question.
