+++ .github/workflows/R-CMD-check.yaml
@@ -6,7 +6,7 @@
 # usethis::use_github_action("check-standard") will install it.
 on:
   push:
-    branches: [Test-Jennick]
+    branches: [Test-Jeannick]
   pull_request:
 
 name: R-CMD-check.yaml
