+++ R/write.cosinor.R
@@ -69,6 +69,7 @@
 #'
 #' # Let's extract the quick summary of the recording
 #' Bdf <- BdfList$Bdf
+#' df <- BdfList$df
 #'
 #' ### Note that since the original data was affected by travel-induced time
 #' ### shift, the recordings would not be properly segmented from 2017-11-02.
@@ -80,10 +81,10 @@
 #' ## Segment Data by Day
 #' dfList <-
 #'     Act2Daily (
-#'         df = FlyEast,
+#'         df = df,
 #'         Bdf = Bdf,
 #'         VAct = "Activity",
-#'         VTm = "X2",
+#'         VTm = "Time",
 #'         Incomplete = TRUE,
 #'         Travel = TRUE
 #'     )
@@ -94,7 +95,7 @@
 #' write.cosinor (
 #'     Dir = getwd (), ## Export to the current working directory
 #'     ID = "JD",
-#'     df = dfList$Daily_df,
+#'     DailyAct = dfList$Daily_df,
 #'     Bdf = Bdf,
 #'     VAct = "Activity",
 #'     VTm = "Time"
@@ -110,19 +111,31 @@
 
 
 write.cosinor <- function (Dir, ID, DailyAct, Bdf, VAct = NULL, VTm = NULL, method = "OLS", tau = 24, ph = FALSE, overwrite = FALSE) {
+
     ## Get Essential Info -----------------
     nT <- length (tau) ### Number of assumed rhythms
     D <- names (DailyAct)
     U <- Bdf$UTC
 
+    ### Set default variable names for activity and time columns
+    if (is.null (VAct)) VAct <- names (df) [[2]] # Default: second column of df
+    if (is.null (VTm)) VTm <- names (df) [[1]] # Default: first column of df
+
     ## Initialize all cosinor results -----------------------
+
+
     if (all (length (tau) == 1, isFALSE (ph))) {
         nVNames <- c ("MESOR", "Amplitude", "Acrophase", "Acrophase.time")
         Bdf [nVNames] <- NA
+        KeyTerm <- "MESOR|Amplitude|Acrophase"
+
     } else {
         nVNames <- c ("MESOR", "Bathyphase.time", "Trough.ph", "Acrophase.time", "Peak", "Amplitude")
         Bdf [nVNames] <- NA
+        KeyTerm <- "MESOR|Bathyphase.time|Trough.ph|Acrophase.time|Peak|Amplitude"
     }
+
+
 
 
     ## Check Directory ----------------
@@ -137,32 +150,32 @@ write.cosinor <- function (Dir, ID, DailyAct, Bdf, VAct = NULL, VTm = NULL, meth
     }
 
     grDevices::pdf (pdfDir, onefile = TRUE, paper = "a4r")
+
+
     ### Plot ----------------
     for (d in D) {
-        print (d)
+        # print (d) ## For fail-check function purposes
         df <- DailyAct [[d]]
 
         if (!is.null (df)) {
-            ### Get Variable Names -------------
 
-            ### Set default variable names for activity and time columns
-            if (is.null (VAct)) VAct <- names (df) [[2]] # Default: second column of df
-            if (is.null (VTm)) VTm <- names (df) [[1]] # Default: first column of df
+            # Check the variable class -----------------------------
+            Tm <- df [[VTm]]
+            Act <- df [[VAct]]
+            if (!inherits (Act, "numeric")) Act <- as.numeric (as.character (Act))
+            if (!inherits (Tm, "numeric")) Tm <- C2T (Time = Tm, Discrete = TRUE)
 
 
-            names (df) <- c ("Tm", "Act")
-            Tm <- as.numeric (df [["Tm"]])
-            Act <- as.numeric (df [["Act"]])
 
[…]
