### 2025-12-11 fdb6d84
Delete .github/workflows/pkgcheck.yaml
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
