+++ .github/workflows/pkgcheck.yaml
@@ -1,23 +1,0 @@
-name: pkgcheck
-
-# This will cancel running jobs once a new run is triggered
-concurrency:
-  group: ${{ github.workflow }}-${{ github.head_ref }}
-  cancel-in-progress: true
-
-on:
-  # Manually trigger the Action under Actions/pkgcheck
-  workflow_dispatch:
-  # Run on every push to main
-  push:
-    branches:
-      - main
-      - Test
-
-jobs:
-  pkgcheck:
-    runs-on: ubuntu-latest
-    permissions:
-      issues: write
-    steps:
-      - uses: ropensci-review-tools/pkgcheck-action@main
