+++ tests/testthat/helper-Demand.R
@@ -1,9 +1,4 @@
 test_that("Demand returns 1 when user selects 1", {
-
-  if (Sys.info()[["sysname"]] %in% "Linux") {
-    skip("Skip on ubuntu-latest (devel)")
-  }
-
   # mock readline to always return "1"
   mockery::stub(Demand, "readline", function(prompt = "") "1")
 
@@ -11,5 +6,6 @@ test_that("Demand returns 1 when user selects 1", {
 
   # ---- Content checks ----
   expect_equal(result, "Option A")
+         
 })
 
