### 2025-12-11 fa40ed1
Add skip condition for Linux and macOS in tests
Skip the test on Linux and macOS due to segfault issues.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
