+++ .github/workflows/pkgcheck.yaml
@@ -1,27 +1,0 @@
-name: pkgcheck
-
-# This will cancel running jobs once a new run is triggered
-concurrency:
-  group: ${{ github.workflow }}-${{ github.head_ref }}
-  cancel-in-progress: true
-
-on:
-  # Manually trigger the Action under Actions/pkgcheck
-  workflow_dispatch:
-  # Run on every push to main
-  push:
-    branches:
-      - main
-      - Test
-
-jobs:
-  pkgcheck:
-    strategy:
-      matrix:
-        os: [ubuntu-latest, windows-latest, macos-latest]
-    runs-on: ${{ matrix.os }}
-    permissions:
-      issues: write
-    steps:
-      - uses: actions/checkout@v4
-      - uses: ropensci-review-tools/pkgcheck-action@main
