+++ tests/testthat/test-write.act.R
@@ -1,6 +1,6 @@
 test_that("write.act exports daily recordings and summary correctly", {
 
-  if (grepl ("Linux|macOS", Sys.info ()[["sysname"]])) {
+  if (grepl ("Linux|Darwin", Sys.info ()[["sysname"]])) {
     skip("Skip on Linux and macOS due to segfault fail")
   } else {
 
