### 2025-12-11 0f2c657
Enhance test to handle OS-specific behavior
Add system info print and skip condition for Linux/macOS.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
