### 2025-12-11 c459fae
Remove print statement in write.act test
Remove print statement for system info in test.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
