### 2025-12-11 d829da5
Add test for Demand function with mock input
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
