+++ tests/testthat/helper-Demand.R
@@ -9,3 +9,14 @@ test_that("Demand returns 1 when user selects 1", {
          
 })
 
+
+test_that("Demand returns Other when user selects 2", {
+  
+  # mock readline to always return "3"
+  mockery::stub(Demand, "readline", function(prompt = "") "2")
+
+  result <- Demand(c("NOOOO", "Yes", "Other"), "option")
+
+  # ---- Content checks ----
+  expect_equal(result, "Yes")
+})
