+++ tests/testthat/helper-Demand.R
@@ -1,4 +1,4 @@
-test_that("Demand returns 1 when user selects 1", {
+test_that("Demand returns proper response", {
   # mock readline to always return "1"
   mockery::stub(Demand, "readline", function(prompt = "") "1")
 
@@ -6,6 +6,13 @@ test_that("Demand returns 1 when user selects 1", {
 
   # ---- Content checks ----
   expect_equal(result, "Option A")
-         
-})
 
+
+  # mock readline to always return "3"
+  mockery::stub(Demand, "readline", function(prompt = "") "2")
+
+  result <- Demand(c("NOOOO", "Yes", "Other"), "option")
+
+  # ---- Content checks ----
+  expect_equal(result, "Yes")
+})
