+++ R/GuessTZ.R
@@ -71,7 +71,6 @@ GuessTZ <- function (aOF, DT = NULL, iTZ = NULL, All = TRUE, fork = FALSE) {
     # Establish initial time zone ----------------
     TZ1 <- ifelse (iTZ == "local", Sys.timezone (), iTZ)
 
-
     # Extract all known time zones ----------------
     oTZs <- OlsonNames ()
 
@@ -80,11 +79,8 @@ GuessTZ <- function (aOF, DT = NULL, iTZ = NULL, All = TRUE, fork = FALSE) {
     if (is.null (iTZ)){
         TZ1 <- NULL
     } else if (!TZ1 %in% oTZs) {
-
-        stop (sprintf (
-            "The provided time zone \"%s\" is not recognized.\n",
-            TZ1
-        ), "Please check the spelling or consult the IANA time zone table (ActiGlobe::IANA).")
+        stop (sprintf ("The provided time zone \"%s\" is not recognized.\n",
+            TZ1), "Please check the spelling or consult the IANA time zone table (ActiGlobe::IANA).")
     }
 
     ## Process DT
@@ -116,45 +112,31 @@ GuessTZ <- function (aOF, DT = NULL, iTZ = NULL, All = TRUE, fork = FALSE) {
         Toffs <- unlist (Toffs)
     } else {
         ## Sequential version
-        Toffs <- vapply (
-            oTZs,
-            function (tz) format (as.POSIXct (DT, tz = tz), "%z"),
-            character (1)
-        )
+        Toffs <- vapply (oTZs, function (tz)
+            format (as.POSIXct (DT, tz = tz), "%z"),
+            character (1) )
     }
 
-
     #### Step 1 Guess all possible TZ indicators -----------
-    pTZs <- sapply (
-        aOF,
-        function (x) oTZs [Toffs %in% x]
-    )
+    pTZs <- sapply (aOF, function (x) oTZs [Toffs %in% x])
 
     #### Step 2 Check if the initial time zone is included
     if (!is.null (TZ1)) {
         if (length (aOF) == 1) {
             pTZs <- ifelse (TZ1 %in% pTZs, TZ1, pTZs)
         } else {
-            pTZs <- sapply (
-                pTZs,
-                function (x) ifelse (TZ1 %in% x, TZ1, x)
-            )
+            pTZs <- sapply (pTZs, function (x) ifelse (TZ1 %in% x, TZ1, x))
         }
     }
-
 
     #### Step 3 Keep only the first one if the All is set to FALSE
     if (!All) {
         if (length (aOF) > 1) {
-            pTZs <- sapply (
-                pTZs,
-                function (x) x [[1]]
-            )
+            pTZs <- sapply (pTZs, function (x) x [[1]])
         } else {
             pTZs <- pTZs [[1]]
         }
     }
-
 
     return (pTZs)
 }
