### 2025-12-11 23e7d5e
rOpenSci Update
Introduced the UnitFactor helper for consistent time unit conversion across functions. Improved input validation and error handling in functions such as Act2Daily, C2T, BriefSum, Date2TotalT, DST2GL, TZ2UTC, Num2UTC, UTC2Num, and others. Updated documentation and test helpers, enhanced robustness for time zone and numeric checks, and improved code clarity and maintainability. Also updated ggplot2 usage for compatibility and added references to documentation.
