+++ R/GuessTZ.R
@@ -24,6 +24,7 @@
 #' initial time zone (`iTZ`) and allowing parallel computation for speed.
 #'
 #' @details
+#' An exhaustive process was
 #' - Uses `OlsonNames()` to retrieve all known time zones.
 #' - Computes offsets for each zone at the reference date/time (`DT`).
 #' - Matches observed offsets (`aOF`) against computed offsets.
@@ -70,11 +71,17 @@ GuessTZ <- function (aOF, DT = NULL, iTZ = NULL, All = TRUE, fork = FALSE) {
     # Establish initial time zone ----------------
     TZ1 <- ifelse (iTZ == "local", Sys.timezone (), iTZ)
 
-    if (is.null (iTZ)) TZ1 <- NULL
-
-
     # Extract all known time zones ----------------
     oTZs <- OlsonNames ()
+
+    # Check Point ----------------------
+    # Validate the time zone
+    if (is.null (iTZ)){
+        TZ1 <- NULL
+    } else if (!TZ1 %in% oTZs) {
+        stop (sprintf ("The provided time zone \"%s\" is not recognized.\n",
+            TZ1), "Please check the spelling or consult the IANA time zone table (ActiGlobe::IANA).")
+    }
 
     ## Process DT
     if (is.null (DT)) {
@@ -105,45 +112,31 @@ GuessTZ <- function (aOF, DT = NULL, iTZ = NULL, All = TRUE, fork = FALSE) {
         Toffs <- unlist (Toffs)
     } else {
         ## Sequential version
-        Toffs <- vapply (
-            oTZs,
-            function (tz) format (as.POSIXct (DT, tz = tz), "%z"),
-            character (1)
-        )
+        Toffs <- vapply (oTZs, function (tz)
+            format (as.POSIXct (DT, tz = tz), "%z"),
+            character (1) )
     }
 
-
     #### Step 1 Guess all possible TZ indicators -----------
-    pTZs <- sapply (
-        aOF,
-        function (x) oTZs [Toffs %in% x]
-    )
+    pTZs <- sapply (aOF, function (x) oTZs [Toffs %in% x])
 
     #### Step 2 Check if the initial time zone is included
     if (!is.null (TZ1)) {
         if (length (aOF) == 1) {
             pTZs <- ifelse (TZ1 %in% pTZs, TZ1, pTZs)
         } else {
-            pTZs <- sapply (
-                pTZs,
-                function (x) ifelse (TZ1 %in% x, TZ1, x)
-            )
+            pTZs <- sapply (pTZs, function (x) ifelse (TZ1 %in% x, TZ1, x))
         }
     }
-
 
     #### Step 3 Keep only the first one if the All is set to FALSE
     if (!All) {
         if (length (aOF) > 1) {
-            pTZs <- sapply (
-                pTZs,
-                function (x) x [[1]]
-            )
+            pTZs <- sapply (pTZs, function (x) x [[1]])
         } else {
             pTZs <- pTZs [[1]]
         }
     }
-
 
     return (pTZs)
 }
