### 2025-12-11 289e685
Simplify R-CMD-check workflow matrix
Removed old R versions from the CI matrix for macOS and Windows.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
