### 2025-12-11 35720df
Remove platform-specific skip and validation tests
Removed checks for Linux and macOS, and deleted various validation tests for the summary CSV and Rad2Hr function.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
