### 2025-12-11 5eb445e
Delete tests/testthat/test-TAdjust.R
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
