### 2025-12-11 b9fbe67
Delete vignettes directory
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
