+++ vignettes/Time-Shift.Rmd
@@ -1,255 +1,0 @@
----
-title: "Time-Shift"
-subtitle: "Pre-process Data Affected by Daylight Saving and Travel-induced Time Shift"
-author: "C. William Yao"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Time-Shift}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
-resource_files:
-  - ./images/Bdf_head.png
-  - ./images/Bdf_demo.png
-  - ./images/Bdf.adj_demo.png
-  - ./images/Unadjusted_Overview.png
-  - ./images/Adjusted_Overview.png
----
-
-
-## What will be covered in this tutorial?
-This specific tutorial is designed specifically to show users how to use ActiGlobe to preprocess longitudinal actigrpahy recordings affected by time shift. For analysis and generation of graphic/excel report of daily actigraphy measures, please go to tutorial title: Graphic-Report.
-
-
-
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-## Load the Libraries
-```{r setup}
-library(ActiGlobe)
-
-### Optional Library
-# library(zeallot)
-```
-If any of the packages has yet been installed, we can always install them using the function ["install.packages()"]. 
-
-
-## Load Example Data: FlyEast
-```{r Load Data, message=FALSE, warning=FALSE, results='hide'}
-data("FlyEast")
-
-head(FlyEast) ### Only the first few lines
-```
-
-
-```{r message=FALSE, warning=FALSE,echo=FALSE, out.width="30%", paged.print=TRUE, tab.cap = "Table 1. An Overview of the FlyEast Dataset"}
-knitr::kable(head(FlyEast))
-```
-
-For more information, type `?FlyEast` or `help(FlyEast)` in your R console.
-
-< Tip for Beginner to R> 
-1. Type: View(FlyEast) in the R console will allow users to review the data structure of FlyEast.
-2. Use <b>Tab</b> button to help finish the code. Try typing: "Fly" in the Console (i.e., the window where we can see a `>`), then press the button <b>Tab</b> on the keyboard.
-
-
-## Step 0. Create a brief summary of the recordings
-In general, when pre-processing or harmonizing large volumes of data, it's good practice to first generate a structured summary of the dataset.
-
-Here, we use `BriefSum()` to generate an initial summary of the actigraphy data recorded in FlyEast. Based on the initial recording time, it runs through the raw recording file to provide a quick overview with simple summaries for each recording day and a newly enriched actigraphy data. This pre-processing step also automatically adds timestamps to the original longitudinal recording. If the original location inherits the practice of daylight-saving transitioning, `BriefSum()` would also automatically handle any gain or loss in time. By default, `BriefSum()` would allow us to store both the summary and the newly enriched recording data in a `data.list`.
-
-```{r BriefSum, message=FALSE, warning=FALSE}
-BdfList <-
-  BriefSum(
-    df = FlyEast,
-    SR = 1 / 60,
-    Start = "2017-10-24 13:45:00"
-  )
-
-
-str(BdfList, max.level = 1) ### An overview of the output structure from BriefSum()
-```
-
-We can also take the advantage of `zeallot` to store multiple outputs at once just like in `Matlab` and `Python`.
-
-```{r eval=FALSE}
-c(Bdf, df) %<-%
-  BriefSum(
-    df = FlyEast,
-    SR = 1 / 60,
-    Start = "2017-10-24 13:45:00"
-  )
-```
-
[…]
