### 2025-12-11 2368b48
Delete tests/testthat/test-boot.seci.R
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
