### 2025-12-11 e842a19
Delete tests/testthat/test-UTCwDST.R
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
