### 2025-12-11 f286f52
Add debug print statements in test-write.act.R
Added print statements to debug directory existence and process flow.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
