### 2025-12-11 0be9b1e
Fix temporary directory setup in write.act tests
Updated the temporary directory creation for testing.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
