### 2025-12-11 6712333
Add skip condition for Linux and macOS tests
Skip tests on Linux and macOS due to segfault issues.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
