### 2025-12-11 13afe98
Insert debug print statements in write.act.R
Added print statements for debugging purposes.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
