### 2025-12-11 2d0b80b
Update write.act.R
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
