### 2025-12-11 5de6889
Refactor write.act function for better defaults
Refactor write.act function to set default VAct to the second column and improve error handling for time and activity variables.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
