### 2025-12-11 f28e5a2
Refactor time parsing logic in C2T.R
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
