+++ R/C2T.R
@@ -61,8 +61,8 @@ C2T <- function (Time, Discrete = FALSE) {
 
     # If All NAs, Parse as Time Strings -------------
     if (length (na.omit (x)) == 0) {
-        Fmt <- TimeFormat (Time)
-        hms <- as.POSIXct (Time, format = Fmt)
+        hms <- TimeFormat (Time, as.time = TRUE)
+        # hms <- as.POSIXct (Time, format = Fmt)
 
         decimal_hours <- as.numeric (format (hms, "%H")) +
             as.numeric (format (hms, "%M")) / 60 +
