### 2025-12-11 dcf8699
Fix time parsing logic for all NA values
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
