### 2025-12-11 986160c
Update print statement in write.act.R
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
