+++ R/write.act.R
@@ -113,6 +113,7 @@ write.act <- function (Dir, ID, df, Bdf, TUnit = "hour", VAct = NULL, VTm = NULL
     if (!inherits (activity, "numeric")) activity <- as.numeric (as.character (activity))
     if (all (activity == 0)) stop ("all activity values are zero")
     if (any (!is.finite (activity))) stop ("activity contains NA/NaN/Inf")
+   print (100)
     if (!inherits (time, "numeric")) time <- C2T (Time = time, Discrete = TRUE)
     if (any (time > 24 | time < 0)) stop ("Currently, the model cannot fit actigraphy recordings lasting longer than a day.
                                        Please, rescale the time coordinate to between 0 and 24.
@@ -120,7 +121,7 @@ write.act <- function (Dir, ID, df, Bdf, TUnit = "hour", VAct = NULL, VTm = NULL
 
 
 
- print (100)
+ print (10011)
     df [[VTm]] <- time 
     df [[VAct]] <- activity
 
