### 2025-12-11 fa76779
Add print statement for debugging
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
