### 2025-12-11 6062e72
Add print statement for 'time' in write.act.R
Added a print statement for the 'time' variable.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
