+++ R/C2T.R
@@ -54,7 +54,7 @@ C2T <- function (Time, Discrete = FALSE) {
         stop ("Input 'Time' must be a pure character time string with no timezone label.
               Please, check the input using TimeFormat().")
 
-
+    hms <- Time
 
     # First Attempt to Coerce to Numeric -------------
     x <- suppressWarnings (as.numeric (as.character (Time)))
@@ -62,8 +62,7 @@ C2T <- function (Time, Discrete = FALSE) {
     # If All NAs, Parse as Time Strings -------------
     if (length (na.omit (x)) == 0) {
         Fmt <- TimeFormat (Time, as.time = FALSE)
-        print(5555)
-        hms <- as.POSIXlt (Time, format = Fmt)
+        hms <- as.POSIXct (Time, format = Fmt)
 
         decimal_hours <- as.numeric (format (hms, "%H")) +
             as.numeric (format (hms, "%M")) / 60 +
