### 2025-12-11 071f424
Refactor time conversion logic in write.act.R
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
