### 2025-12-11 3ec7436
Update C2T.R
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
