### 2025-12-11 80e1a8f
Refactor time handling for different OS in write.act.R
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
