### 2025-12-11 d5fc83e
Enhance validation for activity and time variables
Added checks for activity and time variables, including handling for zero values and non-finite numbers. Improved time conversion based on the operating system and added a constraint for time values.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
