### 2025-12-11 35b87db
Fix formatting issues in test-write.cosinor.R
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
