### 2025-12-11 f48ed65
Update test-write.cosinor.R
