### 2025-12-11 6082ab9
Update test-write.cosinor.R
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
