### 2025-12-11 e387ae4
Clean up print statements in test-write.act.R
Removed print statements for cleaner output.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
