+++ tests/testthat/test-write.act.R
@@ -20,7 +20,7 @@ test_that("write.act exports daily recordings and summary correctly", {
 
   Bdf <- Bdf [1:5, ]
 
-print("prep")
+
  # Call write.act ------------------
 write.act (
   Dir = tmpdir, ## Export to the current working directory
@@ -33,10 +33,10 @@ write.act (
   Travel = TRUE,
   Simple = FALSE
 )
-print (0)
+
   # Construct expected directory ----------------
   fDir <- file.path (paste0 (tmpdir, "/JD"))
-print (1)
+
   # Check that directory was created ------------------
   expect_true (dir.exists (fDir))
 
