### 2025-12-11 7b881c4
Fix variable name in time conversion
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
