### 2025-12-11 39cce78
Update R-CMD-check.yaml
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
