### 2025-12-11 d07e9c1
Refactor input validation with ValInput helper
Replaces repetitive activity and time input validation logic in multiple functions with a new ValInput() helper. This centralizes validation and conversion for activity and time vectors, improving code maintainability and consistency.
