### 2025-12-11 82807d3
Update ValInput.R
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
