+++ man/Act2Daily.Rd
@@ -8,8 +8,6 @@ Act2Daily(
   df,
   Bdf,
   TUnit = "hour",
-  VAct = NULL,
-  VTm = NULL,
   Incomplete = FALSE,
   Travel = TRUE
 )
@@ -93,8 +91,6 @@ dfList <-
     Act2Daily (
         df = df,
         Bdf = Bdf,
-        VAct = "Activity",
-        VTm = "Time",
         Incomplete = TRUE,
         Travel = TRUE
     )
