+++ vignettes/Time-Shift.Rmd
@@ -201,8 +201,6 @@ knitr::include_graphics("images/Unadjusted_Overview.png",dpi = 200)
 dfList <- Act2Daily(
   df = df,
   Bdf = Bdf.adj,
-  VAct = "Activity",
-  VTm = "Time",
   Incomplete = TRUE,
   Travel = TRUE
 )
