### 2025-12-11 203ff9d
Add ValInput helper and refactor input validation
Introduces the ValInput function for unified input validation and conversion of activity and time vectors, with OS-specific handling for time conversion. Refactors CosinorM, CosinorM.KDE, Prob.Inact, write.act, and write.cosinor to use ValInput, improving robustness and reducing code duplication. Updates tests and adds new tests for ValInput and related features. Also adds a GitHub Actions workflow for pkgcheck.
