+++ R/Prob.Inact.R
@@ -101,11 +101,9 @@
 #' @noRd
 
 Prob.Inact <- function (y, T, k = 12, threshold = 3, logical = TRUE) {
-    # Check the variable class -----------------------------
-    if (!inherits (y, "numeric")) y <- as.numeric (as.character (y))
-    if (all (y == 0)) stop ("all activity values are zero")
-    if (any (!is.finite (y))) stop ("activity contains NA/NaN/Inf")
-    if (!inherits (T, "numeric")) T <- C2T (Time = T, Discrete = TRUE)
+    # Check Point and Input Validation -------------------------
+	y <- ValInput(x = y, type = "Act")
+	T <- ValInput(x = T, type = "Tm")
 
 
     # Parameters -------------------
