### 2025-12-11 6065d55
Add Ubuntu devel configuration to R-CMD-check
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
