### 2025-12-11 39c0e51
Remove testthat unit tests
Deleted multiple testthat test files from the tests/testthat directory, removing unit tests for C2T, Demand, GuessTZ, Prob.Inact, R2P, mIANA, and se.CosinorM functions.
