### 2025-12-11 fa5a118
Fix basic pkgcheck issues: remove Date, fix =, T/F, sapply
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
