+++ R/UTC2Num.R
@@ -46,9 +46,9 @@ UTC2Num <- function (x) {
     x <- trimws (x) # Remove white space
 
     # Step 2: Validate the format -------------------------
-    mp <- unlist (sapply (x, function (i) {
+    mp <- unlist (vapply (x, function (i) {
         ifelse (grepl ("-", i), -1, 1)
-    }))
+    }, FUN.VALUE = numeric(1)))
     a <- gsub ("\\+|\\-", "", x) #### Remove plus minus sign
 
     b <- as.numeric (gsub (":.*", "", a))
