### 2025-12-11 54f2fb6
Fix goodpractice issues: remove Date field, replace = with <-, replace T/F with TRUE/FALSE, replace 1:length with seq_along, replace sapply with vapply/lapply
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
