### 2025-12-11 07f3089
Fix additional code style issues: break long lines
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
