### 2025-12-11 2de903b
Add comprehensive documentation report for pkgcheck fixes
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
