+++ PKGCHECK_FIXES_REPORT.md
@@ -1,0 +1,285 @@
+# pkgcheck Issues Fix Report
+
+**Date**: December 11, 2025  
+**Branch**: fix/pkgcheck-issues  
+**Author**: GitHub Copilot Coding Agent
+
+## Executive Summary
+
+This report documents all code improvements made to address issues identified by the pkgcheck action in the ActiGlobe R package. The fixes enhance code quality, maintainability, and adherence to R best practices.
+
+## Issues Identified by pkgcheck
+
+The pkgcheck workflow identified the following issues marked with ✖:
+
+1. ✖ Package fails continuous integration checks
+2. ✖ High cyclomatic complexity in `ggCosinorM` function (74)
+3. ✖ "Date" field present in DESCRIPTION (not required)
+4. ✖ Use of '=' for assignment instead of '<-'
+5. ✖ Long code lines (>80 characters)
+6. ✖ Use of `sapply()` instead of type-safe `vapply()`
+7. ✖ Use of `1:length()` instead of `seq_len()` or `seq_along()`
+8. ✖ Not importing packages as a whole
+9. ✖ Use of 'T' and 'F' instead of TRUE and FALSE
+
+## Changes Made
+
+### 1. DESCRIPTION File ✅
+
+**Issue**: Date field present in DESCRIPTION  
+**Action**: Removed the Date field (line 17)
+
+**File**: `DESCRIPTION`
+**Change**:
+```diff
+- Date: 2025-10-14
+```
+
+**Rationale**: The Date field is not required and becomes outdated quickly. R CMD build automatically adds a build date.
+
+---
+
+### 2. Assignment Operator ✅
+
+**Issue**: Use of '=' instead of '<-' for assignment  
+**Action**: Changed assignment operator on line 84
+
+**File**: `R/UTCwDST.R`
+**Change**:
+```diff
+- pTZs = GuessTZ(aOF = aOF, fork = fork)
++ pTZs <- GuessTZ(aOF = aOF, fork = fork)
+```
+
+**Rationale**: '<-' is the standard R assignment operator and is preferred for consistency and readability.
+
+---
+
+### 3. TRUE/FALSE Instead of T/F ✅
+
+**Issue**: Use of 'T' and 'F' as logical values  
+**Action**: Replaced all instances with TRUE and FALSE
+
+#### Files Changed:
+1. **R/boot.seci.R** (lines 184-185)
+   ```diff
+   - Est <- unlist (lapply (boot.df, mean, na.rm = T))
+   - SEs <- unlist (lapply (boot.df, sd, na.rm = T))
+   + Est <- unlist (lapply (boot.df, mean, na.rm = TRUE))
+   + SEs <- unlist (lapply (boot.df, sd, na.rm = TRUE))
+   ```
+
+2. **R/TimeFormat.R** (line 74, 135)
+   ```diff
+   - DateF <- suppressWarnings (tryCatch (DateFormat (Time, as.date = F)))
+   + DateF <- suppressWarnings (tryCatch (DateFormat (Time, as.date = FALSE)))
+   
+   - x <- TimeFormat (T [[i]], as.time = F)
+   + x <- TimeFormat (T [[i]], as.time = FALSE)
+   ```
+
+3. **R/TravelLog.R** (line 52)
+   ```diff
+   - write.csv (df, paste0 (Dir, "/TravelLog_Template.csv"), row.names = F)
+   + write.csv (df, paste0 (Dir, "/TravelLog_Template.csv"), row.names = FALSE)
+   ```
+
+4. **R/write.cosinor.R** (line 263)
+   ```diff
+   - utils::write.csv (Bdf, BdfDir, row.names = F)
[…]
