### 2025-12-11 ecf9f0e
Fix additional = assignment in UTCwDST.R:106
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
