### 2025-12-11 7dcb6fc
Add executive summary of all changes
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
