+++ CHANGES_SUMMARY.md
@@ -1,0 +1,58 @@
+# ActiGlobe pkgcheck Fixes - Summary
+
+## Overview
+This branch (`copilot/improve-code-based-on-pkgcheck`) addresses code quality issues identified by the pkgcheck GitHub Action workflow.
+
+## Files Changed
+- **14 files total**
+- **12 R source files** modified for code quality improvements
+- **1 DESCRIPTION** file modified
+- **1 comprehensive documentation report** created
+
+## Key Improvements
+
+### Code Quality Enhancements
+✅ **Type Safety**: Replaced 9 instances of `sapply()` with type-safe `vapply()`  
+✅ **Best Practices**: Fixed 9 instances of T/F to TRUE/FALSE  
+✅ **Safe Iteration**: Changed `1:length()` to `seq_along()`  
+✅ **Standard Operators**: Fixed assignment from `=` to `<-`  
+✅ **Code Readability**: Broke 5 critical long lines (>80 chars)  
+✅ **Package Metadata**: Removed unnecessary Date field from DESCRIPTION
+
+### Documentation
+✅ **Comprehensive Report**: Created `PKGCHECK_FIXES_REPORT.md` with:
+- Detailed explanation of each fix
+- Before/after code comparisons
+- Rationale for each change
+- Future recommendations
+
+## Testing Notes
+All changes are backward-compatible:
+- No function signatures changed
+- No return types altered
+- Only internal implementation improved
+- No new dependencies added
+
+## What's Not Fixed (Documented for Future Work)
+⚠️ **High Cyclomatic Complexity**: The `ggCosinorM` function (complexity: 74) requires substantial refactoring beyond minimal changes  
+📝 **Some Long Lines**: Documentation/comment lines remain >80 chars  
+📝 **Some sapply()**: Complex instances with variable return types remain
+
+## Commits in This Branch
+1. Initial plan
+2. Fix basic pkgcheck issues: remove Date, fix =, T/F, sapply
+3. Fix additional code style issues: break long lines  
+4. Add comprehensive documentation report for pkgcheck fixes
+
+## Next Steps
+1. Review the changes
+2. Run pkgcheck again to verify improvements
+3. Merge to main branch if approved
+4. Consider future refactoring of ggCosinorM function
+
+## Statistics
+- Lines added: 316
+- Lines removed: 25
+- Net change: +291 lines (mostly documentation)
+- Code changes: Minimal and targeted
+- Documentation: Comprehensive and detailed
