### 2025-12-12 7a469d8
Address review feedback: revert lapply to sapply in GuessTZ.R, change integer to numeric in UTCwDST.R
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
