### 2025-12-12 6fd33a1
Update pkgcheck.yaml
