+++ .github/workflows/pkgcheck.yaml
@@ -12,7 +12,6 @@ on:
   push:
     branches:
       - main
-      - Test
 
 jobs:
   pkgcheck:
