+++ tests/testthat/helper-Demand.R
@@ -8,7 +8,7 @@ test_that("Demand returns proper response", {
   expect_equal(result, "Option A")
 
 
-# Second Test -----------------------------
+  # Second Test -----------------------------
   # mock readline to always return "2"
   mockery::stub(Demand, "readline", function(prompt = "") "2")
 
