+++ man/Date2TotalT.Rd
@@ -7,17 +7,18 @@
 Date2TotalT(DT, TUnit = "hour", TZ = "local")
 }
 \arguments{
-\item{DT}{A vector of class `Date`, `POSIXct` or `POSIXlt`, sorted in ascending order.
-Each entry represents the start of an interval.  Internally, the function
-appends one extra day beyond the last entry so that the final interval
-covers a full 24 h until the next midnight. See \code{as.POSIXct}.}
+\item{DT}{A vector of class `Date`, `POSIXct` or `POSIXlt`, sorted in ascending
+order. Each entry represents the start of an interval.  Internally, the
+function appends one extra day beyond the last entry so that the final
+interval covers a full 24 h until the next midnight. See \code{as.POSIXct}.}
 
 \item{TUnit}{Character string specifying the unit for the output durations.
 Must be one of `"hour"`, `"minute"`, or `"second"`.  Comparison is
 case-insensitive.  Default is `"hour"`, meaning the returned values
 are in decimal hours.}
 
-\item{TZ}{The time zone when the recording started. (default = "local", which means user's local time zone)}
+\item{TZ}{The time zone when the recording started. (default = "local",
+which means user's local time zone)}
 }
 \value{
 A numeric vector of the same length as `DT`.  Each element is the
