+++ man/TravelLog.Rd
@@ -7,11 +7,14 @@
 TravelLog(Write = FALSE, Dir = NULL)
 }
 \arguments{
-\item{Write}{A binary code to indicate whether to write a .csv file containing the template needed for the travel log. (default = FLASE)
+\item{Write}{A binary code to indicate whether to write a .csv file
+containing the template needed for the travel log. (default = FLASE)
 When set to FALSE, the template will be returned as an object.
-When set to TRUE, user must provide the target directory where they wish to store the template.}
+When set to TRUE, user must provide the target directory where they wish to
+store the template.}
 
-\item{Dir}{The directory where the travel log template to be exported <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}
+\item{Dir}{The directory where the travel log template to be exported <e.g.
+"C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}
 }
 \value{
 a travel-log template as a data.frame or written as a CSV file

[…]
