### 2025-12-12 5c94faf
Fix test structure: move Prob.Inact test from helper to test file, replace sapply with vapply in Act2Daily
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
