### 2025-12-12 836755a
Fix timezone issue in BriefSum and replace sapply with vapply in ValInput
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
