### 2025-12-12 8d6c37e
Update Act2Daily.R
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
