+++ R/C2T.R
@@ -23,7 +23,7 @@
 #' directly. If the input is a time string, it is parsed into hours, minutes,
 #' and seconds, then converted to decimal hours relative to the first entry.
 #'
-#' @param Time A vector of time values. Can be numeric, character, or factor.
+#' @param time A vector of time values. Can be numeric, character, or factor.
 #'   Examples include `"13:45:00"` or `"01:30:15"`. If numeric, values are
 #'   returned as-is (after coercion). If character/factor, values are parsed
 #'   using \code{\link{TimeFormat}}.
@@ -43,55 +43,86 @@
 #' @examples
 #'
 #' # Character input
-#' times <- c ("01:00:00", "02:30:00", "03:15:00")
-#' C2T (times, Discrete = TRUE)
+#' times <- c("01:00:00", "02:30:00", "03:15:00")
+#' C2T(times, Discrete = TRUE)
 #'
 #' @noRd
 
-C2T <- function (Time, Discrete = FALSE) {
-    # Check Point ------------------------
-    if (any(grepl("^[A-Za-z]+$", Time)) || any(is.numeric (Time)))
-        stop ("Input 'Time' must be a pure character time string with no timezone label.
-              Please, check the input using TimeFormat().")
+C2T <- function(time, Discrete = FALSE) {
+  # Check Point ------------------------
+  if (any(grepl("^[A-Za-z]+$", time)) || any(is.numeric(time))) {
+    stop("Input 'time' must be a pure character time string with no
+    timezone label. Please, check the input using TimeFormat().")
+  }
 
 
+  # First Attempt to Coerce to Numeric -------------
+  x <- suppressWarnings(as.numeric(as.character(time)))
 
-    # First Attempt to Coerce to Numeric -------------
-    x <- suppressWarnings (as.numeric (as.character (Time)))
+  # If All NAs, Parse as Time Strings -------------
+  if (length(na.omit(x)) == 0) {
+    Fmt <- TimeFormat(time, as.time = FALSE)
 
-    # If All NAs, Parse as Time Strings -------------
-    if (length (na.omit (x)) == 0) {
-        Fmt <- TimeFormat (Time, as.time = FALSE)
-        hms <- as.POSIXct (Time, format = Fmt)
-
-        decimal_hours <- as.numeric (format (hms, "%H")) +
-            as.numeric (format (hms, "%M")) / 60 +
-            as.numeric (format (hms, "%S")) / 3600
-
-
-        x <- as.numeric (decimal_hours)
-
-    }
-
-    # Adjust for Initial Time if Needed -------------
-    if (!Discrete){
-
-        ini <- x [[1]]
-        x <- x - ini ## For duration
-    }
-
-    if (any(x < 0) || any(x > 24))
-        stop ("Negative or uut of range (0-24) value detected.
-              Please check your input 'Time' values.")
+    decimal_hours <- vapply(time, function(Tm)
+      ParseT(time = Tm, fmt = Fmt),
+      numeric(1)
+    )
 
 
+    x <- as.numeric(decimal_hours)
+  }
 
-    if (any (is.na (x))) warning (paste0 ("NAs introduced by coercion"))
+  # Adjust for Initial Time if Needed -------------
+  if (!Discrete) {
+    ini <- x[[1]]
+    x <- x - ini ## For duration
+  }
+
+  if (any(x < 0) || any(x > 24)) {
+    stop("Negative or uut of range (0-24) value detected.
+              Please check your input 'time' values.")
+  }
 
 
-    # Return Result ------------------------
-    return (x)
+  if (any(is.na(x))) warning(paste0("NAs introduced by coercion"))
+
+
[…]
