### 2025-12-12 ad953a5
Update NEWS.md
