### 2025-12-12 6beb22a
Refactor ggCosinorM into smaller helper functions
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
