+++ R/ggCosinorM.R
@@ -75,10 +75,8 @@
 #' @keywords plot ggplot graph cosinor
 #' @export
 
-ggCosinorM <- function (object, labels = TRUE, ci = TRUE, ci_level = 0.95,
-                        n = 400, point_size = 0.5, title_extra = NULL, legend.position = "right",
-                        ...) {
-    # Accept both parametric CosinorM and KDE-based CosinorM.KDE ----------
+#' @keywords internal
+.validate_ggcosinorm_inputs <- function(object, ci_level, n) {
     if (!inherits (object, c ("CosinorM", "CosinorM.KDE")))
         stop ("ggCosinorM: object must be a CosinorM or CosinorM.KDE fit.", call. = FALSE)
 
@@ -87,15 +85,10 @@ ggCosinorM <- function (object, labels = TRUE, ci = TRUE, ci_level = 0.95,
 
     if (!is.numeric (n) | n <= 0 | n != round (n))
         stop ("N must be a positive integer")
+}
 
-
-    # Check Essential Parameters ----------------------
-    day <- 24
-
-    tau <- if (!is.null (object$tau)) object$tau else 24
-    nT <- length (tau)
-
-    # Observed data --------------
+#' @keywords internal
+.extract_observed_data <- function(object) {
     if (!is.null (object$model) && !is.null (object$model$time)) {
         t_obs <- object$model$time
         y_obs <- if (!is.null (object$model$activity)) {
@@ -106,145 +99,181 @@ ggCosinorM <- function (object, labels = TRUE, ci = TRUE, ci_level = 0.95,
     } else {
         stop ("Object missing model$time/activity for observed data.", call. = FALSE)
     }
-    aug <- data.frame (t_obs = t_obs, y_obs = y_obs)
+    data.frame (t_obs = t_obs, y_obs = y_obs)
+}
 
-    # Parameter source selection -----------------
-    coef_cos <- object$coef.cosinor
-    post <- object$post.hoc
-
-    use_parametric_single <- inherits (object, "CosinorM") && nT == 1
-    use_posthoc <- !use_parametric_single
-
-    # MESOR ------------------
+#' @keywords internal
+.get_mesor <- function(object, use_posthoc, coef_cos, post) {
     if (use_posthoc) {
-        if (is.null (post) || is.null (post ["MESOR.ph"])) stop ("Post-hoc MESOR.ph is required.", call. = FALSE)
-        mesor <- as.numeric (post ["MESOR.ph"])
+        if (is.null (post) || is.null (post ["MESOR.ph"])) 
+            stop ("Post-hoc MESOR.ph is required.", call. = FALSE)
+        as.numeric (post ["MESOR.ph"])
     } else {
-        if (is.null (coef_cos) || is.null (coef_cos ["MESOR"])) stop ("Parametric MESOR is required.", call. = FALSE)
-        mesor <- as.numeric (coef_cos ["MESOR"])
+        if (is.null (coef_cos) || is.null (coef_cos ["MESOR"])) 
+            stop ("Parametric MESOR is required.", call. = FALSE)
+        as.numeric (coef_cos ["MESOR"])
     }
+}
 
-    # Amplitude -----------
+#' @keywords internal
+.get_amplitude <- function(object, use_posthoc, coef_cos, post, tau) {
     if (use_posthoc) {
-        if (is.null (post) || is.null (post ["Amplitude.ph"])) stop ("Post-hoc Amplitude.ph is required.", call. = FALSE)
-        amplitude <- as.numeric (post ["Amplitude.ph"])
+        if (is.null (post) || is.null (post ["Amplitude.ph"])) 
+            stop ("Post-hoc Amplitude.ph is required.", call. = FALSE)
[…]
