### 2025-12-12 0b55c48
Fix condition to maintain backward compatibility
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
