### 2025-12-12 1f45a94
Improve inline documentation following template (80 char limit)
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
