+++ R/ggCosinorM.R
@@ -20,21 +20,38 @@
 #' @title Plot `CosinorM` and `CosinorM.KDE` Fit with `ggplot2`
 #'
 #' @description
-#' Create visualization of a CosinorM or CosinorM.KDE fit using ggplot2. The plot shows the parametric cosinor fit over a fine time grid, optional pointwise confidence bands, observed data points, MESOR line, acrophase verticals, amplitude annotation segments, and labelled parameter values when requested.
+#' Create visualization of a CosinorM or CosinorM.KDE fit using
+#' ggplot2. The plot shows the parametric cosinor fit over a fine time
+#' grid, optional pointwise confidence bands, observed data points,
+#' MESOR line, acrophase verticals, amplitude annotation segments, and
+#' labelled parameter values when requested.
 #'
 #' @import stats ggplot2 ggrepel viridis
 #' @importFrom scales alpha
-#' @param object A fitted model of class \code{\link{CosinorM}} or \code{\link{CosinorM.KDE}}
-#' @param labels Logical; Default `TRUE` places repelled labels on the plot with MESOR, amplitude(s), and acrophase(s).
-#' @param ci Logical; Default `TRUE` computes and draws pointwise parametric confidence bands for the fitted cosinor curve using the model covariance.
-#' @param ci_level Confidence level for the pointwise bands, expressed in numeric value between 0 and 1.
-#' @param n Integer; number of points on the fine prediction grid used to draw the fitted cosinor and confidence ribbon. (default to `400`)
-#' @param point_size Numeric; plotting size for observed points. (default: `0.5`)
-#' @param title_extra Optional character string appended to the plot title for extra context.
-#' @param legend.position Position of the legend on the plot; default is `"right"`. Other options include `"top"`, `"bottom"`, `"left"`, or a numeric vector of length two specifying x and y coordinates.
-#' @param ... Additional arguments (currently ignored) kept for future update
 #'
-#' @return A `ggplot` object representing the cosinor model fit visualization.
+#' @param object A fitted model of class \code{\link{CosinorM}} or
+#'   \code{\link{CosinorM.KDE}}.
+#' @param labels Logical; Default `TRUE` places repelled labels on
+#'   the plot with MESOR, amplitude(s), and acrophase(s).
+#' @param ci Logical; Default `TRUE` computes and draws pointwise
+#'   parametric confidence bands for the fitted cosinor curve using
+#'   the model covariance.
+#' @param ci_level Confidence level for the pointwise bands, expressed
+#'   in numeric value between 0 and 1 (default: `0.95`).
+#' @param n Integer; number of points on the fine prediction grid used
+#'   to draw the fitted cosinor and confidence ribbon (default: `400`).
+#' @param point_size Numeric; plotting size for observed points
+#'   (default: `0.5`).
+#' @param title_extra Optional character string appended to the plot
+#'   title for extra context.
+#' @param legend.position Position of the legend on the plot; default
+#'   is `"right"`. Other options include `"top"`, `"bottom"`, `"left"`,
+#'   or a numeric vector of length two specifying x and y coordinates.
+#' @param ... Additional arguments (currently ignored) kept for future
+#'   update.
+#'
+#' @returns
+#' A `ggplot` object representing the cosinor model fit visualization.
 #'
 #' @examples
 #' \dontrun{
@@ -71,14 +88,24 @@
 #' }
 #'
[…]
