### 2025-12-12 306b907
Address code review feedback: remove extra line, merge theme calls, pass day parameter
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
