+++ R/CosinorM.KDE.R
@@ -29,32 +29,47 @@
 #' Beta, Gamma, Amplitude, Acrophase) are obtained by numerical integration of
 #' the KDE using trapezoid-rule quadrature.
 #'
-#' Cosinor parameters are obtained by numerical integration of the KDE on the grid:
+#' Cosinor parameters are obtained by numerical integration of the
+#' KDE on the grid:
 #' \itemize{
 #'   \item MESOR: mean level of the fitted curve, \eqn{M = I_0 / (2\pi)}.
 #'   \item \eqn{\beta}: cosine coefficient, \eqn{\beta = I_{cos} / \pi}.
 #'   \item \eqn{\gamma}: sine coefficient, \eqn{\gamma = I_{sin} / \pi}.
-#'   \item Amplitude: magnitude of the first harmonic, \eqn{A = \sqrt{\beta^2 + \gamma^2}}.
-#'   \item Acrophase: peak time, \eqn{\phi = atan2(-\gamma, \beta)}, expressed in radians and converted to hours.
+#'   \item Amplitude: magnitude of the first harmonic,
+#'     \eqn{A = \sqrt{\beta^2 + \gamma^2}}.
+#'   \item Acrophase: peak time, \eqn{\phi = atan2(-\gamma, \beta)},
+#'     expressed in radians and converted to hours.
 #' }
 #'
-#' The grid integrals are computed with trapezoid weights \eqn{w_g}:
+#' The grid integrals are computed with trapezoid weights
+#' \eqn{w_g}:
 #' \deqn{I_0 = \text{area}_g \sum_g \frac{pdf_g \, w_g}{den_g}}
-#' \deqn{I_{cos} = \text{area}_g \sum_g \frac{pdf_g \cos(\theta_g)\, w_g}{den_g}}
-#' \deqn{I_{sin} = \text{area}_g \sum_g \frac{pdf_g \sin(\theta_g)\, w_g}{den_g}}
+#' \deqn{I_{cos} = \text{area}_g \sum_g
+#'   \frac{pdf_g \cos(\theta_g)\, w_g}{den_g}}
+#' \deqn{I_{sin} = \text{area}_g \sum_g
+#'   \frac{pdf_g \sin(\theta_g)\, w_g}{den_g}}
 #'
-#' where \eqn{pdf_g} is the normalized KDE at grid angle \eqn{\theta_g}, \eqn{den_g}
-#' is the kernel denominator at that grid point, and \eqn{\text{area}_g} rescales the
-#' fitted density back to the activity scale.
+#' where \eqn{pdf_g} is the normalized KDE at grid angle
+#' \eqn{\theta_g}, \eqn{den_g} is the kernel denominator at that
+#' grid point, and \eqn{\text{area}_g} rescales the fitted density
+#' back to the activity scale.
 #'
 #' Notes:
 #' \itemize{
-#'   \item Normalization factors \eqn{1/\pi} and \eqn{1/(2\pi)} follow from the orthogonality of sine and cosine on \eqn{[0,2\pi)}.
-#'   \item The KDE summarizes the first harmonic of the rest–activity pattern over one cycle; higher harmonics are not extracted.
-#'   \item The trapezoid rule ensures numerical stability by integrating over a dense, evenly spaced grid.
-#'   \item The dense grid excludes the duplicate \eqn{2\pi} endpoint (grid in \eqn{[0,2\pi)}) to avoid duplicated quadrature nodes.
-#'   \item Grid and observation trapezoid weights correct for irregular sampling and yield stable approximations to continuous integrals.
[…]
