### 2025-12-13 f1f4165
Fix data.R documentation and update R-CMD-check workflow
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
