+++ .github/workflows/R-CMD-check.yaml
@@ -6,7 +6,7 @@
 # usethis::use_github_action("check-standard") will install it.
 on:
   push:
-    branches: [main, master]
+    branches: [main, master, copilot/refactor-ggcosinorm-function]
   pull_request:
 
 name: R-CMD-check.yaml
