### 2025-12-13 73a4b78
Fix CosinorM.R documentation: all lines now ≤80 chars
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
