+++ R/CosinorM.R
@@ -22,16 +22,19 @@
 #'
 #'
 #' @details
-#' The `Cosinor` model is a cosine-based harmonic regression used to estimate circadian rhythm parameters.
+#' The `Cosinor` model is a cosine-based harmonic regression used
+#' to estimate circadian rhythm parameters.
 #'
 #' \strong{Single-phase equation}:
 #'       \deqn{y = M + A \cos \left(\frac{2\pi t}{\tau} - \phi\right) }
 #' \itemize{
-#'   \item \eqn{M }: MESOR (mid-line estimating statistic of rhythm), the intercept
+#'   \item \eqn{M }: MESOR (mid-line estimating statistic of
+#'     rhythm), the intercept
 #'   \item \eqn{A }: Amplitude, peak deviation from M
 #'   \item \eqn{t }: Time coordinate within the cycle
 #'   \item \eqn{\tau }: The assumed period length
-#'   \item \eqn{\phi }: Acrophase (time-of-peak), computed from fitted sine and cosine coefficients
+#'   \item \eqn{\phi }: Acrophase (time-of-peak), computed from
+#'     fitted sine and cosine coefficients
 #' }
 #'
 #'
@@ -45,7 +48,8 @@
 #' \strong{Linearized form}:
 #'      \deqn{\hat{y} = M + \beta x + \gamma z + \epsilon }
 #' \itemize{
-#'   \item \eqn{\beta = A * cos(\phi)}, estimated coefficient for the cosine term
+#'   \item \eqn{\beta = A * cos(\phi)}, estimated coefficient for
+#'     the cosine term
 #'   \item \eqn{x = cos(\frac{2 * \pi * t} {\tau})}, cosine-transformed time
 #'   \item \eqn{\gamma = A * sin(\phi)}, estimated coefficient for the sine term
 #'   \item \eqn{z = sin(\frac{2 * \pi * t} {\tau})}, sine-transformed time
@@ -63,8 +67,11 @@
 #'
 #'  Note,
 #' \itemize{
-#'   \item \eqn{\phi } can be converted to clock time to identify the peak activity time AFTER midnight.
-#'   \item if the model is not a mono-phase (i.e., more than one \eqn{\tau }), there will be \eqn{\frac{24 (hours)} {\tau} } number of peaks within a day.
+#'   \item \eqn{\phi } can be converted to clock time to identify
+#'     the peak activity time AFTER midnight.
+#'   \item if the model is not a mono-phase (i.e., more than one
+#'     \eqn{\tau }), there will be \eqn{\frac{24 (hours)} {\tau} }
+#'     number of peaks within a day.
 #' }
 #'
 #' \strong{Amplitude \eqn{(A) } estimation}:
@@ -77,22 +84,31 @@
 #'
 #' @param time Numeric vector of time coordinates for each data point
 #' @param activity Numeric vector of activity counts from an actigraphy device
-#' @param tau Numeric scalar or vector for the assumed circadian period. Default is 24 for single-phase; multiple phases can be supplied via [c()]
+#' @param tau Numeric scalar or vector for the assumed circadian
+#'   period. Default is 24 for single-phase; multiple phases can be
+#'   supplied via [c()]
 #' @param method Character string specifying estimation method
 #' \itemize{
 #'   \item "OLS": Ordinary least squares via \code{\link[stats]{lm}} (default)
-#'   \item "FGLS": Feasible generalized least squares; models heteroskedasticity via a log-variance fit to squared OLS residuals, computes weights, and refits by weighted least squares
+#'   \item "FGLS": Feasible generalized least squares; models
[…]
