### 2025-12-13 34c441f
Fix workflow and documentation for se.CosinorM.R and write.cosinor.R
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
