+++ R/write.cosinor.R
@@ -21,35 +21,63 @@
 #' @title Generate Report for Daily Cosinor and Generate Summary Plot
 #'
 #' @description
-#' A multi-wrapper function which generate data trend using \code{\link[ggplot2]{geom_smooth}} and model plot using \code{\link{CosinorM}} or \code{\link{CosinorM.KDE}}. All daily results could be found in the exported report. A summary file in .csv format is also generated to update the original `Bdf` with cosinor model coefficients for each day.
+#' A multi-wrapper function which generate data trend using
+#' \code{\link[ggplot2]{geom_smooth}} and model plot using
+#' \code{\link{CosinorM}} or \code{\link{CosinorM.KDE}}. All daily
+#' results could be found in the exported report. A summary file in
+#' .csv format is also generated to update the original `Bdf` with
+#' cosinor model coefficients for each day.
 #'
 #' @import gridExtra ggplot2 grDevices utils stats
 #'
-#' @param Dir The directory where the recordings to be exported <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">
+#' @param Dir The directory where the recordings to be exported
+#'   <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../
+#'   FOLDER NAME/">
 #' @param ID The subject's ID which would be used to create a folder.
 #' @param DailyAct A data.frame list generated by `Act2Daily()`
-#' @param Bdf A \code{\link{BriefSum}} object. Note, if jet lag occurred during the recording, please, update the metadata using \code{\link{TAdjust}} before passing to this function.
-#' @param VAct Optional character. Name of the activity column in each data.frame object (`df`) in `DailyAct`. If NULL,
-#'   defaults to the second column of `df`.
-#' @param VTm Optional character. Name of the activity column in each data.frame object (`df`) in `DailyAct`. If NULL,
-#'   defaults to the first column of `df`.
+#' @param Bdf A \code{\link{BriefSum}} object. Note, if jet lag
+#'   occurred during the recording, please, update the metadata using
+#'   \code{\link{TAdjust}} before passing to this function.
+#' @param VAct Optional character. Name of the activity column in
+#'   each data.frame object (`df`) in `DailyAct`. If NULL, defaults
+#'   to the second column of `df`.
+#' @param VTm Optional character. Name of the activity column in each
+#'   data.frame object (`df`) in `DailyAct`. If NULL, defaults to the
+#'   first column of `df`.
 #' @param method Character string specifying estimation method
 #' \itemize{
-#'   \item "OLS": Ordinary least square cosinor model via \code{\link{CosinorM}}. (default)
-#'   \item "FGLS": Feasible generalized least squares cosinor model via \code{\link{CosinorM}}.
-#'   \item "KDE": Gaussian kernel density estimate (KDE) on circular time (hours of day) via \code{\link{CosinorM.KDE}}.
+#'   \item "OLS": Ordinary least square cosinor model via
+#'     \code{\link{CosinorM}}. (default)
+#'   \item "FGLS": Feasible generalized least squares cosinor model
+#'     via \code{\link{CosinorM}}.
+#'   \item "KDE": Gaussian kernel density estimate (KDE) on circular
+#'     time (hours of day) via \code{\link{CosinorM.KDE}}.
 #' }
-#' @param tau Numeric scalar or vector for the assumed circadian period, passed to \code{\link{CosinorM}}. Default is 24 for single-phase; multiple phases can be supplied via [c()].
-#' @param ph Optional logical scaler; if only a single period is assigned to tau, TRUE will force the function to append the post-hoc circadian estimates instead of the model estimates. (default: FALSE)
[…]
