### 2025-12-13 3f2acda
Fix documentation for BriefSum.R, Act2Daily.R, and write.act.R
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
