+++ R/Act2Daily.R
@@ -31,34 +31,45 @@
 #' which represents "time lose".
 #'
 #'
-#' @details Transforms a continuous actigraphy time-series into a day-by-day list
-#'  of aligned epoch and produces a fully annotated data frame. Based on the
-#'  summary table generated by either \code{\link{BriefSum}} or \code{\link{TAdjust}}, it re-organize and
-#'  segment the raw recording data by each day. The function is designed to
-#'  process incomplete recordings while accounting for travel and daylight-saving
-#'  adjustments. All labels contained in the summary table (i.e., Bdf) will be
-#'  assigned to the corresponding data points. This workflow ensures that each day's recording is bounded by its
-#'  calendar date while accounts for travel--induced overlaps, and clearly documents any incomplete or unallocated points.
+#' @details Transforms a continuous actigraphy time-series into a
+#'  day-by-day list of aligned epoch and produces a fully annotated
+#'  data frame. Based on the summary table generated by either
+#'  \code{\link{BriefSum}} or \code{\link{TAdjust}}, it re-organize
+#'  and segment the raw recording data by each day. The function is
+#'  designed to process incomplete recordings while accounting for
+#'  travel and daylight-saving adjustments. All labels contained in
+#'  the summary table (i.e., Bdf) will be assigned to the
+#'  corresponding data points. This workflow ensures that each day's
+#'  recording is bounded by its calendar date while accounts for
+#'  travel--induced overlaps, and clearly documents any incomplete or
+#'  unallocated points.
 #'
 #'
-#' @param df A data.frame of raw actigraphy recording. Both time and activity count
-#' should be included in the `df`. See `VAct` and `VTm` for further detail.
-#' @param Bdf A \code{\link{BriefSum}} object containing per-day metadata for the recording. Note, if jet lag occurred during the recording, please, update the metadata using \code{\link{TAdjust}} before passing to this function.
+#' @param df A data.frame of raw actigraphy recording. Both time
+#'   and activity count should be included in the `df`. See `VAct`
+#'   and `VTm` for further detail.
+#' @param Bdf A \code{\link{BriefSum}} object containing per-day
+#'   metadata for the recording. Note, if jet lag occurred during the
+#'   recording, please, update the metadata using
+#'   \code{\link{TAdjust}} before passing to this function.
 #' @param TUnit Character; time--unit for the x--axis of each day's timeline.
 #'   Must be one of `day`, `hour`, `minute` or `second`.  Default is `hour`.
 #' @param VAct Optional character. Name of the activity column in `df`. If NULL,
 #'   defaults to the second column of `df`.
-#' @param VTm Optional character.  Name of the date.time index column in `df`. If NULL,
-#'   defaults to the first column of `df`.
-#' @param Incomplete Logical; if TRUE, days flagged `Incomplete Recording` (i.e. <24 h)
-#'   are retained in the data list with recordings segmented by day. Default = FALSE (these days are removed).
-#' @param Travel Logical; if TRUE, days flagged `Travel` are retained although
-#'   some data points from an earlier adjacent calendar day may be duplicated (a
-#'   warning is issued). If FALSE, travel days and the day before/after are excluded.
-#'   Default = TRUE.
+#' @param VTm Optional character.  Name of the date.time index
+#'   column in `df`. If NULL, defaults to the first column of `df`.
+#' @param Incomplete Logical; if TRUE, days flagged `Incomplete
+#'   Recording` (i.e. <24 h) are retained in the data list with
+#'   recordings segmented by day. Default = FALSE (these days are
+#'   removed).
+#' @param Travel Logical; if TRUE, days flagged `Travel` are retained
+#'   although some data points from an earlier adjacent calendar day
[…]
