### 2025-12-13 2b942b7
Fix documentation for Prob.Inact.R, TAdjust.R, ggActiGlobe.R
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
