+++ R/ggActiGlobe.R
@@ -18,19 +18,26 @@
 #' @title Plot an Overview of an `ActiGlobe` Activity Time Series
 #'
 #' @description
-#' Creates a time-series `scatterplot` of activity counts from an `ActiGlobe` data.frame,marking each midnight boundary with a vertical dashed line and `coloring` points that were flagged (e.g. travel overlaps or unallocated epochs).
+#' Creates a time-series `scatterplot` of activity counts from an
+#' `ActiGlobe` data.frame,marking each midnight boundary with a
+#' vertical dashed line and `coloring` points that were flagged (e.g.
+#' travel overlaps or unallocated epochs).
 #'
 #' @import ggplot2
 #'
 #' @param df A data.frame of annotated actigraphy epochs.  Must include:
 #'   - An activity column named by `VAct`.
 #'   - A datetime column named by `VDT`.
-#'   - Optionally, a `Note` column to flag affected epochs. See \code{\link{BriefSum}} and \code{\link{TAdjust}} for details.
-#' @param Bdf A \code{\link{BriefSum}} object containing per-day metadata for the recording. Note, if jet lag occurred during the recording, please, update the metadata using \code{\link{TAdjust}} before passing to this function.
-#' @param VAct Optional character. Name of the activity column in `df`. If NULL,
-#'   defaults to the second column of `df`.
-#' @param VDT Optional character. Name of the POSIXct datetime column in `df`. If NULL,
-#'   defaults to "DateTime" of `df`.
+#'   - Optionally, a `Note` column to flag affected epochs. See
+#'   \code{\link{BriefSum}} and \code{\link{TAdjust}} for details.
+#' @param Bdf A \code{\link{BriefSum}} object containing per-day
+#'   metadata for the recording. Note, if jet lag occurred during the
+#'   recording, please, update the metadata using
+#'   \code{\link{TAdjust}} before passing to this function.
+#' @param VAct Optional character. Name of the activity column in
+#'   `df`. If NULL, defaults to the second column of `df`.
+#' @param VDT Optional character. Name of the POSIXct datetime column
+#'   in `df`. If NULL, defaults to "DateTime" of `df`.
 #'
 #' @return A `ggplot` object showing:
 #'   - Activity counts vs. time.
@@ -73,7 +80,8 @@
 #' )
 #' print (p2)
 #'
-#' # Pro-tip: [`cowplot`] can help stack the time series graphs in one single plot
+#' # Pro-tip: [`cowplot`] can help stack the time series graphs in
+#' # one single plot
 #' }
 #'
 #' @keywords visualization actigraphy
