### 2025-12-13 477fe8c
Complete documentation fixes for all remaining R files
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
