+++ R/Rad2Hr.R
@@ -23,7 +23,9 @@
 #' Convert `acrophase` in radian to clock time based on tau.
 #'
 #' @param x Numeric vector of phases in radian.
-#' @param tau Numeric scalar. Period length in the same time units you want returned (for hours use 24). Must be a single numeric value > 0 and <= 24.
+#' @param tau Numeric scalar. Period length in the same time units
+#'   you want returned (for hours use 24). Must be a single numeric
+#'   value > 0 and <= 24.
 #'
 #' @return Numeric vector of times in [0, tau). NA values propagate.
 #'
