+++ R/DST.R
@@ -28,7 +28,9 @@
 #'
 #' @param DT A vector of class `Date`, `POSIXct`, or `POSIXlt`, representing
 #'   the dates/times to test.  These need not be sorted.
-#' @param TZ A single string naming an IANA time zone (e.g. `America/New_York`). If `"local"`, it will extract local time zone using `Sys.timezone()`.  Must match one entry in
+#' @param TZ A single string naming an IANA time zone (e.g.
+#'   `America/New_York`). If `"local"`, it will extract local time
+#'   zone using `Sys.timezone()`.  Must match one entry in
 #'   `OlsonNames()`.
 #'
 #' @return
