+++ R/Num2UTC.R
@@ -21,10 +21,11 @@
 #' @title Convert Numbers to Standard UTC Offset
 #'
 #' @description
-#' Convert numeric values representing UTC offsets (in hours, with optional fractional
-#' parts for minutes) into standardized character strings of the form
-#' `"UTC+HH:MM"` or `"UTC-HH:MM"`. This is useful for translating numeric
-#' time zone offsets into human-readable UTC offset notation.
+#' Convert numeric values representing UTC offsets (in hours, with
+#' optional fractional parts for minutes) into standardized character
+#' strings of the form `"UTC+HH:MM"` or `"UTC-HH:MM"`. This is useful
+#' for translating numeric time zone offsets into human-readable UTC
+#' offset notation.
 #'
 #' @param x UTC Offsets <e.g., "UTC+09:30" or "UTC-07:00">
 #'
