### 2025-12-13 7e1b704
Fix DateFormat duplicate title, update pkgcheck.yaml, fix TAdjust inline comments
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
