+++ R/TAdjust.R
@@ -133,10 +133,12 @@ TAdjust <- function (Bdf, TLog, TZ = NULL, fork = FALSE) {
 
     ################## Adjust for Daylight Saving Changes #####################
 
-    ### Guess if the UTC may experience Time Change due to daylight saving calender
+    ### Guess if the UTC may experience Time Change due to DST
     pUDST <- UTCwDST (U)
-    SameUTC <- UTCs == U ### Determine if the original UTC from BriefSum is the same as the new UTC from R2P
-    nuUDST <- ifelse (SameUTC, aDST, pUDST) ### New logical indicator of whether daylight saving occurs...
+    ### Determine if original UTC from BriefSum matches new UTC from R2P
+    SameUTC <- UTCs == U
+    ### New logical indicator of whether daylight saving occurs
+    nuUDST <- ifelse (SameUTC, aDST, pUDST)
 
 
     ### If TZ is not specified, guess it.
@@ -153,13 +155,18 @@ TAdjust <- function (Bdf, TLog, TZ = NULL, fork = FALSE) {
 
 
     ### Step 1 Change NDPs
-    #### Based on the new daylight saving, we will change the NDPs and cumulative time...
-    A1 <- ifelse (nuUDST & GL0 == 0, 0, -1 * GL0) ### Adjusting factor for incorrect initial GL guesses
-    GL <- GL0 + A1 ### Update GL....
+    #### Based on new DST, change NDPs and cumulative time
+    ### Adjusting factor for incorrect initial GL guesses
+    A1 <- ifelse (nuUDST & GL0 == 0, 0, -1 * GL0)
+    ### Update GL
+    GL <- GL0 + A1
 
-    NDP <- ifelse (A1 == 0, nDP, nDP + (A1 * 3600 / Epc)) ### Remove the inappropriately adjusted gain or loss due to suspected time shift
-    mDP <- max (cumsum (nDP)) - max (cumsum (NDP)) ### Number of data points needed to be added.
-    NDP [length (NDP)] <- NDP [length (NDP)] + mDP ### Add the remainant to the last day. Leave the question whether it is more than one fDP to later step.
+    ### Remove inappropriately adjusted gain/loss from suspected shift
+    NDP <- ifelse (A1 == 0, nDP, nDP + (A1 * 3600 / Epc))
+    ### Number of data points needed to be added
+    mDP <- max (cumsum (nDP)) - max (cumsum (NDP))
+    ### Add the remainant to the last day
+    NDP [length (NDP)] <- NDP [length (NDP)] + mDP
 
 
     ### Step 2 Edit cumulative time based on NDPs
@@ -187,7 +194,8 @@ TAdjust <- function (Bdf, TLog, TZ = NULL, fork = FALSE) {
     y [y > LstP] <- NA
 
 
-    ### Adjust Cumulative Start DataPoint V2 will give you the same as the above------------
+    ### Adjust Cumulative Start DataPoint V2 (same as above)
+    ### ----------
     #  if ( Version == 2) {
     #    T2A = c(0,diff(H2J))
     #    x = a1
@@ -209,7 +217,8 @@ TAdjust <- function (Bdf, TLog, TZ = NULL, fork = FALSE) {
     Idxl <- length (DT)
     Nl <- as.integer (N [Idxl]) #### Last number of data
     Dl <- DT [Idxl] #### Last date of the recording
-    Tl <- as.numeric (as.POSIXct (Dl, tz = gTZ [Idxl])) ### The starting second of the last day in number
+    ### The starting second of the last day in number
[…]
