### 2025-12-13 afd2ba1
Fix inline comments in CosinorM.KDE.R (all ≤80 chars)
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
