+++ R/CosinorM.KDE.R
@@ -259,7 +259,8 @@
 #' @keywords circular cosinor KDE circadian
 #' @export
 
-CosinorM.KDE <- function (time, activity, bw = 0.8, grid = 360L, arctan2 = TRUE, dilute = FALSE) {
+CosinorM.KDE <- function (time, activity, bw = 0.8, grid = 360L,
+                          arctan2 = TRUE, dilute = FALSE) {
     # Check Point and Input Validation -------------------------
 	activity <- ValInput(x = activity, type = "Act")
 	time  <- ValInput(x = time, type = "Tm")
@@ -292,7 +293,8 @@ CosinorM.KDE <- function (time, activity, bw = 0.8, grid = 360L, arctan2 = TRUE,
     ### Trapzoid weight - radian on observation points
     w <- trap_weights (theta = theta)
 
-    ### Pairwise wrapped angular differences and kernel matrix (observations x observations)
+    ### Pairwise wrapped angular differences and kernel matrix
+    ### (observations x observations)
     diffs_mat <- wrap_diff (outer (theta, theta, "-")) # n x n
     bw_rad <- bw * 2 * pi / tau
     K <- dnorm (diffs_mat, mean = 0, sd = bw_rad) # kernel matrix (obs x obs)
@@ -304,14 +306,17 @@ CosinorM.KDE <- function (time, activity, bw = 0.8, grid = 360L, arctan2 = TRUE,
 
     ### integrate area on observation angles (coarse check)
     area <- trap_int (theta, y_h)
-    if (!is.finite (area) || area <= .Machine$double.eps) stop ("area is zero or invalid; check activity and bw")
+    if (!is.finite (area) || area <= .Machine$double.eps)
+        stop ("area is zero or invalid; check activity and bw")
 
     Pdf <- dens / area
 
 
     ##  Variance / se for fitted curve -----------------------------
-    W <- K * matrix (rep (w, each = nrow (K)), nrow = nrow (K), byrow = FALSE)
-    W <- W / matrix (rep (den, times = ncol (W)), nrow = nrow (W), byrow = FALSE)
+    W <- K * matrix (rep (w, each = nrow (K)), nrow = nrow (K),
+                     byrow = FALSE)
+    W <- W / matrix (rep (den, times = ncol (W)), nrow = nrow (W),
+                     byrow = FALSE)
 
     resid <- activity - y_h
     RS <- resid^2
@@ -320,14 +325,17 @@ CosinorM.KDE <- function (time, activity, bw = 0.8, grid = 360L, arctan2 = TRUE,
     trW <- sum (diag (W))
     trWW <- sum (W * W)
     df_denom <- n - (2 * trW) + trWW
-    if (df_denom <= 0) stop ("Nonpositive denominator for sigma^2 estimate; increase bw or check W")
+    if (df_denom <= 0)
+        stop ("Nonpositive denominator for sigma^2 estimate; ",
+              "increase bw or check W")
 
     sigma_hat <- RSS / df_denom
     var_fitted <- sigma_hat * rowSums (W^2)
     se_fitted <- sqrt (var_fitted)
 
 
-    # Evaluate on dense grid for stable integrals (grid points) --------------------
+    # Evaluate on dense grid for stable integrals (grid points)
+    # ----------
     # gtheta in radians, exclude duplicate 2*pi (2pi = 0)
     gtheta <- seq (0, 2 * pi, length.out = as.integer (grid) + 1)
     gtheta <- gtheta [-(length (gtheta))] # length = grid
[…]
