### 2025-12-13 159b548
Fix inline comments in write.cosinor.R (all ≤80 chars)
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
