### 2025-12-13 575baf7
Update helper-Demand.R
