+++ vignettes/Introduction.Rmd
@@ -1,0 +1,221 @@
+---
+title: "Introduction"
+subtitle: "What is actigraphy? How do we use R?"
+author: "C. William Yao"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Introduction}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+resource_files:
+  - ./images/Bdf_head.png
+  - ./images/Bdf_demo.png
+  - ./images/Bdf.adj_demo.png
+  - ./images/Unadjusted_Overview.png
+  - ./images/Adjusted_Overview.png
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+## What will be covered in this tutorial?
+
+This specific tutorial is designed to give users a gentle introduction
+to actigraphy and R. For tutorials on how to use ActiGlobe, please go
+to the rest of the tutorials in this package.
+
+## Load the Libraries
+
+```{r setup, eval=FALSE}
+library(ActiGlobe)
+
+### Optional Library
+# library(zeallot)
+```
+
+If any of the packages have yet to be installed, we can always install
+them using the function `install.packages()`.
+
+## Introduction
+
+### What is Actigraphy?
+
+Actigraphy is a non-invasive method used for measuring sleep/wake
+cycles by using a small device, usually in the form of a wrist-worn
+watch. It is the easiest way to measure sleep in the field and is
+widely used in the realm of sleep research due to its extended
+recording time and ease of use (Smith et al., 2018). An actigraphy
+device has a built-in accelerometer that measures movement, allowing
+researchers to objectively track activity patterns over days, weeks, or
+even months.
+
+### Understanding Accelerometers and Gyroscopes
+
+Modern actigraphy devices rely on sophisticated sensors to capture
+movement data:
+
+#### Accelerometers
+
+An **accelerometer** is a sensor that measures acceleration, or more
+simply, motion in one direction. It works by converting physical
+movement into electrical signals that the device can interpret and
+store. A 3-axis (or tri-axial) accelerometer measures acceleration
+along three perpendicular axes (X, Y, and Z), which correspond to
+different directions in three-dimensional space. This allows the device
+to track all types of motion, whether you are walking, running, or
+lying in bed.
+
+More specifically, an actigraph outputs the magnitude of acceleration
+(i.e., the vector magnitude of acceleration across all three axes),
+which is then used by algorithms to classify periods of activity versus
+rest. For instance, if you are out on a run, as your feet pound against
+the pavement and your arm swings back and forth, the device on your
+wrist will detect a sharp increase in the magnitude of acceleration,
+signaling that you are wide awake and on the move. In contrast, if you
+are curled up in bed fast asleep, the magnitude will greatly decrease
+and stabilize, allowing the device to easily discern that you are in a
+period of sleep or rest.
+
+#### Gyroscopes
+
+While accelerometers measure linear acceleration, **gyroscopes** measure
+angular velocity or rotational movement. Some advanced actigraphy
+devices incorporate gyroscopes to provide a more complete picture of
+movement. By combining data from both accelerometers and gyroscopes,
+these devices can better distinguish between different types of
+activities (e.g., walking vs. cycling) and reduce errors caused by
+device orientation.
+
+### On-Device Filtering and Data Acquisition
+
+To enable prolonged and uninterrupted data collection while managing
+storage limitations, most clinical and research-grade actigraphy devices
+implement **on-device filtering** and data processing:
+
+#### Signal Processing
+
+Raw acceleration signals from the accelerometer are typically sampled at
+high frequencies (e.g., 32 Hz or higher). However, storing raw
[…]
