### 2025-12-13 f1d426e
Add Introduction vignette for ActiGlobe package
This vignette introduces actigraphy, its applications, and the use of the ActiGlobe package for data analysis. It covers key concepts, library loading, and the importance of time zone adjustments in actigraphy research.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
