### 2025-12-13 854184d
Enable job cancellation on new workflow runs
Add concurrency settings to the R CMD check workflow

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
