### 2025-12-13 94bc77a
Update pkgcheck.yaml
