+++ .github/workflows/pkgcheck.yaml
@@ -2,7 +2,7 @@ name: pkgcheck
 
 # This will cancel running jobs once a new run is triggered
 concurrency:
-  group: ${{ github.workflow }}-${{ github.head_ref }}
+  group: ${{ github.workflow }}-${{ github.ref_name }}
   cancel-in-progress: true
 
 on:
