### 2025-12-15 2926ead
Act2Daily
