### 2025-12-15 0c5ce19
text length
