+++ R/Date2TotalT.R
@@ -25,10 +25,11 @@
 #' requested time unit.
 #'
 #' @param DT
-#'   A vector of class `Date`, `POSIXct` or `POSIXlt`, sorted in ascending
-#'   order. Each entry represents the start of an interval.  Internally, the
-#'   function appends one extra day beyond the last entry so that the final
-#'   interval covers a full 24 h until the next midnight. See \code{as.POSIXct}.
+#'   A vector of class `Date`, `POSIXct` or `POSIXlt`, sorted in
+#'   ascending order.
+#'   Each entry represents the start of an interval.  Internally, the function
+#'   appends one extra day beyond the last entry so that the final interval
+#'   covers a full 24 h until the next midnight. See \code{as.POSIXct}.
 #'
 #' @param TUnit
 #'   Character string specifying the unit for the output durations.
@@ -36,8 +37,8 @@
 #'   case-insensitive.  Default is `"hour"`, meaning the returned values
 #'   are in decimal hours.
 #'
-#' @param TZ The time zone when the recording started. (default = "local",
-#' which means user's local time zone)
+#' @param TZ The time zone when the recording started. (default =
+#'   "local", which means user's local time zone)
 #'
 #' @return
 #'   A numeric vector of the same length as `DT`.  Each element is the
@@ -48,20 +49,21 @@
 #' @examples
 #' \donttest{
 #' # Two calendar days: returns c(24, 24) hours
-#' Date2TotalT(as.Date(c("2021-01-01", "2021-01-02")), "hour")
+#' Date2TotalT (as.Date (c ("2021-01-01", "2021-01-02")), "hour")
 #'
 #'
 #' # Working in minutes
-#' Date2TotalT(as.POSIXct(c(
-#'   "2021-06-10 08:00:00",
-#'   "2021-06-10 14:30:00"
+#' Date2TotalT (as.POSIXct (c (
+#'     "2021-06-10 08:00:00",
+#'     "2021-06-10 14:30:00"
 #' )), "minute")
 #' }
 #'
 #' # In seconds (case-insensitive unit name)
-#' Date2TotalT(as.Date("2022-12-31"), "SeCoNd")
+#' Date2TotalT (as.Date ("2022-12-31"), "SeCoNd")
 #'
 #' @export
+
 Date2TotalT <- function(DT, TUnit = "hour", TZ = "local") {
   TZ <- ifelse(TZ == "local", Sys.timezone(), TZ)
 
