+++ R/TravelLog.R
@@ -19,21 +19,22 @@
 #'
 #' @import utils
 #'
-#' @param Write A binary code to indicate whether to write a .csv file
-#' containing the template needed for the travel log. (default = FLASE)
-#' When set to FALSE, the template will be returned as an object.
-#' When set to TRUE, user must provide the target directory where they wish to
-#' store the template.
-#' @param Dir The directory where the travel log template to be exported <e.g.
-#' "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">
+#' @param Write A binary code to indicate whether to write a .csv
+#'   file containing the template needed for the travel log. (default
+#'   = FLASE) When set to FALSE, the template will be returned as an
+#'   object. When set to TRUE, user must provide the target directory
+#'   where they wish to store the template.
+#' @param Dir The directory where the travel log template to be
+#'   exported <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../
+#'   FOLDER NAME/">
 #'
 #' @return a travel-log template as a data.frame or written as a CSV file
 #'
 #' @examples
 #'
-#' Tlg <- TravelLog(Write = FALSE)
+#' Tlg <- TravelLog (Write = FALSE)
 #'
-#' print(Tlg)
+#' print (Tlg)
 #'
 #' @keywords Travel Log Template Timezone Shift
 #' @export
