+++ R/write.act.R
@@ -20,48 +20,44 @@
 #' @title Export Daily Activity Score and Recording Summary
 #'
 #' @description
-#' A function that exports two types of files locally. 1. A summary
-#' of the entire recording (.csv). 2. Daily activity recordings in
-#' individual files (.txt). The default deliminator is set to be
-#' space, with `.` as the decimal indicator. As such, it should be
-#' noted that some transformation may be needed when opening in EXCEL
-#' on a computer with non-English EU settings.
+#' A function that exports two types of files locally. 1. A summary of the
+#' entire recording (.csv). 2. Daily activity recordings in individual files
+#' (.txt). The default deliminator is set to be space, with `.` as the decimal
+#' indicator. As such, it should be noted that some transformation may be
+#' needed when opening in EXCEL on a computer with non-English EU settings.
 #'
 #' @import utils
 #' @importFrom readr write_delim
 #'
-#' @param Dir The directory where the recordings to be exported
-#'   <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../
-#'   FOLDER NAME/">
+#' @param Dir The directory where the recordings to be exported <e.g.
+#' "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">
 #' @param ID The subject's ID which would be used to create a folder.
-#' @param df A data.frame of raw actigraphy recording. Both time and
-#'   activity count should be included in the `df`. See `VAct` and
-#'   `VTm` for further detail.
-#' @param Bdf A \code{\link{BriefSum}} object. Note, if jet lag
-#'   occurred during the recording, please, update the metadata using
-#'   \code{\link{TAdjust}} before passing to this function.
+#' @param df A data.frame of raw actigraphy recording. Both time and activity
+#' count should be included in the `df`. See `VAct` and `VTm` for further
+#' detail.
+#' @param Bdf A \code{\link{BriefSum}} object. Note, if jet lag occurred during
+#' the recording, please, update the metadata using \code{\link{TAdjust}}
+#' before passing to this function.
 #' @param TUnit Character; time--unit for the x--axis of each day's timeline.
 #'   Must be one of `day`, `hour`, `minute` or `second`.  Default is `hour`.
-#' @param VAct Optional character.  Name of the activity column in
-#'   `df`. If NULL, defaults to the second column of `df`.
-#' @param VTm Optional character.  Name of the time index column in
-#'   `df`. If NULL, defaults to the first column of `df`.
-#' @param Incomplete Logical; if TRUE, days flagged `Incomplete
-#'   Recording` (i.e. <24 h) are retained in the data list with
-#'   recordings segmented by day. Default = FALSE (these days are
-#'   removed).
-#' @param Travel Logical; if TRUE, days flagged `Travel` are retained
-#'   although some data points from an earlier adjacent calendar day
-#'   may be duplicated (a warning is issued). If FALSE, travel days
-#'   and the day before/after are excluded. Default = TRUE.
-#' @param Simple Logical; if TRUE, only columns stored in the
-#'   original recordings will be exported. If FALSE, all information
-#'   stored in `df` will be generated. Default = FALSE
-#'
+#' @param VAct Optional character.  Name of the activity column in `df`.
+#' If NULL, defaults to the second column of `df`.
+#' @param VTm Optional character.  Name of the time index column in `df`.
[…]
