+++ R/boot.seci.R
@@ -15,8 +15,7 @@
 #  You should have received a copy of the GNU Affero General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
-#' @title Bootstrap Standard Errors and Confidence Intervals for
-#'   Model Parameters
+#' @title Bootstrap Standard Errors and Confidence Intervals for Parameters
 #'
 #' @description Computes standard errors and confidence intervals for cosinor
 #' and post-hoc parameters via non-parametric bootstrap
@@ -52,161 +51,163 @@
 #' FlyEast
 #'
 #' BdfList <-
-#'     BriefSum (
-#'         df = FlyEast,
-#'         SR = 1 / 60,
-#'         Start = "2017-10-24 13:45:00"
-#'     )
+#'   BriefSum(
+#'     df = FlyEast,
+#'     SR = 1 / 60,
+#'     Start = "2017-10-24 13:45:00"
+#'   )
 #'
 #' # Let's extract actigraphy data from a single day
 #' df <- BdfList$df
-#' df <- subset (df, df$Date == "2017-10-27")
+#' df <- subset(df, df$Date == "2017-10-27")
 #'
 #' # Multicomponent Cosinor Model
-#' fit <- CosinorM (
-#'     time = df$Time,
-#'     activity = df$Activity,
-#'     tau = c (12, 24),
-#'     method = "OLS"
+#' fit <- CosinorM(
+#'   time = df$Time,
+#'   activity = df$Activity,
+#'   tau = c(12, 24),
+#'   method = "OLS"
 #' )
 #'
 #' # inspect coefficients
 #'
-#' boot.seci (
-#'     object = fit,
-#'     ci_level = 0.95,
-#'     n = 500
+#' boot.seci(
+#'   object = fit,
+#'   ci_level = 0.95,
+#'   n = 500
 #' )
 #'
 #'
 #' # Gaussian Kernel Density Estimation
-#' fit2 <- CosinorM.KDE (
-#'     time = df$Time,
-#'     activity = df$Activity
+#' fit2 <- CosinorM.KDE(
+#'   time = df$Time,
+#'   activity = df$Activity
 #' )
 #'
 #' # inspect coefficients
 #'
-#' boot.seci (
-#'     object = fit2,
-#'     ci_level = 0.95,
-#'     n = 500
+#' boot.seci(
+#'   object = fit2,
+#'   ci_level = 0.95,
+#'   n = 500
 #' )
 #' }
 #'
 #' @keywords boot bootstrap ci se
 #' @export
 
-boot.seci <- function (object, ci_level = 0.95, n = 500, digits = 2) {
-    ## Checkpoint -----------------------
-    if ("param" %in% names (object)) stop ("Object cannot be a diluted model structure.")
-    if (!inherits (object, c ("CosinorM", "CosinorM.KDE"))) {
-        stop ("object must be of class 'CosinorM' or 'CosinorM.KDE'")
+boot.seci <- function(object, ci_level = 0.95, n = 500, digits = 2) {
+  ## Checkpoint -----------------------
+  if ("param" %in% names(object))
+    stop("Object cannot be a diluted model structure.")
+
+  if (!inherits(object, c("CosinorM", "CosinorM.KDE"))) {
+    stop("object must be of class 'CosinorM' or 'CosinorM.KDE'")
+  }
+  if (!is.numeric(ci_level) || ci_level <= 0 || ci_level >= 1) {
+    stop("ci_level must be a numeric value between 0 and 1")
+  }
+  if (!is.numeric(n) || n <= 0 || n != round(n)) {
+    stop("n must be a positive integer")
+  }
[…]
