### 2025-12-15 851a59b
Add Test-Jeannick branch to pkgcheck workflow
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
