### 2025-12-15 a7815a3
Update R-CMD-check workflow matrix configurations
Removed deprecated R versions from CI matrix.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
