### 2025-12-15 004a4f4
Add comment for return output step
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
