+++ .github/workflows/R-CMD-check.yaml
@@ -11,7 +11,7 @@ concurrency:
 # usethis::use_github_action("check-standard") will install it.
 on:
   push:
-    branches: [Test]
+    branches: [Test-Copilot]
   pull_request:
 
 name: R-CMD-check.yaml
