### 2025-12-15 19ea28d
Update pkgdown.yaml
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
