### 2025-12-15 e86457d
Replace sapply with vapply for type safety
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
