+++ R/UTCwDST.R
@@ -98,11 +98,13 @@ UTCwDST <- function(UTCs, fork = FALSE) {
   ### Yes to daylight saving time for the north hemispher but NO for the south
 
   ## Check if any time zone may experience time change.
-  Out <- sapply(pTZs, function(tzs) {
-    wDST <- sapply(tzs, function(tz) as.POSIXlt(wDT, tz = tz)$isdst)
-    sDST <- sapply(tzs, function(tz) as.POSIXlt(sDT, tz = tz)$isdst)
+  Out <- vapply(pTZs, function(tzs) {
+    wDST <- vapply(tzs, function(tz) as.POSIXlt(wDT, tz = tz)$isdst,
+                   FUN.VALUE = integer(1))
+    sDST <- vapply(tzs, function(tz) as.POSIXlt(sDT, tz = tz)$isdst,
+                   FUN.VALUE = integer(1))
     any(wDST != sDST)
-  })
+  }, FUN.VALUE = logical(1))
 
   if (length(Out) == length(UTCs)) {
     names(Out) <- UTCs
