+++ R/GuessTZ.R
@@ -128,21 +128,21 @@ GuessTZ <- function(aOF, DT = NULL, iTZ = NULL, All = TRUE, fork = FALSE) {
 
 
   ## Step 1 Guess all possible TZ indicators -----------
-  pTZs <- sapply(aOF, function(x) oTZs[Toffs %in% x])
+  pTZs <- lapply(aOF, function(x) oTZs[Toffs %in% x])
 
   ## Step 2 Check if the initial time zone is included
   if (!is.null(TZ1)) {
     if (length(aOF) == 1) {
-      pTZs <- ifelse(TZ1 %in% pTZs, TZ1, pTZs)
+      pTZs <- ifelse(TZ1 %in% pTZs[[1]], list(TZ1), pTZs)
     } else {
-      pTZs <- sapply(pTZs, function(x) ifelse(TZ1 %in% x, TZ1, x))
+      pTZs <- lapply(pTZs, function(x) ifelse(TZ1 %in% x, TZ1, x))
     }
   }
 
   ## Step 3 Keep only the first one if the All is set to FALSE
   if (!All) {
     if (length(aOF) > 1) {
-      pTZs <- sapply(pTZs, function(x) x[[1]])
+      pTZs <- vapply(pTZs, function(x) x[[1]], FUN.VALUE = character(1))
     } else {
       pTZs <- pTZs[[1]]
     }
