### 2025-12-15 6506cfd
Fix long lines in vignettes and global variable bindings
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
