+++ vignettes/Circadian-Analysis.Rmd
@@ -22,7 +22,11 @@ library(ActiGlobe)
 
 
 ## Cosinor Analysis
-After we segmented the data by date, we can now model the circadian rhythm using the activity scores. For this, we will use the pre-travel recording on the 27<sup>th</sup> of October. Since the recording was not affected by the time change. We can simply store the unaffected segments in a new variable called `df`, which stands for data frame.
+After we segmented the data by date, we can now model the circadian 
+rhythm using the activity scores. For this, we will use the pre-travel 
+recording on the 27<sup>th</sup> of October. Since the recording was 
+not affected by the time change. We can simply store the unaffected 
+segments in a new variable called `df`, which stands for data frame.
 ```{r Import 20171027, message=FALSE, warning=FALSE, results='asis'}
 # Import data
 data("FlyEast")
@@ -52,7 +56,8 @@ df <- subset(df, df$Date == "2017-10-27")
 ### Single-phase Model (i.e., single phase circadian rhythm)
 #### Traditional Cosinor Model fitted via Ordinary Least Sqaure (OLS)
 
-```{r Model 1 OLS, message=FALSE, warning=FALSE, fig.height=5, fig.width=7, fig.alt=  "Figure 1: Fitting of OLS-based Cosinor Model"}
+```{r Model 1 OLS, message=FALSE, warning=FALSE, fig.height=5, 
+fig.width=7, fig.alt="Figure 1: Fitting of OLS-based Cosinor Model"}
 fit.ols <-
   CosinorM(
     time = df$Time,
@@ -72,13 +77,25 @@ ggCosinorM(fit.ols)
 
 
 #### Note 
-If we process `FlyEast` using the `cosinor::cosinor.lm`, we will notice that the results look a bit different. This is because ActiGlobe treats the first time point of the day as `00:00:00`, whereas this is sometimes treated as `00:00:00` plus one unit of the device sampling rate due to the use of the sequence generator. 
+If we process `FlyEast` using the `cosinor::cosinor.lm`, we will 
+notice that the results look a bit different. This is because 
+ActiGlobe treats the first time point of the day as `00:00:00`, 
+whereas this is sometimes treated as `00:00:00` plus one unit of the 
+device sampling rate due to the use of the sequence generator. 
 
 
 
 #### Ecnometry-modified Cosinor Model with Feasible General Least Sqaure (FGLS)
-Besides, the traditional OLS-based cosinor model,`CosinorM` also allows users to estimate the model using the FGLS-based method. In ActiGlobe, the FGLS was implemented via a weighted residual least square approximation. This two-step process often produces a better representation of the "border" between sleep and wakefulness than traditional OLS. Nevertheless, neither method can realistically capture our activity patterns, given that they are simply single-phase cosinor models.
-```{r Model 1 FGLS, message=FALSE, warning=FALSE, fig.height=5, fig.width=7, fig.alt=  "Figure 2: Fitting of FGLS-based Cosinor Model"}
+Besides, the traditional OLS-based cosinor model,`CosinorM` also 
+allows users to estimate the model using the FGLS-based method. In 
+ActiGlobe, the FGLS was implemented via a weighted residual least 
+square approximation. This two-step process often produces a better 
+representation of the "border" between sleep and wakefulness than 
+traditional OLS. Nevertheless, neither method can realistically 
+capture our activity patterns, given that they are simply 
+single-phase cosinor models.
+```{r Model 1 FGLS, message=FALSE, warning=FALSE, fig.height=5, 
[…]
