### 2025-12-15 7147ce3
Refactor ggCosinorM internal functions to match coding style
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
