+++ R/ggCosinorM.R
@@ -98,72 +98,74 @@
 ggCosinorM <- function (object, labels = TRUE, ci = TRUE, ci_level = 0.95,
                         n = 400, point_size = 0.5, title_extra = NULL, 
                         legend.position = "right", ...) {
-    # Validate inputs
-    .validate_ggcosinorm_inputs(object, ci_level, n)
+    # Step 0 Validate inputs ------------------------
+    ValidateInputs (object, ci_level, n)
 
-    # Extract basic parameters
-    day <- 24
-    tau <- if (!is.null (object$tau)) object$tau else 24
-    nT <- length (tau)
+    # Step 1 Extract Essential Parameters from Model Object ---------------
+    Day <- 24
+    Tau <- if (!is.null (object$tau)) object$tau else 24
+    nTau <- length (Tau)
 
-    # Extract observed data
-    aug <- .extract_observed_data(object)
+    ## Extract observed data
+    ObsData <- ExtractObservedData (object)
 
-    # Determine parameter source
-    coef_cos <- object$coef.cosinor
-    post <- object$post.hoc
-    use_parametric_single <- inherits (object, "CosinorM") && nT == 1
-    use_posthoc <- !use_parametric_single
+    ## Determine parameter source
+    CoefCos <- object$coef.cosinor
+    PostHoc <- object$post.hoc
+    UseParametricSingle <- inherits (object, "CosinorM") && nTau == 1
+    UsePostHoc <- !UseParametricSingle
 
-    # Extract cosinor parameters
-    mesor <- .get_mesor(object, use_posthoc, coef_cos, post)
-    amplitude <- .get_amplitude(object, use_posthoc, coef_cos, post, tau)
-    acrophase_time <- .get_acrophase_time(object, use_posthoc, coef_cos, 
-                                          post, tau)
-    bathy <- .get_bathyphase(object, use_posthoc, post, acrophase_time, 
-                             tau, day)
-    pv <- .get_peak_trough_values(use_posthoc, post, mesor, amplitude)
-    peak_value <- pv$peak
-    trough_value <- pv$trough
+    # Step 2 Extract Cosinor Parameters ------------------------
+    Mesor <- GetMesor (object, UsePostHoc, CoefCos, PostHoc)
+    Amplitude <- GetAmplitude (object, UsePostHoc, CoefCos, PostHoc, Tau)
+    AcrophaseTime <- GetAcrophaseTime (object, UsePostHoc, CoefCos, 
+                                       PostHoc, Tau)
+    Bathy <- GetBathyphase (object, UsePostHoc, PostHoc, AcrophaseTime, 
+                            Tau, Day)
+    PeakTroughVals <- GetPeakTroughValues (UsePostHoc, PostHoc, Mesor, 
+                                            Amplitude)
+    PeakValue <- PeakTroughVals$peak
+    TroughValue <- PeakTroughVals$trough
 
-    # Compute fitted curve
-    fit_result <- .compute_fitted_curve(object, aug, tau, n)
-    newt <- fit_result$newt
-    fit_pred <- fit_result$fit_pred
-    se_fit <- fit_result$se_fit
-    uses_kdf <- fit_result$uses_kdf
+    # Step 3 Compute Fitted Curve ------------------------
+    FitResult <- ComputeFittedCurve (object, ObsData, Tau, n)
+    NewT <- FitResult$newt
+    FitPred <- FitResult$fit_pred
+    SeFit <- FitResult$se_fit
+    UsesKdf <- FitResult$uses_kdf
 
-    # Compute confidence bands
[…]
