### 2025-12-15 d84a09f
Fix code review issues: spelling and type consistency
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
