+++ R/GuessTZ.R
@@ -133,9 +133,9 @@ GuessTZ <- function(aOF, DT = NULL, iTZ = NULL, All = TRUE, fork = FALSE) {
   ## Step 2 Check if the initial time zone is included
   if (!is.null(TZ1)) {
     if (length(aOF) == 1) {
-      pTZs <- ifelse(TZ1 %in% pTZs[[1]], list(TZ1), pTZs)
+      if (TZ1 %in% pTZs[[1]]) pTZs <- list(TZ1)
     } else {
-      pTZs <- lapply(pTZs, function(x) ifelse(TZ1 %in% x, TZ1, x))
+      pTZs <- lapply(pTZs, function(x) if (TZ1 %in% x) TZ1 else x)
     }
   }
 
