### 2025-12-15 b76d4bd
Shorten internal function names and fix vignette code chunks
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
