+++ R/ggCosinorM.R
@@ -102,7 +102,7 @@ ggCosinorM <- function (object, labels = TRUE, ci = TRUE, ci_level = 0.95,
                         n = 400, point_size = 0.5, title_extra = NULL, 
                         legend.position = "right", ...) {
     # Step 0 Validate inputs ------------------------
-    ValidateInputs (object, ci_level, n)
+    CheckInput (object, ci_level, n)
 
     # Step 1 Extract Essential Parameters from Model Object ---------------
     Day <- 24
@@ -110,7 +110,7 @@ ggCosinorM <- function (object, labels = TRUE, ci = TRUE, ci_level = 0.95,
     nTau <- length (Tau)
 
     ## Extract observed data
-    ObsData <- ExtractObservedData (object)
+    ObsData <- GetObsData (object)
 
     ## Determine parameter source
     CoefCos <- object$coef.cosinor
@@ -119,33 +119,31 @@ ggCosinorM <- function (object, labels = TRUE, ci = TRUE, ci_level = 0.95,
     UsePostHoc <- !UseParametricSingle
 
     # Step 2 Extract Cosinor Parameters ------------------------
-    Mesor <- GetMesor (object, UsePostHoc, CoefCos, PostHoc)
-    Amplitude <- GetAmplitude (object, UsePostHoc, CoefCos, PostHoc, Tau)
-    AcrophaseTime <- GetAcrophaseTime (object, UsePostHoc, CoefCos, 
-                                       PostHoc, Tau)
-    Bathy <- GetBathyphase (object, UsePostHoc, PostHoc, AcrophaseTime, 
-                            Tau, Day)
-    PeakTroughVals <- GetPeakTroughValues (UsePostHoc, PostHoc, Mesor, 
-                                            Amplitude)
+    MESOR <- GetParam ("MESOR", UsePostHoc, CoefCos, PostHoc, Tau)
+    Amplitude <- GetParam ("Amplitude", UsePostHoc, CoefCos, PostHoc, Tau)
+    AcrophaseTime <- GetParam ("Acrophase", UsePostHoc, CoefCos, PostHoc, Tau)
+    Bathy <- GetParam ("Bathyphase", UsePostHoc, PostHoc, NULL, Tau, Day, 
+                       AcrophaseTime)
+    PeakTroughVals <- GetParam ("PeakTrough", UsePostHoc, PostHoc, NULL, NULL, 
+                                NULL, NULL, MESOR, Amplitude)
     PeakValue <- PeakTroughVals$peak
     TroughValue <- PeakTroughVals$trough
 
     # Step 3 Compute Fitted Curve ------------------------
-    FitResult <- ComputeFittedCurve (object, ObsData, Tau, n)
+    FitResult <- GetFit (object, ObsData, Tau, n)
     NewT <- FitResult$newt
     FitPred <- FitResult$fit_pred
     SeFit <- FitResult$se_fit
     UsesKdf <- FitResult$uses_kdf
 
     # Step 4 Compute Confidence Bands ------------------------
-    CiResult <- ComputeConfidenceBands (ci, UsesKdf, FitPred, SeFit, 
-                                        ci_level, object)
+    CiResult <- GetCI (ci, UsesKdf, FitPred, SeFit, ci_level, object)
     YMin <- CiResult$ym
     YMax <- CiResult$yM
 
     # Step 5 Fallback Markers from Fitted Curve ------------------------
-    FallbackMarkers <- FallbackMarkersFromFit (Bathy, PeakValue, TroughValue, 
-                                                AcrophaseTime, NewT, FitPred)
+    FallbackMarkers <- FallbackFit (Bathy, PeakValue, TroughValue, 
[…]
