### 2025-12-15 cfee6c1
Add default case to GetParam switch for error handling
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
