### 2025-12-15 948920d
Use ifelse for scalar value assignments where appropriate
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
