### 2025-12-15 720c0a3
Fix helper function documentation style per contributing.md
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
