+++ R/ggCosinorM.R
@@ -198,18 +198,6 @@ GetObsData <- function (object) {
 }
 
 #' @title Get Cosinor Parameter
-#' @description
-#' Unified function to extract cosinor parameters with switch statement.
-#' @param param Parameter name: "MESOR", "Amplitude", "Acrophase", 
-#'   "Bathyphase", or "PeakTrough"
-#' @param use_posthoc Logical; if TRUE, use post-hoc estimates
-#' @param coef_cos Cosinor coefficients vector (for MESOR, Amplitude, Acrophase)
-#' @param post Post-hoc estimates vector
-#' @param tau Period(s) of the cosinor model
-#' @param day Length of day (for Bathyphase)
-#' @param acrophase_time Acrophase time (for Bathyphase)
-#' @param mesor MESOR value (for PeakTrough)
-#' @param amplitude Amplitude value (for PeakTrough)
 #' @noRd
 GetParam <- function (param, use_posthoc, coef_cos = NULL, post = NULL, 
                       tau = NULL, day = NULL, acrophase_time = NULL, 
@@ -404,16 +392,6 @@ FallbackFit <- function (bathy, peak_value, trough_value,
 }
 
 #' @title Plot Fit
-#' @description
-#' Adds the fitted cosinor curve and optional confidence ribbon to
-#' a ggplot object.
-#' @param g A ggplot object.
-#' @param newt Time points for fitted curve.
-#' @param fit_pred Fitted values.
-#' @param ci Logical; if TRUE, add confidence ribbon.
-#' @param ym Lower confidence bound.
-#' @param yM Upper confidence bound.
-#' @returns Updated ggplot object.
 #' @noRd
 PlotFit <- function(g, newt, fit_pred, ci, ym, yM) {
     g <- g + ggplot2::geom_line (
@@ -434,18 +412,6 @@ PlotFit <- function(g, newt, fit_pred, ci, ym, yM) {
 }
 
 #' @title Plot MESOR
-#' @description
-#' Adds MESOR horizontal line, observed data points, peak and trough
-#' markers to a ggplot object.
-#' @param g A ggplot object.
-#' @param mesor MESOR value.
-#' @param aug Data frame with observed data.
-#' @param point_size Size for observed data points.
-#' @param acrophase_time Acrophase time value.
-#' @param peak_value Peak activity value.
-#' @param bathy Bathyphase time value.
-#' @param trough_value Trough activity value.
-#' @returns Updated ggplot object.
 #' @noRd
 PlotMESOR <- function(g, mesor, aug, point_size, acrophase_time, 
                                   peak_value, bathy, trough_value) {
@@ -478,12 +444,6 @@ PlotMESOR <- function(g, mesor, aug, point_size, acrophase_time,
 }
 
 #' @title Plot Acrophase
-#' @description
-#' Adds vertical line(s) at the acrophase time(s) to a ggplot object.
-#' @param g A ggplot object.
-#' @param acrophase_time Acrophase time value(s).
-#' @param aug Data frame with observed data.
-#' @returns Updated ggplot object.
 #' @noRd
 PlotAcro <- function(g, acrophase_time, aug) {
     y_sf <- floor (max (aug$y_obs, na.rm = TRUE) * 0.8)
@@ -498,12 +458,6 @@ PlotAcro <- function(g, acrophase_time, aug) {
 }
 
 #' @title Plot Inactive
-#' @description
-#' Identifies and adds shaded rectangles for detected inactive periods
-#' to a ggplot object.
-#' @param g A ggplot object.
-#' @param aug Data frame with observed data.
-#' @returns Updated ggplot object.
 #' @noRd
 PlotInactive <- function(g, aug) {
[…]
