### 2025-12-15 2bd6e23
Restore full documentation for helper functions
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
