+++ R/ggCosinorM.R
@@ -167,6 +167,12 @@ ggCosinorM <- function (object, labels = TRUE, ci = TRUE, ci_level = 0.95,
 
 # Pre-defined helper functions -------------------------------
 #' @title Check Input
+#' @description
+#' Validates that the object is of correct class and that ci_level and
+#' n parameters are within acceptable ranges.
+#' @param object A fitted model object to validate.
+#' @param ci_level Confidence level to validate.
+#' @param n Number of grid points to validate.
 #' @noRd
 CheckInput <- function (object, ci_level, n) {
     if (!inherits (object, c ("CosinorM", "CosinorM.KDE")))
@@ -181,6 +187,11 @@ CheckInput <- function (object, ci_level, n) {
 }
 
 #' @title Get Observed Data
+#' @description
+#' Extracts time and activity observations from a fitted CosinorM or
+#' CosinorM.KDE model object.
+#' @param object A fitted model object.
+#' @returns A data.frame with columns t_obs and y_obs.
 #' @noRd
 GetObsData <- function (object) {
     if (!is.null (object$model) && !is.null (object$model$time)) {
@@ -198,6 +209,19 @@ GetObsData <- function (object) {
 }
 
 #' @title Get Cosinor Parameter
+#' @description
+#' Unified function to extract cosinor parameters with switch statement.
+#' @param param Parameter name: "MESOR", "Amplitude", "Acrophase", 
+#'   "Bathyphase", or "PeakTrough"
+#' @param use_posthoc Logical; if TRUE, use post-hoc estimates
+#' @param coef_cos Cosinor coefficients vector (for MESOR, Amplitude, Acrophase)
+#' @param post Post-hoc estimates vector
+#' @param tau Period(s) of the cosinor model
+#' @param day Length of day (for Bathyphase)
+#' @param acrophase_time Acrophase time (for Bathyphase)
+#' @param mesor MESOR value (for PeakTrough)
+#' @param amplitude Amplitude value (for PeakTrough)
+#' @returns Numeric value or list depending on param type.
 #' @noRd
 GetParam <- function (param, use_posthoc, coef_cos = NULL, post = NULL, 
                       tau = NULL, day = NULL, acrophase_time = NULL, 
@@ -269,6 +293,14 @@ GetParam <- function (param, use_posthoc, coef_cos = NULL, post = NULL,
 }
 
 #' @title Get Parametric Fit
+#' @description
+#' Computes the fitted values and standard errors for a parametric
+#' cosinor model over a fine time grid.
+#' @param object A fitted CosinorM object.
+#' @param aug Data frame with observed data.
+#' @param tau Period(s) of the cosinor model.
+#' @param n Number of grid points for prediction.
+#' @returns A list with newt, fit_pred, se_fit, and uses_kdf.
 #' @noRd
 GetParamFit <- function (object, aug, tau, n) {
     nTau <- length (tau)
@@ -320,6 +352,13 @@ GetParamFit <- function (object, aug, tau, n) {
 }
 
 #' @title Get KDE Fit
+#' @description
+#' Extracts fitted values and standard errors from a CosinorM.KDE
+#' model object.
+#' @param object A fitted CosinorM.KDE object.
+#' @param t_obs Observed time points.
+#' @param tau Period(s) of the cosinor model.
+#' @returns A list with newt, fit_pred, se_fit, and uses_kdf, or NULL.
 #' @noRd
 GetKDEFit <- function (object, t_obs, tau) {
     if (!is.null (object$kdf)) {
@@ -341,6 +380,14 @@ GetKDEFit <- function (object, t_obs, tau) {
 }
 
 #' @title Get Fit
+#' @description
[…]
