+++ R/TimeFormat.R
@@ -30,7 +30,7 @@
 #' @returns
 #' If \code{as.time = TRUE}, returns a character vector of the input times
 #' reformatted according to the detected format. Each element corresponds to
-#' the respective entry in \code{Time}.
+#' the respective entry in \code{time}.
 #'
 #' If \code{as.time = FALSE}, returns a character scalar giving the
 #' best-matching time format string.
@@ -47,7 +47,7 @@
 #' )
 #'
 #' ## Ask TimeFormat to tell us the format of the time.
-#' TimeFormat (Time, as.time = FALSE)
+#' TimeFormat (time - Time, as.time = FALSE)
 #'
 #' # Example 2: When multiple formats co-exist in a variable
 #' ## Create and store dates and time in a variable called Time
@@ -68,7 +68,7 @@
 #'
 #'
 #' ### DO NOT!
-#' ### print(TimeFormat(Time))
+#' ### print(TimeFormat(time = Time))
 #' ### Note, this process will fail because there are multiple formats
 #'
 #' @export
