+++ R/ggCosinorM.R
@@ -26,8 +26,12 @@
 #' MESOR line, acrophase verticals, amplitude annotation segments, and
 #' labelled parameter values when requested.
 #'
-#' @import stats ggplot2 ggrepel viridis
+#' @importFrom stats coef df.residual model.frame model.response qt vcov
+#' @importFrom ggplot2 aes coord_cartesian element_blank element_text geom_hline geom_line geom_point geom_rect geom_ribbon geom_segment ggplot labs margin scale_colour_manual scale_fill_manual scale_x_continuous theme theme_minimal
+#' @importFrom ggrepel geom_label_repel
+#' @importFrom viridis scale_color_viridis scale_fill_viridis
 #' @importFrom scales alpha
+#' @importFrom utils globalVariables
 #'
 #' @param object A fitted model of class \code{\link{CosinorM}} or
 #'   \code{\link{CosinorM.KDE}}.
@@ -97,87 +101,72 @@
 ggCosinorM <- function (object, labels = TRUE, ci = TRUE, ci_level = 0.95,
                         n = 400, point_size = 0.5, title_extra = NULL,
                         legend.position = "right", ...) {
-    # Validate inputs
-    .validate_ggcosinorm_inputs (object, ci_level, n)
+    # Step 0 Validate inputs ------------------------
+    CheckInput (object, ci_level, n)
 
-    # Extract basic parameters
-    day <- 24
-    tau <- if (!is.null (object$tau)) object$tau else 24
-    nT <- length (tau)
+    # Step 1 Extract Essential Parameters from Model Object ---------------
+    Day <- 24
+    Tau <- if (!is.null (object$tau)) object$tau else 24
+    nTau <- length (Tau)
 
-    # Extract observed data
-    aug <- .extract_observed_data (object)
+    ## Extract observed data
+    ObsData <- GetObsData (object)
 
-    # Determine parameter source
-    coef_cos <- object$coef.cosinor
-    post <- object$post.hoc
-    use_parametric_single <- inherits (object, "CosinorM") && nT == 1
-    use_posthoc <- !use_parametric_single
+    ## Determine parameter source
+    CoefCos <- object$coef.cosinor
+    PostHoc <- object$post.hoc
+    UseParametricSingle <- inherits (object, "CosinorM") && nTau == 1
+    UsePostHoc <- !UseParametricSingle
 
-    # Extract cosinor parameters
-    mesor <- .get_mesor (object, use_posthoc, coef_cos, post)
-    amplitude <- .get_amplitude (object, use_posthoc, coef_cos, post, tau)
-    acrophase_time <- .get_acrophase_time (
-        object, use_posthoc, coef_cos,
-        post, tau
-    )
-    bathy <- .get_bathyphase (
-        object, use_posthoc, post, acrophase_time,
-        tau, day
-    )
-    pv <- .get_peak_trough_values (use_posthoc, post, mesor, amplitude)
-    peak_value <- pv$peak
-    trough_value <- pv$trough
+    # Step 2 Extract Cosinor Parameters ------------------------
+    MESOR <- GetParam ("MESOR", UsePostHoc, CoefCos, PostHoc, Tau)
+    Amplitude <- GetParam ("Amplitude", UsePostHoc, CoefCos, PostHoc, Tau)
+    AcrophaseTime <- GetParam ("Acrophase", UsePostHoc, CoefCos, PostHoc, Tau)
+    Bathy <- GetParam ("Bathyphase", UsePostHoc, PostHoc, NULL, Tau, Day,
+                       AcrophaseTime)
+    PeakTroughVals <- GetParam ("PeakTrough", UsePostHoc, PostHoc, NULL, NULL,
[…]
