+++ R/CosinorM.R
@@ -79,7 +79,8 @@
 #' \deqn{A= \sqrt {(\beta^2 + \gamma^2)}}
 #'
 #'
-#' @import stats sandwich
+#' @importFrom stats lm coef resid predict update as.formula vcov
+#' @importFrom sandwich vcovHC
 #'
 #'
 #' @param time Numeric vector of time coordinates for each data point
@@ -159,26 +160,26 @@
 #'
 #' \dontrun{
 #' # Import data
-#' data (FlyEast)
+#' data( FlyEast)
 #'
 #'
 #' # Create quick summary of the recording with adjustment for daylight saving.
 #' BdfList <-
-#'     BriefSum (
-#'         df = FlyEast,
-#'         SR = 1 / 60,
-#'         Start = "2017-10-24 13:45:00"
-#'     )
+#'   BriefSum (
+#'     df = FlyEast,
+#'     SR = 1 / 60,
+#'     Start = "2017-10-24 13:45:00"
+#'   )
 #'
 #' # Let's extract actigraphy data from a single day
 #' df <- BdfList$df
 #' df <- subset (df, df$Date == "2017-10-27")
 #'
 #' fit <- CosinorM (
-#'     time = df$Time,
-#'     activity = df$Activity,
-#'     tau = 24,
-#'     method = "OLS"
+#'   time = df$Time,
+#'   activity = df$Activity,
+#'   tau = 24,
+#'   method = "OLS"
 #' )
 #'
 #'
@@ -187,55 +188,50 @@
 #'
 #'
 #' # plot Cosinor in hours
-#' plot (fit$time,
-#'     fit$fitted.values,
-#'     type = "l",
-#'     xlab = "Hour",
-#'     ylab = "24-Hour Cosinor Model"
-#' )
+#' ggCosinorM (fit)
 #' }
 #'
 #' @keywords cosinor
 #' @export
 
 
-CosinorM <- function (time, activity, tau, method = "OLS", arctan2 = TRUE,
-                      type = "HC3", dilute = FALSE) {
+CosinorM <- function(time, activity, tau, method = "OLS", arctan2 = TRUE,
+                     type = "HC3", dilute = FALSE) {
     # Check Point and Input Validation -------------------------
-    activity <- ValInput (x = activity, type = "Act")
-    time <- ValInput (x = time, type = "Tm")
-    if (!method %in% c ("OLS", "FGLS")) stop ("Unsupported method specified!")
+    activity <- ValInput(x = activity, type = "Act")
+    time <- ValInput(x = time, type = "Tm")
+    if (!method %in% c("OLS", "FGLS")) stop("Unsupported method specified!")
 
 
     # Get Essential Info -----------------
-    nT <- length (tau) ### Number of assumed rhythms
-    Vars <- as.vector (outer (c ("C", "S"), 1:nT, paste0))
+    nT <- length(tau) ### Number of assumed rhythms
+    Vars <- as.vector(outer(c("C", "S"), 1:nT, paste0))
     ### C = x and S = z in the linear equation
 
     day <- 24 ### 24 hours per-day for now
     factor <- day / tau
 
-    dt <- diff (time)
-    dt <- dt [dt > 0]
-    Epc <- 1 / min (dt)
+    dt <- diff(time)
+    dt <- dt[dt > 0]
+    Epc <- 1 / min(dt)
 
 
     # Build cosine and sine columns for each period -------------------
     ## Create a model data.frame
-    Mdf <- data.frame (matrix (nrow = length (time), ncol = nT * 2))
-    names (Mdf) <- Vars
+    Mdf <- data.frame(matrix(nrow = length(time), ncol = nT * 2))
[…]
