+++ R/ggActiGlobe.R
@@ -23,7 +23,9 @@
 #' `coloring` points that were flagged (e.g. travel overlaps or unallocated
 #' epochs).
 #'
-#' @import ggplot2
+#' @importFrom ggplot2 ggplot aes element_line element_text geom_point
+#' @importFrom ggplot2 geom_vline labs margin scale_x_continuous
+#' @importFrom ggplot2 scale_y_continuous theme theme_classic
 #'
 #' @param df A data.frame of annotated actigraphy epochs.  Must include:
 #'   - An activity column named by `VAct`.
@@ -46,39 +48,39 @@
 #'
 #' @examples
 #' \dontrun{
-#' library (ActiGlobe)
+#' library(ActiGlobe)
 #'
 #' # Overview the Uncorrected Longitudinal Recording
-#' data (FlyEast)
+#' data(FlyEast)
 #'
 #' BdfList <-
-#'     BriefSum (
-#'         df = FlyEast,
-#'         SR = 1 / 60,
-#'         Start = "2017-10-24 13:45:00"
-#'     )
+#'   BriefSum(
+#'     df = FlyEast,
+#'     SR = 1 / 60,
+#'     Start = "2017-10-24 13:45:00"
+#'   )
 #'
-#' p <- ggActiGlobe (
-#'     df = BdfList$df,
-#'     Bdf = BdfList$Bdf,
-#'     VAct = "Activity",
-#'     VDT = "DateTime"
+#' p <- ggActiGlobe(
+#'   df = BdfList$df,
+#'   Bdf = BdfList$Bdf,
+#'   VAct = "Activity",
+#'   VDT = "DateTime"
 #' )
 #'
-#' print (p)
+#' print(p)
 #'
 #'
 #' # Overview the Corrected Longitudinal Recording
-#' data (TLog)
+#' data(TLog)
 #'
-#' BdfList$Bdf.adj <- TAdjust (BdfList$Bdf, TLog)
-#' p2 <- ggActiGlobe (
-#'     df = BdfList$df,
-#'     Bdf = BdfList$Bdf,
-#'     VAct = "Activity",
-#'     VDT = "DateTime"
+#' BdfList$Bdf.adj <- TAdjust(BdfList$Bdf, TLog)
+#' p2 <- ggActiGlobe(
+#'   df = BdfList$df,
+#'   Bdf = BdfList$Bdf,
+#'   VAct = "Activity",
+#'   VDT = "DateTime"
 #' )
-#' print (p2)
+#' print(p2)
 #'
 #' # Pro-tip: [`cowplot`] can help stack the time series graphs in one
 #' # single plot
@@ -88,139 +90,141 @@
 #' @export
 
 
-ggActiGlobe <- function (df, Bdf, VAct = NULL, VDT = "DateTime") {
+ggActiGlobe <- function(df, Bdf, VAct = NULL, VDT = "DateTime") {
     ## Ensure Note column exists --------------
-    if (!"Note" %in% names (df)) {
+    if (!"Note" %in% names(df)) {
         df$Note <- ""
     }
 
-    NR <- seq_len (nrow (df))
-    A <- df [[VAct]]
-    Mx <- round (max (A, na.rm = TRUE))
-    mn <- round (min (A, na.rm = TRUE))
+    NR <- seq_len(nrow(df))
+    A <- df[[VAct]]
+    Mx <- round(max(A, na.rm = TRUE))
+    mn <- round(min(A, na.rm = TRUE))
 
 
     VD <- "Date"
-    D <- df [[VD]]
-    DT <- df [[VDT]]
-    Tc <- sub ("^\\S+\\s+", "", as.character (DT))
+    D <- df[[VD]]
[…]
