+++ NAMESPACE
@@ -31,6 +31,7 @@ import(utils)
 importFrom(ggplot2,aes)
 importFrom(ggplot2,coord_cartesian)
 importFrom(ggplot2,element_blank)
+importFrom(ggplot2,element_line)
 importFrom(ggplot2,element_text)
 importFrom(ggplot2,geom_hline)
 importFrom(ggplot2,geom_line)
@@ -38,13 +39,16 @@ importFrom(ggplot2,geom_point)
 importFrom(ggplot2,geom_rect)
 importFrom(ggplot2,geom_ribbon)
 importFrom(ggplot2,geom_segment)
+importFrom(ggplot2,geom_vline)
 importFrom(ggplot2,ggplot)
 importFrom(ggplot2,labs)
 importFrom(ggplot2,margin)
 importFrom(ggplot2,scale_colour_manual)
 importFrom(ggplot2,scale_fill_manual)
 importFrom(ggplot2,scale_x_continuous)
+importFrom(ggplot2,scale_y_continuous)
 importFrom(ggplot2,theme)
+importFrom(ggplot2,theme_classic)
 importFrom(ggplot2,theme_minimal)
 importFrom(ggrepel,geom_label_repel)
 importFrom(lubridate,dst)
