+++ R/mIANA.R
@@ -18,8 +18,8 @@
 #' @title Mini IANA Table for Internal Uses
 #'
 #' @description
-#' A miniturized IANA table generator that only extract the first timezone based
-#' on the abbreviated time zone code. An internal function to resolve the
+#' A miniturized IANA table generator that only extract the first timezone
+#' based on the abbreviated time zone code. An internal function to resolve the
 #' internal data access issue.
 #'
 #' @import utils
@@ -47,93 +47,102 @@
 #' @noRd
 
 mIANA <- function (Write = FALSE, Dir = NULL) {
-    Timezone_IANA <-
-        c (
-            "Europe/Amsterdam", "America/Vancouver", "America/Chicago",
-            "Europe/London", "Africa/El_Aaiun", "Australia/Adelaide", "Europe/Athens",
-            "America/New_York", "America/Denver", "Australia/Sydney",
-            "America/Anchorage", "America/Halifax", "America/Santiago",
-            "Atlantic/Azores", "America/St_Johns", "Pacific/Auckland", "America/Adak",
-            "Asia/Jerusalem", "Europe/Dublin", "America/Godthab", "America/Miquelon",
-            "Antarctica/Troll", "Australia/Lord_Howe", "Pacific/Easter",
-            "Pacific/Chatham", "Pacific/Norfolk", "Asia/Aden",
-            "America/Argentina/Salta", "Africa/Lagos", "Asia/Tehran",
-            "Africa/Algiers", "America/Belem", "America/Asuncion", "Asia/Krasnoyarsk",
-            "America/Lima", "Pacific/Port_Moresby", "Asia/Karachi", "Africa/Abidjan",
-            "Asia/Dubai", "Asia/Damascus", "America/Porto_Velho",
-            "America/Mexico_City", "America/Caracas", "America/La_Paz",
-            "Africa/Khartoum", "Europe/Istanbul", "Africa/Addis_Ababa",
-            "Australia/Darwin", "Pacific/Guam", "Asia/Baku", "Asia/Chita",
-            "America/Hermosillo", "America/Montevideo", "Africa/Tripoli",
-            "Europe/Astrakhan", "Asia/Tokyo", "Asia/Aqtau", "Asia/Yekaterinburg",
-            "Europe/Moscow", "America/Guayaquil", "Australia/Perth", "Asia/Sakhalin",
-            "Africa/Johannesburg", "Asia/Rangoon", "America/Panama",
-            "Asia/Kuala_Lumpur", "Asia/Hovd", "Asia/Jayapura", "Asia/Vladivostok",
-            "Asia/Anadyr", "Asia/Pyongyang", "Asia/Tashkent", "Asia/Kabul",
-            "Asia/Bangkok", "Asia/Irkutsk", "Asia/Manila", "America/Guyana",
-            "Pacific/Apia", "America/Bogota", "Asia/Ulaanbaatar", "Asia/Ashgabat",
-            "Asia/Bishkek", "Australia/Brisbane", "Asia/Taipei", "Asia/Kathmandu",
-            "Asia/Makassar", "Asia/Jakarta", "Asia/Brunei", "Asia/Tbilisi",
-            "America/Paramaribo", "America/Whitehorse", "America/Cayenne",
[…]
