+++ R/Rad2Hr.R
@@ -29,29 +29,29 @@
 #' @return Numeric vector of times in [0, tau). NA values propagate.
 #'
 #' @examples
-#' Rad2Hr(pi / 2, tau = 24)
+#' Rad2Hr (pi / 2, tau = 24)
 #'
-#' Rad2Hr(c(-pi / 2, 0, pi, 3 * pi / 2), tau = 24)
+#' Rad2Hr (c (-pi / 2, 0, pi, 3 * pi / 2), tau = 24)
 #'
 #' @keywords acrophase radian
 #' @export
 
-Rad2Hr <- function(x, tau) {
-  # coerce x to numeric while preserving NA -------------
-  if (!is.numeric(x)) x <- as.numeric(as.character(x))
-  if (!is.numeric(tau)) tau <- as.numeric(as.character(tau))
+Rad2Hr <- function (x, tau) {
+    # coerce x to numeric while preserving NA -------------
+    if (!is.numeric (x)) x <- as.numeric (as.character (x))
+    if (!is.numeric (tau)) tau <- as.numeric (as.character (tau))
 
-  # validate tau --------------
-  if (length(tau) != 1 || is.na(tau) || !is.finite(tau)) {
-    stop("tau must be a single finite numeric value.")
-  }
-  if (!(tau > 0 && tau <= 24)) {
-    stop("tau must be greater than 0 and less than or equal to 24.")
-  }
+    # validate tau --------------
+    if (length (tau) != 1 || is.na (tau) || !is.finite (tau)) {
+        stop ("tau must be a single finite numeric value.")
+    }
+    if (!(tau > 0 && tau <= 24)) {
+        stop ("tau must be greater than 0 and less than or equal to 24.")
+    }
 
-  # conversion and wrap to [0, tau) ---------------
-  acrophase_time <- (x * tau) / (2 * pi)
-  out <- acrophase_time %% tau
+    # conversion and wrap to [0, tau) ---------------
+    acrophase_time <- (x * tau) / (2 * pi)
+    out <- acrophase_time %% tau
 
-  return(as.numeric(out))
+    return (as.numeric (out))
 }
