+++ R/se.CosinorM.R
@@ -108,115 +108,115 @@
 #' @examples
 #' \dontrun{
 #' # Import data
-#' data(FlyEast)
+#' data (FlyEast)
 #'
 #'
 #' # Create quick summary of the recording with adjustment for daylight saving.
 #' BdfList <-
-#'   BriefSum(
-#'     df = FlyEast,
-#'     SR = 1 / 60,
-#'     Start = "2017-10-24 13:45:00"
-#'   )
+#'     BriefSum (
+#'         df = FlyEast,
+#'         SR = 1 / 60,
+#'         Start = "2017-10-24 13:45:00"
+#'     )
 #'
 #' # Let's extract actigraphy data from a single day
 #' df <- BdfList$df
-#' df <- subset(df, df$Date == "2017-10-27")
+#' df <- subset (df, df$Date == "2017-10-27")
 #'
-#' fit <- CosinorM(
-#'   time = df$Time,
-#'   activity = df$Activity,
-#'   tau = 24,
-#'   method = "OLS"
+#' fit <- CosinorM (
+#'     time = df$Time,
+#'     activity = df$Activity,
+#'     tau = 24,
+#'     method = "OLS"
 #' )
 #'
 #' # Compute variance and Delta SEs
-#' res <- se.CosinorM(object = fit)
-#' print(res)
+#' res <- se.CosinorM (object = fit)
+#' print (res)
 #' }
 #'
 #' @noRd
 
 
-se.CosinorM <- function(object, method = "delta") {
-  # Check Point --------------------------
-  if (method != "delta") {
-    stop("Only method = 'delta' is currently supported.")
-  }
+se.CosinorM <- function (object, method = "delta") {
+    # Check Point --------------------------
+    if (method != "delta") {
+        stop ("Only method = 'delta' is currently supported.")
+    }
 
 
-  # Extract Essential Parameters ------------------
-  VCOV <- object$vcov
-  tau <- object$tau
-  lT <- length(tau)
-  Coefs <- object$coef.cosinor
+    # Extract Essential Parameters ------------------
+    VCOV <- object$vcov
+    tau <- object$tau
+    lT <- length (tau)
+    Coefs <- object$coef.cosinor
 
-  # Debug parameters ------------------------
-  ## CSidx <- grep("^[CS][0-9]+$", row.names(VCOV))
-  ## SE <- sqrt(diag(VCOV))
+    # Debug parameters ------------------------
+    ## CSidx <- grep("^[CS][0-9]+$", row.names(VCOV))
+    ## SE <- sqrt(diag(VCOV))
 
-  # Collect the betas and gammas
-  # Compute Variance for Acrophase and Amplitude  --------------
-  VARs <-
-    lapply(1:lT, function(i) {
-      Tu <- tau[i]
-      VarB <- VCOV[paste0("C", i), paste0("C", i)]
-      VarG <- VCOV[paste0("S", i), paste0("S", i)]
-      CovBG <- VCOV[paste0("C", i), paste0("S", i)]
+    # Collect the betas and gammas
+    # Compute Variance for Acrophase and Amplitude  --------------
+    VARs <-
+        lapply (1:lT, function (i) {
+            Tu <- tau [i]
+            VarB <- VCOV [paste0 ("C", i), paste0 ("C", i)]
+            VarG <- VCOV [paste0 ("S", i), paste0 ("S", i)]
+            CovBG <- VCOV [paste0 ("C", i), paste0 ("S", i)]
 
-      amp <- Coefs[grep(paste0("Amplitude.", Tu), names(Coefs))]
[…]

[…]
