+++ vignettes/Circadian-Analysis.Rmd
@@ -274,7 +274,8 @@ cosinor_peak_hour <- (phi %% (2 * pi)) * 24 / (2 * pi) %% 24
 true_peak_hour <- plot_df$hour [which.max (df$Activity)]
 
 # Kernel density (assumes kdf$hour in hours)
-kdf_plot <- data.frame (hour = fit.KDE$model$time, density = fit.KDE$kdf$density)
+kdf_plot <- 
+  data.frame (hour = fit.KDE$model$time, density = fit.KDE$kdf$density)
 y_primary_max <- 2000
 kdf_plot$density_scaled <-
     kdf_plot$density / max (kdf_plot$density) * y_primary_max
@@ -372,7 +373,8 @@ cosinor_peak_hour <- ((phi * 24 / (2 * pi)) %% 24)
 true_peak_hour <- plot_df$hour [which.max (df$Activity)]
 
 # Kernel density (assumes kdf$hour in hours)
-kdf_plot <- data.frame (hour = fit.KDE$model$time, density = fit.KDE$kdf$density)
+kdf_plot <- 
+  data.frame (hour = fit.KDE$model$time, density = fit.KDE$kdf$density)
 y_primary_max <- 2000
 kdf_plot$density_scaled <- kdf_plot$density / max (kdf_plot$density) *
     y_primary_max
