+++ man/CosinorM.KDE.Rd
@@ -251,7 +251,7 @@ Using R With Examples in Cognitive Psychology. Methods. Front Psychol.
 2018-October-30 2018;9(2040):2040. doi:10.3389/fpsyg.2018.02040
 }
 \seealso{
-\code{\link{CosinorM}}
+\code{\link{CosinorM}}  \code{\link{ggCosinorM}}
 }
 \keyword{KDE}
 \keyword{circadian}
