+++ man/CosinorM.Rd
@@ -132,17 +132,17 @@ Amplitude is calculated from fitted sine and cosine coefficients as:
 \deqn{A= \sqrt {(\beta^2 + \gamma^2)}}
 }
 \examples{
-require(stats)
-require(graphics)
+require (stats)
+require (graphics)
 
 \dontrun{
 # Import data
-data(FlyEast)
+data( FlyEast)
 
 
 # Create quick summary of the recording with adjustment for daylight saving.
 BdfList <-
-  BriefSum(
+  BriefSum (
     df = FlyEast,
     SR = 1 / 60,
     Start = "2017-10-24 13:45:00"
@@ -150,9 +150,9 @@ BdfList <-
 
 # Let's extract actigraphy data from a single day
 df <- BdfList$df
-df <- subset(df, df$Date == "2017-10-27")
+df <- subset (df, df$Date == "2017-10-27")
 
-fit <- CosinorM(
+fit <- CosinorM (
   time = df$Time,
   activity = df$Activity,
   tau = 24,
@@ -188,6 +188,6 @@ Harvey, A. C. (1976). Estimating Regression Models with Multiplicative
 Heteroscedasticity. Econometrica, 44(3), 461-465. doi:10.2307/1913974
 }
 \seealso{
-\code{\link[stats]{lm}}  \code{\link{CosinorM.KDE}}
+\code{\link[stats]{lm}}  \code{\link{CosinorM.KDE}}  \code{\link{ggCosinorM}}
 }
 \keyword{cosinor}
