### 2025-12-15 5744ede
Update pkgcheck.yaml
