+++ .github/workflows/pkgcheck.yaml
@@ -15,8 +15,10 @@ on:
 
 jobs:
   pkgcheck:
-    runs-on: ubuntu-latest
+    runs-on: windows-latest
     permissions:
       issues: write
     steps:
+      - uses: actions/checkout@v4
       - uses: ropensci-review-tools/pkgcheck-action@main
+        continue-on-error: true
