### 2025-12-15 ee55b39
Update pkgcheck.yaml
