+++ .github/workflows/pkgcheck.yaml
@@ -15,7 +15,7 @@ on:
 
 jobs:
   pkgcheck:
-    runs-on: windows-latest
+    runs-on: ubuntu-latest
     permissions:
       issues: write
     steps:
