### 2025-12-15 ff0c291
Add test coverage and reduce cyclomatic complexity
- Added comprehensive tests for write.cosinor covering KDE, FGLS methods, multiple tau values, and default parameters
- Added tests for GuessTZ covering fork mode, invalid timezone handling, multiple DT values, and iTZ prioritization
- Refactored boot.seci (cyclocomp 27→~8) by extracting validation, CI computation, and bootstrap helpers
- Refactored GetParam (cyclocomp 23→~5) by extracting parameter extraction functions
- Refactored CosinorM.KDE (cyclocomp 18→~7) by extracting observation KDE, variance, grid KDE, and output builders
- Refactor
[…]
