+++ tests/testthat/test-write.cosinor.R
@@ -133,3 +133,202 @@ test_that("write.cosinor exports PDF and summary CSV correctly", {
   expect_true(all(c("MESOR", "Bathyphase.time", "Trough.ph", "Acrophase.time",
                     "Peak", "Amplitude") %in% names(out)))
 })
+
+
+test_that("write.cosinor handles method='KDE' correctly", {
+  tmpdir <- tempdir()
+  ID <- "TEST_KDE"
+
+  BdfList <-
+    BriefSum(
+      df = FlyEast,
+      SR = 1 / 60,
+      Start = "2017-10-24 13:45:00",
+      TZ = "America/New_York"
+    )
+
+  Bdf <- BdfList$Bdf[5:6, ]
+  df <- BdfList$df
+
+  dfList <-
+    Act2Daily(
+      df = df,
+      Bdf = Bdf,
+      VAct = "Activity",
+      VTm = "Time"
+    )
+
+  # Test KDE method
+  write.cosinor(
+    Dir = tmpdir,
+    ID = ID,
+    DailyAct = dfList$Daily_df,
+    Bdf = Bdf,
+    VAct = "Activity",
+    VTm = "Time",
+    method = "KDE",
+    overwrite = TRUE
+  )
+
+  # ---- Structure checks ----
+  fDir <- file.path(paste0(tmpdir, "/", ID, "/"))
+  expect_true(dir.exists(fDir))
+
+  pdfFile <- file.path(fDir, paste0(ID, ".pdf"))
+  expect_true(file.exists(pdfFile))
+
+  csvFile <- file.path(fDir, "Summary.csv")
+  expect_true(file.exists(csvFile))
+
+  # ---- Content checks ----
+  out <- utils::read.csv(csvFile, stringsAsFactors = FALSE)
+  expect_equal(nrow(out), nrow(Bdf))
+})
+
+
+test_that("write.cosinor handles method='FGLS' correctly", {
+  tmpdir <- tempdir()
+  ID <- "TEST_FGLS"
+
+  BdfList <-
+    BriefSum(
+      df = FlyEast,
+      SR = 1 / 60,
+      Start = "2017-10-24 13:45:00",
+      TZ = "America/New_York"
+    )
+
+  Bdf <- BdfList$Bdf[5:6, ]
+  df <- BdfList$df
+
+  dfList <-
+    Act2Daily(
+      df = df,
+      Bdf = Bdf,
+      VAct = "Activity",
+      VTm = "Time"
+    )
+
+  # Test FGLS method
+  write.cosinor(
+    Dir = tmpdir,
+    ID = ID,
+    DailyAct = dfList$Daily_df,
+    Bdf = Bdf,
+    VAct = "Activity",
+    VTm = "Time",
+    method = "FGLS",
+    overwrite = TRUE
+  )
+
+  # ---- Structure checks ----
+  fDir <- file.path(paste0(tmpdir, "/", ID, "/"))
+  expect_true(dir.exists(fDir))
+
+  pdfFile <- file.path(fDir, paste0(ID, ".pdf"))
+  expect_true(file.exists(pdfFile))
+
+  csvFile <- file.path(fDir, "Summary.csv")
+  expect_true(file.exists(csvFile))
+
+  # ---- Content checks ----
+  out <- utils::read.csv(csvFile, stringsAsFactors = FALSE)
+  expect_equal(nrow(out), nrow(Bdf))
+})
+
+
+test_that("write.cosinor handles multiple tau values", {
+  tmpdir <- tempdir()
+  ID <- "TEST_MULTITAU"
+
+  BdfList <-
+    BriefSum(
+      df = FlyEast,
+      SR = 1 / 60,
[…]
