+++ R/CosinorM.KDE.R
@@ -262,32 +262,84 @@ CosinorM.KDE <- function (time, activity, bw = 0.8, grid = 360L,
     if (!is.integer (grid)) grid <- as.integer (grid)
 
     ## Get Essential Info -----------------
-    ### Assume tau = 24 hours
     tau <- 24
     dt <- diff (time)
     dt <- dt [dt > 0]
     Epc <- 1 / min (dt)
 
+    ## Convert to radians and compute observation-level KDE
+    obs_results <- compute_obs_kde (time, activity, tau, bw)
+    theta <- obs_results$theta
+    y_h <- obs_results$y_h
+    Pdf <- obs_results$Pdf
+    w <- obs_results$w
+    K <- obs_results$K
+    den <- obs_results$den
+    n <- obs_results$n
 
-    ## convert hours to radian in [0,2*pi)
+    ## Compute variance for fitted curve
+    variance_results <- compute_variance (activity, y_h, K, w, den, n)
+    resid <- variance_results$resid
+    df_denom <- variance_results$df_denom
+    sigma_hat <- variance_results$sigma_hat
+    var_fitted <- variance_results$var_fitted
+    se_fitted <- variance_results$se_fitted
+
+
+    ## Evaluate on grid and compute cosinor parameters
+    grid_results <- compute_grid_kde (theta, activity, w, bw, tau,
+                                      grid, sigma_hat)
+    gtheta <- grid_results$gtheta
+    y.g <- grid_results$y.g
+    pdf.g <- grid_results$pdf.g
+    w.g <- grid_results$w.g
+    den.g <- grid_results$den.g
+    var_fitted.g <- grid_results$var_fitted.g
+    se_fitted.g <- grid_results$se_fitted.g
+    area.g <- grid_results$area.g
+
+    ## Compute cosinor parameters
+    cosinor_params <- compute_cosinor_params (pdf.g, gtheta, w.g, den.g,
+                                              area.g, arctan2)
+    mesor <- cosinor_params$mesor
+    beta <- cosinor_params$beta
+    gamma <- cosinor_params$gamma
+    amplitude <- cosinor_params$amplitude
+    acrophase <- cosinor_params$acrophase
+
+    ## Prepare output coefficients
+    coef.cosinor <- prepare_coef_output (mesor, amplitude, acrophase,
+                                         beta, gamma)
+    post.hoc <- compute_posthoc (y_h, Epc)
+
+    ## Generate final output structure
+    fit <- build_kde_output (dilute, coef.cosinor, post.hoc, activity, time,
+                            tau, bw, grid, arctan2, resid, df_denom, theta,
+                            Pdf, w, den, y_h, var_fitted, se_fitted, gtheta,
+                            pdf.g, w.g, den.g, y.g, var_fitted.g, se_fitted.g)
+
+    class (fit) <- c ("CosinorM.KDE")
+    return (fit)
+}
+
+
+## Helper functions for CosinorM.KDE ---------------
+
+#' @title Compute Observation-Level KDE
+#' @noRd
+compute_obs_kde <- function (time, activity, tau, bw) {
     theta <- (time %% tau) * 2 * pi / tau
     n <- length (activity)
-
-    ### Trapzoid weight - radian on observation points
     w <- trap_weights (theta = theta)
 
-    ### Pairwise wrapped angular differences and kernel matrix
-    ### (observations x observations)
-    diffs_mat <- wrap_diff (outer (theta, theta, "-")) # n x n
+    diffs_mat <- wrap_diff (outer (theta, theta, "-"))
     bw_rad <- bw * 2 * pi / tau
[…]
