### 2025-12-15 d5dae6d
Fix code review issues: use file.exists for file checks and improve default parameter handling
- Changed dir.exists() to file.exists() for PDF and CSV file checks in write.cosinor.R
- Fixed default VAct/VTm parameter handling with proper safety checks
- Original code had a bug referencing undefined 'df' variable

Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
