### 2025-12-15 d8eeb92
Optimize default parameter logic for better clarity and efficiency
- Separated VAct and VTm null checks into independent blocks
- Fixed ordering of safety checks (length before access)
- Improved error messages for edge cases

Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
