### 2025-12-15 7b0f615
Update pkgcheck.yaml
