+++ .github/workflows/pkgcheck.yaml
@@ -19,5 +19,4 @@ jobs:
     permissions:
       issues: write
     steps:
-      - uses: actions/checkout@v4
       - uses: ropensci-review-tools/pkgcheck-action@main
