### 2025-12-15 f4fd3be
Remove \donttest tags from example sections in R scripts
Eliminated the \donttest tags from the example code blocks in DST.R and Date2TotalT.R to ensure all examples are included in documentation and testing. This change improves the visibility and test coverage of the provided examples.
