+++ R/Date2TotalT.R
@@ -46,7 +46,7 @@
 #'   the last date), expressed in the units given by `TUnit`.
 #'
 #' @examples
-#' \donttest{
+#' 
 #' # Two calendar days: returns c(24, 24) hours
 #' Date2TotalT (as.Date (c ("2021-01-01", "2021-01-02")), "hour")
 #'
@@ -56,7 +56,7 @@
 #'     "2021-06-10 08:00:00",
 #'     "2021-06-10 14:30:00"
 #' )), "minute")
-#' }
+#' 
 #'
 #' # In seconds (case-insensitive unit name)
 #' Date2TotalT (as.Date ("2022-12-31"), "SeCoNd")
