+++ tests/testthat/test-write.cosinor.R
@@ -138,4 +138,43 @@ test_that ("write.cosinor exports PDF and summary CSV correctly", {
         "MESOR", "Bathyphase.time", "Trough.ph", "Acrophase.time",
         "Peak", "Amplitude"
     ) %in% names (out)))
+
+
+
+
+    # Test FGLS method --------------------------------
+    write.cosinor(
+        Dir = tmpdir,
+        ID = ID,
+        DailyAct = dfList$Daily_df,
+        Bdf = Bdf,
+        VAct = "Activity",
+        VTm = "Time",
+        method = "FGLS",
+        overwrite = TRUE
+    )
+
+    # ---- Structure checks ----
+    fDir <- file.path (paste0 (tmpdir, "/", ID, "/"))
+    expect_true (dir.exists (fDir)) # directory exists
+
+    pdfFile <- file.path (fDir, paste0 (ID, ".pdf"))
+    expect_true (file.exists (pdfFile)) # PDF exists
+
+    csvFile <- file.path (fDir, "Summary.csv")
+    expect_true (file.exists (csvFile)) # CSV exists
+
+    # ---- Relationship checks ----
+    # Verify that the summary CSV relates correctly to the input Bdf
+    out <- utils::read.csv (csvFile, stringsAsFactors = FALSE)
+    expect_equal (nrow (out), nrow (Bdf)) # same number of rows as input subset
+    expect_true (length (out) > length (Bdf))
+
+    # ---- Content checks ----
+    # Ensure the summary CSV contains expected cosinor coefficient columns
+    expect_true (all (c (
+        "MESOR", "Amplitude", "Acrophase",
+        "Acrophase.time"
+    ) %in% names (out)))
+
 })
