+++ R/is.DST.R
@@ -44,14 +44,14 @@
 #' dates <- as.POSIXct (c ("2021-06-15", "2021-12-15"), tz = "America/New_York")
 #' DST (dates, TZ = "America/New_York")
 #'
-#' \donttest{
+#'
 #' # Around the spring-forward transition
 #' trans <- as.Date (c ("2021-03-13", "2021-03-14", "2021-03-15"))
 #' DST (trans, TZ = "America/New_York")
 #'
 #' # Using the local system time zone
 #' DST (as.POSIXct ("2023-07-01 12:00:00"), TZ = "local")
-#' }
+#' 
 #' @keywords daylight dst
 #' @export
 
