### 2025-12-15 92ff136
Fix pkgcheck issues: replace sapply with vapply, fix long lines, replace whole package imports
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
