+++ R/UTCwDST.R
@@ -101,11 +101,13 @@ UTCwDST <- function (UTCs, fork = FALSE) {
     ### but NO for the south
 
     ## Check if any time zone may experience time change.
-    Out <- sapply (pTZs, function (tzs) {
-        wDST <- sapply (tzs, function (tz) as.POSIXlt (wDT, tz = tz)$isdst)
-        sDST <- sapply (tzs, function (tz) as.POSIXlt (sDT, tz = tz)$isdst)
+    Out <- vapply (pTZs, function (tzs) {
+        wDST <- vapply (tzs, function (tz) as.POSIXlt (wDT, tz = tz)$isdst,
+                        integer (1))
+        sDST <- vapply (tzs, function (tz) as.POSIXlt (sDT, tz = tz)$isdst,
+                        integer (1))
         any (wDST != sDST)
-    })
+    }, logical (1))
 
     if (length (Out) == length (UTCs)) {
         names (Out) <- UTCs
