### 2025-12-15 48a09d0
Fix type consistency in GuessTZ.R Step 2
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
