### 2025-12-15 9d07073
Simplify GuessTZ Step 2 logic to handle all cases uniformly
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
