### 2025-12-15 5ba3943
Refactor and clean up circadian analysis code
Removed commented and unused code from Act2Daily.R and BriefSum.R, improved code formatting and readability in Date2TotalT.R, TravelLog.R, and write.cosinor.R, and enhanced code clarity and line breaks in the Circadian-Analysis vignette. Also updated import statements and improved ggplot2 usage for better maintainability.
