+++ R/write.cosinor.R
@@ -28,7 +28,7 @@
 #' also generated to update the original `Bdf` with cosinor model coefficients
 #' for each day.
 #'
-#' @import gridExtra ggplot2 grDevices utils stats
+#' @import gridExtra ggplot2 grDevices utils stats themes
 #'
 #' @param Dir The directory where the recordings to be exported <e.g.
 #' "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">
@@ -138,174 +138,174 @@
 write.cosinor <- function (Dir, ID, DailyAct, Bdf, VAct = NULL, VTm = NULL,
                            method = "OLS", tau = 24, ph = FALSE,
                            overwrite = FALSE) {
-    ## Get Essential Info -----------------
-    nT <- length (tau) ### Number of assumed rhythms
-    D <- names (DailyAct)
-    U <- Bdf$UTC
+  ## Get Essential Info -----------------
+  nT <- length (tau) ### Number of assumed rhythms
+  D <- names (DailyAct)
+  U <- Bdf$UTC
 
-    ### Set default variable names for activity and time columns
-    if (is.null (VAct)) VAct <- names (df) [[2]] # Default: second column of df
-    if (is.null (VTm)) VTm <- names (df) [[1]] # Default: first column of df
+  ### Set default variable names for activity and time columns
+  if (is.null (VAct)) VAct <- names (df) [[2]] # Default: second column of df
+  if (is.null (VTm)) VTm <- names (df) [[1]] # Default: first column of df
 
-    ## Initialize all cosinor results -----------------------
-    if (all (length (tau) == 1, isFALSE (ph))) {
-        nVNames <- c ("MESOR", "Amplitude", "Acrophase", "Acrophase.time")
-        Bdf [nVNames] <- NA
-        KeyTerm <- "MESOR|Amplitude|Acrophase"
-    } else {
-        nVNames <- c (
-            "MESOR", "Bathyphase.time", "Trough.ph", "Acrophase.time",
-            "Peak", "Amplitude"
-        )
-        Bdf [nVNames] <- NA
-        KeyTerm <-
-            "MESOR|Bathyphase.time|Trough.ph|Acrophase.time|Peak|Amplitude"
-    }
+  ## Initialize all cosinor results -----------------------
+  if (all (length (tau) == 1, isFALSE (ph))) {
+    nVNames <- c ("MESOR", "Amplitude", "Acrophase", "Acrophase.time")
+    Bdf [nVNames] <- NA
+    KeyTerm <- "MESOR|Amplitude|Acrophase"
+  } else {
+    nVNames <- c (
+      "MESOR", "Bathyphase.time", "Trough.ph", "Acrophase.time",
+      "Peak", "Amplitude"
+    )
+    Bdf [nVNames] <- NA
+    KeyTerm <-
+      "MESOR|Bathyphase.time|Trough.ph|Acrophase.time|Peak|Amplitude"
+  }
 
-    ## Check Directory ----------------
-    fDir <- paste0 (Dir, "/", ID)
-    if (!dir.exists (fDir)) dir.create (fDir)
-
-
-    ## Create a PDF file for the ID  ----------------
-    pdfDir <- paste0 (fDir, "/", ID, ".pdf")
-    if (isFALSE (overwrite)) {
-        if (dir.exists (pdfDir)) pdfDir <- paste0 (pdfDir, " ", Sys.time ())
-    }
-
-    grDevices::pdf (pdfDir, onefile = TRUE, paper = "a4r")
+  ## Check Directory ----------------
+  fDir <- paste0 (Dir, "/", ID)
[…]

[…]
