+++ R/CosinorM.R
@@ -20,13 +20,12 @@
 #' @description
 #' Fit a cosine-based harmonic linear regression on circular time (hours of day)
 #'
-#'
 #' @details
 #' The `Cosinor` model is a cosine-based harmonic regression used to estimate
 #' circadian rhythm parameters.
 #'
 #' \strong{Single-phase equation}:
-#'       \deqn{y = M + A \cos \left(\frac{2\pi t}{\tau} - \phi\right) }
+#'     \deqn{y = M + A \cos \left(\frac{2\pi t}{\tau} - \phi\right) }
 #' \itemize{
 #'   \item \eqn{M }: MESOR (mid-line estimating statistic of rhythm), the
 #'   intercept
@@ -37,7 +36,6 @@
 #'   cosine coefficients
 #' }
 #'
-#'
 #' Note that because the model is parameterized with a negative phase inside the
 #' cosine, this means the the derived peak time from acrophase corresponds to
 #' the peak time forward from midnight (not backward). In other words, the
@@ -46,7 +44,7 @@
 #'
 #'
 #' \strong{Linearized form}:
-#'      \deqn{\hat{y} = M + \beta x + \gamma z + \epsilon }
+#'    \deqn{\hat{y} = M + \beta x + \gamma z + \epsilon }
 #' \itemize{
 #'   \item \eqn{\beta = A * cos(\phi)}, estimated coefficient for the cosine
 #'   term
@@ -118,13 +116,13 @@
 #'   \item time: The time coordinates of the recording
 #'   \item method: The estimation method used
 #'   \item coef.cosinor: Named numeric vector with entries:
-#'     \itemize{
-#'       \item MESOR: Mid-line estimating statistic of rhythm
-#'       \item Amplitude: Predicted peak activity magnitude
-#'       \item Acrophase: Acrophase in radian (time-of-peak)
-#'       \item Beta: the coefficient of the cosine function
-#'       \item Gamma: the coefficient of the sine function
-#'     }
+#'   \itemize{
+#'     \item MESOR: Mid-line estimating statistic of rhythm
+#'     \item Amplitude: Predicted peak activity magnitude
+#'     \item Acrophase: Acrophase in radian (time-of-peak)
+#'     \item Beta: the coefficient of the cosine function
+#'     \item Gamma: the coefficient of the sine function
+#'   }
 #'   \item post.hoc: Post-hoc peak/trough diagnostics derived from
 #'   fitted.values at observation angles (MESOR.ph, Bathyphase.ph.time,
 #'   Trough.ph, Acrophase.ph.time, Peak.ph, Amplitude.ph)
@@ -166,9 +164,9 @@
 #' # Create quick summary of the recording with adjustment for daylight saving.
 #' BdfList <-
 #'   BriefSum (
-#'     df = FlyEast,
-#'     SR = 1 / 60,
-#'     Start = "2017-10-24 13:45:00"
+#'   df = FlyEast,
+#'   SR = 1 / 60,
+#'   Start = "2017-10-24 13:45:00"
 #'   )
 #'
 #' # Let's extract actigraphy data from a single day
@@ -189,10 +187,10 @@
 #'
 #' # plot Cosinor in hours
 #' plot(fit$time,
-#'      fit$fitted.values,
-#'      type = "l",
-#'      xlab = "Hour",
-#'      ylab = "24-Hour Cosinor Model")
+#'    fit$fitted.values,
+#'    type = "l",
+#'    xlab = "Hour",
+#'    ylab = "24-Hour Cosinor Model")
 #' }
 #'
 #' @keywords cosinor
@@ -200,211 +198,209 @@
 
 
 CosinorM <- function(time, activity, tau, method = "OLS", arctan2 = TRUE,
[…]
