### 2025-12-16 8951002
Update NEWS.md
