+++ README.Rmd
@@ -18,7 +18,8 @@ knitr::opts_chunk$set (
 <!-- badges: start -->
 ![version](https://img.shields.io/badge/version-0.1.9-green)
 [![R-CMD-check](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml/badge.svg?branch=Test)](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml)
-
+[![pkgcheck](https://github.com/cwy20030/ActiGlobe/actions/workflows/pkgcheck.yaml/badge.svg?branch=Test)](https://github.com/cwy20030/ActiGlobe/actions/workflows/pkgcheck.yaml)
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 <!-- badges: end -->
 
 ActiGlobe is designed to make it easy to pre-process longitudinal actigraphy recordings, especially for the recordings affected by time shift due to daylight saving changes and or cross-continental travel. It streamlines the process from end-to-end, to simplify the process of pre-processing, analyzing, and exporting daily actigraphy data and visual reports, making it an essential tool for researchers and professionals in the field. To learn how to use it, check the vignettes for a quick tutorial.
