+++ R/BriefSum.R
@@ -113,7 +113,7 @@ BriefSum <- function (df, SR, Start, TZ = "local") {
     # Prepare Basic Variables ------------------
     # MN <- hms::as_hms("00:00:00")
     # MN2 <- hms::as_hms("24:00:00")
-    TZ <- ifelse (TZ == "local", Sys.timezone (), TZ)
+    if (TZ == "local") TZ <- Sys.timezone ()
     Epc <- 1 / SR # Compute epoch length
     # DP <- nrow(df) # Number of Data Points
     # TT <- DP * Epc # Total time of recordings in seconds
