+++ R/Date2TotalT.R
@@ -63,7 +63,7 @@
 #'
 #' @export
 Date2TotalT <- function (DT, TUnit = "hour", TZ = "local") {
-    TZ <- ifelse (TZ == "local", Sys.timezone (), TZ)
+    if (TZ == "local") TZ <- Sys.timezone ()
 
     # Validate the time zone
     valid_zones <- OlsonNames ()
