+++ man/write.act.Rd
@@ -72,16 +72,16 @@ needed when opening in EXCEL on a computer with non-English EU settings.
 \dontrun{
 
 # Import data
-data(FlyEast)
+data (FlyEast)
 
 
 # Create quick summary of the recording with adjustment for daylight saving.
 BdfList <-
-  BriefSum(
-    df = FlyEast,
-    SR = 1 / 60,
-    Start = "2017-10-24 13:45:00"
-  )
+    BriefSum (
+        df = FlyEast,
+        SR = 1 / 60,
+        Start = "2017-10-24 13:45:00"
+    )
 
 # Let's extract the quick summary of the recording
 Bdf <- BdfList$Bdf
@@ -90,23 +90,23 @@ Bdf <- BdfList$Bdf
 ### shift, the recordings would not be properly segmented from 2017-11-02.
 
 ## To avoid time shift due to travelling, we will keep only the first 8 days.
-Bdf <- Bdf[1:8, ]
+Bdf <- Bdf [1:8, ]
 
 df <- BdfList$df
 # Export daily recordings and the summary report
 
-write.act(
-  Dir = getwd(), ## Export to the current working directory
-  ID = "JD",
-  df = df,
-  Bdf = Bdf,
-  VAct = "Activity",
-  VTm = "Time"
+write.act (
+    Dir = getwd (), ## Export to the current working directory
+    ID = "JD",
+    df = df,
+    Bdf = Bdf,
+    VAct = "Activity",
+    VTm = "Time"
 )
 
 ## To see what was exported copy the path generated by the code below and
 ## paste it as an url in a web browser or the file explorer.
-getwd()
+getwd ()
 }
 
 }
