+++ man/Date2TotalT.Rd
@@ -33,19 +33,19 @@ following calendar day at midnight) and returns these durations in the
 requested time unit.
 }
 \examples{
-\donttest{
+
 # Two calendar days: returns c(24, 24) hours
-Date2TotalT(as.Date(c("2021-01-01", "2021-01-02")), "hour")
+Date2TotalT (as.Date (c ("2021-01-01", "2021-01-02")), "hour")
 
 
 # Working in minutes
-Date2TotalT(as.POSIXct(c(
-  "2021-06-10 08:00:00",
-  "2021-06-10 14:30:00"
+Date2TotalT (as.POSIXct (c (
+    "2021-06-10 08:00:00",
+    "2021-06-10 14:30:00"
 )), "minute")
-}
+
 
 # In seconds (case-insensitive unit name)
-Date2TotalT(as.Date("2022-12-31"), "SeCoNd")
+Date2TotalT (as.Date ("2022-12-31"), "SeCoNd")
 
 }
