+++ man/BriefSum.Rd
@@ -71,17 +71,17 @@ start/end times (HH:MM:SS), any warning labels (e.g. `Travel`,
 epochs for a full day.
 }
 \examples{
-\dontrun{
-# Import data
-data(FlyEast)
 
-BdfList <- BriefSum(
-  df = FlyEast,
-  SR = 1 / 60,
-  Start = "2017-10-19 13:45:00"
+# Import data
+data (FlyEast)
+
+BdfList <- BriefSum (
+    df = FlyEast,
+    SR = 1 / 60,
+    Start = "2017-10-19 13:45:00"
 )
 
-str(BdfList)
+str (BdfList)
 # View(BdfList)
 
 ## install library "zeallot"
@@ -90,7 +90,7 @@ str(BdfList)
 ## BriefSum(df = FlyEast,
 ##          SR = 1/60,
 ##          Start = "2017-10-24 13:45:00")
-}
+
 
 }
 \keyword{actigraphy}
