+++ man/boot.seci.Rd
@@ -41,45 +41,45 @@ and post-hoc parameters via non-parametric bootstrap
 FlyEast
 
 BdfList <-
-  BriefSum(
-    df = FlyEast,
-    SR = 1 / 60,
-    Start = "2017-10-24 13:45:00"
-  )
+    BriefSum (
+        df = FlyEast,
+        SR = 1 / 60,
+        Start = "2017-10-24 13:45:00"
+    )
 
 # Let's extract actigraphy data from a single day
 df <- BdfList$df
-df <- subset(df, df$Date == "2017-10-27")
+df <- subset (df, df$Date == "2017-10-27")
 
 # Multicomponent Cosinor Model
-fit <- CosinorM(
-  time = df$Time,
-  activity = df$Activity,
-  tau = c(12, 24),
-  method = "OLS"
+fit <- CosinorM (
+    time = df$Time,
+    activity = df$Activity,
+    tau = c (12, 24),
+    method = "OLS"
 )
 
 # inspect coefficients
 
-boot.seci(
-  object = fit,
-  ci_level = 0.95,
-  n = 500
+boot.seci (
+    object = fit,
+    ci_level = 0.95,
+    n = 500
 )
 
 
 # Gaussian Kernel Density Estimation
-fit2 <- CosinorM.KDE(
-  time = df$Time,
-  activity = df$Activity
+fit2 <- CosinorM.KDE (
+    time = df$Time,
+    activity = df$Activity
 )
 
 # inspect coefficients
 
-boot.seci(
-  object = fit2,
-  ci_level = 0.95,
-  n = 500
+boot.seci (
+    object = fit2,
+    ci_level = 0.95,
+    n = 500
 )
 }
 

[…]
