+++ vignettes/Time-Shift.Rmd
@@ -31,10 +31,10 @@ knitr::opts_chunk$set(
 
 ## Load the Libraries
 ```{r setup}
-library(ActiGlobe)
+library (ActiGlobe)
 
 ### Optional Library
-# library(zeallot)
+# library (zeallot)
 ```
 If any of the packages has yet been installed, we can always install them using the function ["install.packages()"]. 
 
@@ -43,18 +43,18 @@ If any of the packages has yet been installed, we can always install them using 
 ```{r Load Data, message=FALSE, warning=FALSE, results='hide'}
 data("FlyEast")
 
-head(FlyEast) ### Only the first few lines
+head (FlyEast) ### Only the first few lines
 ```
 
 
 ```{r message=FALSE, warning=FALSE,echo=FALSE, out.width="30%", paged.print=TRUE, tab.cap = "Table 1. An Overview of the FlyEast Dataset"}
-knitr::kable(head(FlyEast))
+knitr::kable (head (FlyEast))
 ```
 
-For more information, type `?FlyEast` or `help(FlyEast)` in your R console.
+For more information, type `?FlyEast` or `help (FlyEast)` in your R console.
 
 < Tip for Beginner to R> 
-1. Type: View(FlyEast) in the R console will allow users to review the data structure of FlyEast.
+1. Type: View (FlyEast) in the R console will allow users to review the data structure of FlyEast.
 2. Use <b>Tab</b> button to help finish the code. Try typing: "Fly" in the Console (i.e., the window where we can see a `>`), then press the button <b>Tab</b> on the keyboard.
 
 
@@ -72,13 +72,13 @@ BdfList <-
   )
 
 
-str(BdfList, max.level = 1) ### An overview of the output structure from BriefSum()
+str (BdfList, max.level = 1) ### An overview of the output structure from BriefSum()
 ```
 
 We can also take the advantage of `zeallot` to store multiple outputs at once just like in `Matlab` and `Python`.
 
 ```{r eval=FALSE}
-c(Bdf, df) %<-%
+c (Bdf, df) %<-%
   BriefSum(
     df = FlyEast,
     SR = 1 / 60,
@@ -90,13 +90,12 @@ In the brief summary of daily recording, we would have thirteen parameters, incl
 
 ```{r message=FALSE, warning=FALSE, results='hide'}
 Bdf <- BdfList$Bdf
-head(Bdf)
+head (Bdf)
 ```
 
 
-```{r echo=FALSE, message=FALSE, warning=FALSE, fig.cap="Table 2 The Header of Bdf", fig.width=8}
-knitr::include_graphics("images/Bdf_head.png")
-
+```{r echo=FALSE, message=FALSE, warning=FALSE, tab.cap="Table 2 The Header of Bdf", fig.width=8}
+knitr::kable (head (Bdf), format = "html", table.attr = "style='width:800px;'")
 ```
 
 In the enriched data - `df`, we have both the original data stored in `FlyEast` and some new information, such as the time stamp of each data point.
@@ -104,11 +103,13 @@ In the enriched data - `df`, we have both the original data stored in `FlyEast` 
 ```{r View df, message=FALSE, warning=FALSE, results='hide'}
 df <- BdfList$df
 
-head(df) ### This should give us the same first few lines of FlyEast dataset with a few new columns created by [BrifSum()].
+head (df) 
+### This should give us the same first few lines of 
+### FlyEast dataset with a few new columns created by [BrifSum()].
 ```
 
 ```{r message=FALSE, warning=FALSE,echo=FALSE, out.width="30%", paged.print=TRUE,tab.cap = "Table 3. BriefSum Enriched Longitudinal Recording"}
[…]
