### 2025-12-16 9eab268
Update TAdjust.R
