+++ R/TZ2UTC.R
@@ -53,7 +53,7 @@
 #' @export
 
 TZ2UTC <- function (DT, TZ = "local") {
-    TZ <- ifelse (TZ == "local", Sys.timezone (), TZ)
+    if (TZ == "local") TZ <- Sys.timezone ()
 
 
     # Validate the time zone
