+++ man/DST.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/is.DST.R
+% Please edit documentation in R/DST.R
 \name{DST}
 \alias{DST}
 \title{Daylight Saving Time}
