+++ reference/BriefSum.md
@@ -101,7 +101,6 @@ c("ActiGlobe","data.frame") on both returned data.frames.
 ## Examples
 
 ``` r
-if (FALSE) { # \dontrun{
 # Import data
 data (FlyEast)
 
@@ -112,6 +111,31 @@ BdfList <- BriefSum (
 )
 
 str (BdfList)
+#> List of 2
+#>  $ Bdf:Classes 'ActiGlobe' and 'data.frame': 35 obs. of  13 variables:
+#>   ..$ Date                   : chr [1:35] "2017-10-19" "2017-10-20" "2017-10-21" "2017-10-22" ...
+#>   ..$ Epoch                  : num [1:35] 60 60 60 60 60 60 60 60 60 60 ...
+#>   ..$ UTC                    : chr [1:35] "UTC-04:00" "UTC-04:00" "UTC-04:00" "UTC-04:00" ...
+#>   ..$ TZ_code                : chr [1:35] "EDT" "EDT" "EDT" "EDT" ...
+#>   ..$ Daylight_Saving        : logi [1:35] TRUE TRUE TRUE TRUE TRUE TRUE ...
+#>   ..$ Recording_Start        : chr [1:35] "13:45:00" "00:00:00" "00:00:00" "00:00:00" ...
+#>   ..$ Recording_End          : chr [1:35] "23:59:00" "23:59:00" "23:59:00" "23:59:00" ...
+#>   ..$ GL_Offset              : num [1:35] 0 0 0 0 0 0 0 0 0 0 ...
+#>   ..$ nDataPoints            : num [1:35] 615 1440 1440 1440 1440 1440 1440 1440 1440 1440 ...
+#>   ..$ Cumulative_Start_Second: num [1:35] 60 36960 123360 209760 296160 ...
+#>   ..$ Cumulative_End_Second  : num [1:35] 36900 123300 209700 296100 382500 ...
+#>   ..$ Excluded               : logi [1:35] TRUE FALSE FALSE FALSE FALSE FALSE ...
+#>   ..$ Warning                : chr [1:35] "Incomplete Recording" "" "" "" ...
+#>  $ df :Classes 'ActiGlobe' and 'data.frame': 48847 obs. of  9 variables:
+#>   ..$ Activity      : int [1:48847] 11 0 0 0 0 0 0 0 0 0 ...
+#>   ..$ X2            : chr [1:48847] "0.00" "0.00" "0.00" "0.00" ...
+#>   ..$ Marker        : int [1:48847] 0 0 0 0 0 0 0 0 0 0 ...
+#>   ..$ DateTime      : POSIXct[1:48847], format: "2017-10-19 13:45:00" "2017-10-19 13:46:00" ...
+#>   ..$ Date          : chr [1:48847] "2017-10-19" "2017-10-19" "2017-10-19" "2017-10-19" ...
+#>   ..$ Time          : chr [1:48847] "13:45:00" "13:46:00" "13:47:00" "13:48:00" ...
+#>   ..$ UTC           : chr [1:48847] "UTC-04:00" "UTC-04:00" "UTC-04:00" "UTC-04:00" ...
+#>   ..$ DaylightSaving: logi [1:48847] TRUE TRUE TRUE TRUE TRUE TRUE ...
+#>   ..$ nPoint        : int [1:48847] 1 2 3 4 5 6 7 8 9 10 ...
 # View(BdfList)
 
 ## install library "zeallot"
@@ -120,5 +144,5 @@ str (BdfList)
 ## BriefSum(df = FlyEast,
 ##          SR = 1/60,
 ##          Start = "2017-10-24 13:45:00")
-} # }
+
 ```
