+++ vignettes/Time-Shift.Rmd
@@ -16,6 +16,25 @@ resource_files:
 ---
 
 
+<style>
+/* container that forces a fixed pixel width and enables scrolling */
+.fullsize-scroll {
+  width: 700px;
+  overflow: auto;
+  -webkit-overflow-scrolling: touch;
+}
+
+/* prevent theme CSS from shrinking the image */
+.fullsize-scroll img {
+  width: 1200px !important;
+  max-width: none !important;
+  height: auto !important;
+  display: block;
+}
+
+</style>
+
+
 ## What will be covered in this tutorial?
 This specific tutorial is designed specifically to show users how to use ActiGlobe to preprocess longitudinal actigrpahy recordings affected by time shift. For analysis and generation of graphic/excel report of daily actigraphy measures, please go to tutorial title: Graphic-Report.
 
@@ -31,10 +50,10 @@ knitr::opts_chunk$set(
 
 ## Load the Libraries
 ```{r setup}
-library (ActiGlobe)
+library(ActiGlobe)
 
 ### Optional Library
-# library (zeallot)
+# library(zeallot)
 ```
 If any of the packages has yet been installed, we can always install them using the function ["install.packages()"]. 
 
@@ -43,18 +62,18 @@ If any of the packages has yet been installed, we can always install them using 
 ```{r Load Data, message=FALSE, warning=FALSE, results='hide'}
 data("FlyEast")
 
-head (FlyEast) ### Only the first few lines
+head(FlyEast) ### Only the first few lines
 ```
 
 
 ```{r message=FALSE, warning=FALSE,echo=FALSE, out.width="30%", paged.print=TRUE, tab.cap = "Table 1. An Overview of the FlyEast Dataset"}
-knitr::kable (head (FlyEast))
+knitr::kable(head(FlyEast))
 ```
 
-For more information, type `?FlyEast` or `help (FlyEast)` in your R console.
+For more information, type `?FlyEast` or `help(FlyEast)` in your R console.
 
 < Tip for Beginner to R> 
-1. Type: View (FlyEast) in the R console will allow users to review the data structure of FlyEast.
+1. Type: View(FlyEast) in the R console will allow users to review the data structure of FlyEast.
 2. Use <b>Tab</b> button to help finish the code. Try typing: "Fly" in the Console (i.e., the window where we can see a `>`), then press the button <b>Tab</b> on the keyboard.
 
 
@@ -72,13 +91,13 @@ BdfList <-
   )
 
 
-str (BdfList, max.level = 1) ### An overview of the output structure from BriefSum()
+str(BdfList, max.level = 1) ### An overview of the output structure from BriefSum()
 ```
 
 We can also take the advantage of `zeallot` to store multiple outputs at once just like in `Matlab` and `Python`.
 
 ```{r eval=FALSE}
-c (Bdf, df) %<-%
+c(Bdf, df) %<-%
   BriefSum(
     df = FlyEast,
     SR = 1 / 60,
@@ -90,26 +109,31 @@ In the brief summary of daily recording, we would have thirteen parameters, incl
 
 ```{r message=FALSE, warning=FALSE, results='hide'}
 Bdf <- BdfList$Bdf
-head (Bdf)
+head(Bdf)
 ```
 
+<style>
+.fullsize-scroll { width:700px; text-align:center; }
+.fullsize-scroll .image-title { display:inline-block; margin:8px 0; font-weight:400; font-size:14px; }
+.fullsize-scroll img { width:700px; max-width:none; display:block; margin:0 auto; }
+</style>
 
-```{r echo=FALSE, message=FALSE, warning=FALSE, tab.cap="Table 2 The Header of Bdf", fig.width=8}
-knitr::kable (head (Bdf), format = "html", table.attr = "style='width:800px;'")
-```
+
+<div class="fullsize-scroll">
[…]
