+++ vignettes/Time-Shift.Rmd
@@ -42,7 +42,7 @@ This specific tutorial is designed specifically to show users how to use ActiGlo
 
 
 ```{r, include = FALSE}
-knitr::opts_chunk$set(
+knitr::opts_chunk$set (
   collapse = TRUE,
   comment = "#>"
 )
@@ -50,66 +50,67 @@ knitr::opts_chunk$set(
 
 ## Load the Libraries
 ```{r setup}
-library(ActiGlobe)
+library (ActiGlobe)
 
 ### Optional Library
-# library(zeallot)
+# library (zeallot)
 ```
-If any of the packages has yet been installed, we can always install them using the function ["install.packages()"]. 
+If any of the packages has yet been installed, we can always install them using the function ["install.packages ()"]. 
 
 
 ## Load Example Data: FlyEast
 ```{r Load Data, message=FALSE, warning=FALSE, results='hide'}
-data("FlyEast")
+data ("FlyEast")
 
-head(FlyEast) ### Only the first few lines
+head (FlyEast) ### Only the first few lines
 ```
 
 
 ```{r message=FALSE, warning=FALSE,echo=FALSE, out.width="30%", paged.print=TRUE, tab.cap = "Table 1. An Overview of the FlyEast Dataset"}
-knitr::kable(head(FlyEast))
+knitr::kable (head (FlyEast))
 ```
 
-For more information, type `?FlyEast` or `help(FlyEast)` in your R console.
+For more information, type `?FlyEast` or `help (FlyEast)` in your R console.
 
 < Tip for Beginner to R> 
-1. Type: View(FlyEast) in the R console will allow users to review the data structure of FlyEast.
-2. Use <b>Tab</b> button to help finish the code. Try typing: "Fly" in the Console (i.e., the window where we can see a `>`), then press the button <b>Tab</b> on the keyboard.
+1. Type: View (FlyEast) in the R console will allow users to review the data structure of FlyEast.
+2. Use <b>Tab</b> button to help finish the code. Try typing: "Fly" in the Console  (i.e., the window where we can see a `>`), then press the button <b>Tab</b> on the keyboard.
 
 
 ## Step 0. Create a brief summary of the recordings
 In general, when pre-processing or harmonizing large volumes of data, it's good practice to first generate a structured summary of the dataset.
 
-Here, we use `BriefSum()` to generate an initial summary of the actigraphy data recorded in FlyEast. Based on the initial recording time, it runs through the raw recording file to provide a quick overview with simple summaries for each recording day and a newly enriched actigraphy data. This pre-processing step also automatically adds timestamps to the original longitudinal recording. If the original location inherits the practice of daylight-saving transitioning, `BriefSum()` would also automatically handle any gain or loss in time. By default, `BriefSum()` would allow us to store both the summary and the newly enriched recording data in a `data.list`.
+Here, we use `BriefSum ()` to generate an initial summary of the actigraphy data recorded in FlyEast. Based on the initial recording time, it runs through the raw recording file to provide a quick overview with simple summaries for each recording day and a newly enriched actigraphy data. This pre-processing step also automatically adds timestamps to the original longitudinal recording. If the original location inherits the practice of daylight-saving transitioning, `BriefSum ()` would also automatically handle any gain or loss in time. By default, `BriefSum ()` would allow us to store both the summary and the newly enriched recording data in a `data.list`.
 
[…]
