+++ vignettes/Time-Shift.Rmd
@@ -68,7 +68,8 @@ BdfList <-
     BriefSum (
         df = FlyEast,
         SR = 1 / 60,
-        Start = "2017-10-24 13:45:00"
+        Start = "2017-10-24 13:45:00",
+        TZ = "America/New_York"
     )
 
 ### An overview of the output structure from BriefSum()
