+++ vignettes/Time-Shift.Rmd
@@ -16,25 +16,6 @@ resource_files:
 ---
 
 
-<style>
-/* container that forces a fixed pixel width and enables scrolling */
-.fullsize-scroll {
-  width: 700px;
-  overflow: auto;
-  -webkit-overflow-scrolling: touch;
-}
-
-/* prevent theme CSS from shrinking the image */
-.fullsize-scroll img {
-  width: 1200px !important;
-  max-width: none !important;
-  height: auto !important;
-  display: block;
-}
-
-</style>
-
-
 ## What will be covered in this tutorial?
 This specific tutorial is designed specifically to show users how to use ActiGlobe to preprocess longitudinal actigrpahy recordings affected by time shift. For analysis and generation of graphic/excel report of daily actigraphy measures, please go to tutorial title: Graphic-Report.
 
@@ -43,8 +24,8 @@ This specific tutorial is designed specifically to show users how to use ActiGlo
 
 ```{r, include = FALSE}
 knitr::opts_chunk$set (
-  collapse = TRUE,
-  comment = "#>"
+    collapse = TRUE,
+    comment = "#>"
 )
 ```
 
@@ -53,9 +34,9 @@ knitr::opts_chunk$set (
 library (ActiGlobe)
 
 ### Optional Library
-# library (zeallot)
+# library(zeallot)
 ```
-If any of the packages has yet been installed, we can always install them using the function ["install.packages ()"]. 
+If any of the packages has yet been installed, we can always install them using the function ["install.packages()"]. 
 
 
 ## Load Example Data: FlyEast
@@ -70,67 +51,62 @@ head (FlyEast) ### Only the first few lines
 knitr::kable (head (FlyEast))
 ```
 
-For more information, type `?FlyEast` or `help (FlyEast)` in your R console.
+For more information, type `?FlyEast` or `help(FlyEast)` in your R console.
 
 < Tip for Beginner to R> 
-1. Type: View (FlyEast) in the R console will allow users to review the data structure of FlyEast.
-2. Use <b>Tab</b> button to help finish the code. Try typing: "Fly" in the Console  (i.e., the window where we can see a `>`), then press the button <b>Tab</b> on the keyboard.
+1. Type: View(FlyEast) in the R console will allow users to review the data structure of FlyEast.
+2. Use <b>Tab</b> button to help finish the code. Try typing: "Fly" in the Console (i.e., the window where we can see a `>`), then press the button <b>Tab</b> on the keyboard.
 
 
 ## Step 0. Create a brief summary of the recordings
 In general, when pre-processing or harmonizing large volumes of data, it's good practice to first generate a structured summary of the dataset.
 
-Here, we use `BriefSum ()` to generate an initial summary of the actigraphy data recorded in FlyEast. Based on the initial recording time, it runs through the raw recording file to provide a quick overview with simple summaries for each recording day and a newly enriched actigraphy data. This pre-processing step also automatically adds timestamps to the original longitudinal recording. If the original location inherits the practice of daylight-saving transitioning, `BriefSum ()` would also automatically handle any gain or loss in time. By default, `BriefSum ()` would allow us to store both the summary and the newly enriched recording data in a `data.list`.
[…]
