+++ articles/Circadian-Analysis.md
@@ -16,7 +16,8 @@ BdfList <-
     BriefSum (
         df = FlyEast,
         SR = 1 / 60,
-        Start = "2017-10-24 13:45:00"
+        Start = "2017-10-24 13:45:00",
+        TZ = "America/New_York"
     )
 
 
