+++ articles/Time-Shift.md
@@ -13,7 +13,7 @@ actigraphy measures, please go to tutorial title: Graphic-Report.
 library (ActiGlobe)
 
 ### Optional Library
-# library (zeallot)
+# library(zeallot)
 ```
 
 If any of the packages has yet been installed, we can always install
@@ -22,7 +22,7 @@ them using the function \[“install.packages()”\].
 ## Load Example Data: FlyEast
 
 ``` r
-data("FlyEast")
+data ("FlyEast")
 
 head (FlyEast) ### Only the first few lines
 ```
@@ -41,10 +41,10 @@ Table 1. An Overview of the FlyEast Dataset
 For more information, type
 [`?FlyEast`](https://cwy20030.github.io/ActiGlobe/reference/FlyEast.md)
 or
-[`help (FlyEast)`](https://cwy20030.github.io/ActiGlobe/reference/FlyEast.md)
+[`help(FlyEast)`](https://cwy20030.github.io/ActiGlobe/reference/FlyEast.md)
 in your R console.
 
-\< Tip for Beginner to R\> 1. Type: View (FlyEast) in the R console will
+\< Tip for Beginner to R\> 1. Type: View(FlyEast) in the R console will
 allow users to review the data structure of FlyEast. 2. Use **Tab**
 button to help finish the code. Try typing: “Fly” in the Console (i.e.,
 the window where we can see a `>`), then press the button **Tab** on the
@@ -73,14 +73,15 @@ recording data in a `data.list`.
 
 ``` r
 BdfList <-
-  BriefSum(
-    df = FlyEast,
-    SR = 1 / 60,
-    Start = "2017-10-24 13:45:00"
-  )
+    BriefSum (
+        df = FlyEast,
+        SR = 1 / 60,
+        Start = "2017-10-24 13:45:00",
+        TZ = "America/New_York"
+    )
 
-
-str (BdfList, max.level = 1) ### An overview of the output structure from BriefSum()
+### An overview of the output structure from BriefSum()
+str (BdfList, max.level = 1)
 #> List of 2
 #>  $ Bdf:Classes 'ActiGlobe' and 'data.frame': 35 obs. of  13 variables:
 #>  $ df :Classes 'ActiGlobe' and 'data.frame': 48847 obs. of  9 variables:
@@ -91,11 +92,11 @@ once just like in `Matlab` and `Python`.
 
 ``` r
 c (Bdf, df) %<-%
-  BriefSum(
-    df = FlyEast,
-    SR = 1 / 60,
-    Start = "2017-10-24 13:45:00"
-  )
+    BriefSum (
+        df = FlyEast,
+        SR = 1 / 60,
+        Start = "2017-10-24 13:45:00"
+    )
 ```
 
 In the brief summary of daily recording, we would have thirteen
@@ -109,14 +110,7 @@ Bdf <- BdfList$Bdf
 head (Bdf)
 ```
 
-| Date | Epoch | UTC | TZ_code | Daylight_Saving | Recording_Start | Recording_End | GL_Offset | nDataPoints | Cumulative_Start_Second | Cumulative_End_Second | Excluded | Warning |
-|:---|---:|:---|:---|:---|:---|:---|---:|---:|---:|---:|:---|:---|
[…]
