+++ vignettes/Time-Shift.Rmd
@@ -16,6 +16,25 @@ resource_files:
 ---
 
 
+<style>
+/* container that forces a fixed pixel width and enables scrolling */
+.fullsize-scroll {
+  width: 700px;
+  overflow: auto;
+  -webkit-overflow-scrolling: touch;
+}
+
+/* prevent theme CSS from shrinking the image */
+.fullsize-scroll img {
+  width: 1200px !important;
+  max-width: none !important;
+  height: auto !important;
+  display: block;
+}
+
+</style>
+
+
 ## What will be covered in this tutorial?
 This specific tutorial is designed specifically to show users how to use ActiGlobe to preprocess longitudinal actigrpahy recordings affected by time shift. For analysis and generation of graphic/excel report of daily actigraphy measures, please go to tutorial title: Graphic-Report.
 
@@ -23,15 +42,15 @@ This specific tutorial is designed specifically to show users how to use ActiGlo
 
 
 ```{r, include = FALSE}
-knitr::opts_chunk$set (
-    collapse = TRUE,
-    comment = "#>"
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
 )
 ```
 
 ## Load the Libraries
 ```{r setup}
-library (ActiGlobe)
+library(ActiGlobe)
 
 ### Optional Library
 # library(zeallot)
@@ -41,14 +60,14 @@ If any of the packages has yet been installed, we can always install them using 
 
 ## Load Example Data: FlyEast
 ```{r Load Data, message=FALSE, warning=FALSE, results='hide'}
-data ("FlyEast")
+data("FlyEast")
 
-head (FlyEast) ### Only the first few lines
+head(FlyEast) ### Only the first few lines
 ```
 
 
 ```{r message=FALSE, warning=FALSE,echo=FALSE, out.width="30%", paged.print=TRUE, tab.cap = "Table 1. An Overview of the FlyEast Dataset"}
-knitr::kable (head (FlyEast))
+knitr::kable(head(FlyEast))
 ```
 
 For more information, type `?FlyEast` or `help(FlyEast)` in your R console.
@@ -65,52 +84,56 @@ Here, we use `BriefSum()` to generate an initial summary of the actigraphy data 
 
 ```{r BriefSum, message=FALSE, warning=FALSE}
 BdfList <-
-    BriefSum (
-        df = FlyEast,
-        SR = 1 / 60,
-        Start = "2017-10-24 13:45:00",
-        TZ = "America/New_York"
-    )
+  BriefSum(
+    df = FlyEast,
+    SR = 1 / 60,
+    Start = "2017-10-24 13:45:00"
+  )
 
-### An overview of the output structure from BriefSum()
-str (BdfList, max.level = 1)
+
+str(BdfList, max.level = 1) ### An overview of the output structure from BriefSum()
 ```
 
 We can also take the advantage of `zeallot` to store multiple outputs at once just like in `Matlab` and `Python`.
 
 ```{r eval=FALSE}
-c (Bdf, df) %<-%
-    BriefSum (
-        df = FlyEast,
-        SR = 1 / 60,
-        Start = "2017-10-24 13:45:00"
-    )
+c(Bdf, df) %<-%
+  BriefSum(
+    df = FlyEast,
+    SR = 1 / 60,
+    Start = "2017-10-24 13:45:00"
+  )
 ```
 
 In the brief summary of daily recording, we would have thirteen parameters, including the information on the beginning and the end of each daily recording period, labels warning the total time of a recording is less than 24 hours and the presence of daylight-saving time (based on the initial time zone).
 
 ```{r message=FALSE, warning=FALSE, results='hide'}
 Bdf <- BdfList$Bdf
-head (Bdf)
[…]
