+++ COMPARISON_REPORT.md
@@ -1,0 +1,110 @@
+# Version Comparison Report
+
+## Comparison Details
+
+**Base Version:** c71974c (Update Time-Shift.Rmd)
+- **Date:** Tue Dec 16 01:10:03 2025 -0500
+- **Author:** Willie (Chun) Yao
+
+**Current Version:** d592e65 (Initial plan)
+- **Date:** Tue Dec 16 15:51:28 2025 +0000
+- **Author:** copilot-swe-agent[bot]
+
+## Summary of Changes
+
+**Total files changed:** 1
+**Total insertions:** 3 lines
+**Total deletions:** 2 lines
+
+---
+
+## Files Changed
+
+### 1. vignettes/Graphic-Report.Rmd
+
+**Status:** Modified (M)
+
+#### Changes:
+- **Line 36:** Modified code chunk options
+  - **Before:** `{r BriefSum, message=FALSE, warning=FALSE, include = FALSE}`
+  - **After:** `{r BriefSum, message=FALSE, warning=FALSE, include=FALSE, eval=FALSE}`
+  - **Change:** Removed space before `=` in `include = FALSE` → `include=FALSE`, and added `eval=FALSE` option
+
+- **Line 41:** Added new parameter to `BriefSum()` function call
+  - **Before:** Function call with only `df`, `SR`, and `Start` parameters
+  - **After:** Added `TZ = "America/New_York"` parameter
+  
+- **Line 42:** Added trailing comma
+  - **Before:** `Start = "2017-10-24 13:45:00"` (no comma at end)
+  - **After:** `Start = "2017-10-24 13:45:00",` (comma added)
+
+#### Full Diff:
+
+```diff
+@@ -33,12 +33,13 @@ library (ActiGlobe)
+ data ("FlyEast")
+ ```
+ 
+-```{r BriefSum, message=FALSE, warning=FALSE, include = FALSE}
++```{r BriefSum, message=FALSE, warning=FALSE, include=FALSE, eval=FALSE}
+ BdfList <-
+     BriefSum (
+         df = FlyEast,
+         SR = 1 / 60,
+-        Start = "2017-10-24 13:45:00"
++        Start = "2017-10-24 13:45:00",
++        TZ = "America/New_York"
+     )
+ ```
+```
+
+---
+
+## Detailed Analysis
+
+### What Changed?
+
+1. **Code Chunk Options Update:**
+   - Standardized spacing in `include=FALSE` (removed space)
+   - Added `eval=FALSE` to prevent code evaluation during vignette build
+
+2. **Function Call Enhancement:**
+   - Added timezone parameter `TZ = "America/New_York"` to the `BriefSum()` function call
+   - This provides explicit timezone context for the temporal data analysis
+
+### Impact:
+
+- **Build/Evaluation:** The code chunk will no longer be evaluated during vignette rendering (`eval=FALSE`)
+- **Functionality:** When manually run, the function now includes timezone information
+- **Code Quality:** Improved parameter alignment with trailing comma for easier future modifications
+
+---
+
+## Commit History Between Versions
+
+The following commits occurred between c71974c and the current version:
+
+1. **c71974c** - Update Time-Shift.Rmd (Base version)
+2. **3741c14** - Update Time-Shift.Rmd
+3. **9eab268** - Update TAdjust.R
+4. **9b844ee** - Update R2P.R
+5. **11f9a99** - Update Time-Shift.Rmd
+6. **f7afd86** - Update Time-Shift.Rmd
+7. **f864c71** - rOpenSci Update 4
+8. **710e233** - Update Graphic-Report.Rmd (This contains the visible changes)
+9. **ee222e0** - Update Time-Shift.Rmd
+10. **ffde2f0** - Refactor BriefSum function call parameters
[…]
