### 2025-12-16 5ff9704
Add quick summary document for version comparison
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
