+++ COMPARISON_SUMMARY.md
@@ -1,0 +1,60 @@
+# Comparison Summary: c71974c vs Current Version
+
+## Quick Summary
+
+This document provides a quick overview of the comparison between commit **c71974c** and the **current version** on the repository.
+
+## Key Findings
+
+- **Only 1 file changed:** `vignettes/Graphic-Report.Rmd`
+- **Changes made:** 3 insertions, 2 deletions (net +1 line)
+- **Type of changes:** Code refactoring and parameter enhancement
+
+## Main Differences
+
+### File: vignettes/Graphic-Report.Rmd
+
+**Three specific changes were made:**
+
+1. **Code chunk option formatting:**
+   - Changed `include = FALSE` to `include=FALSE` (spacing standardization)
+   - Added `eval=FALSE` option to prevent evaluation during build
+
+2. **Added timezone parameter:**
+   - Added `TZ = "America/New_York"` to the `BriefSum()` function call
+   - This provides explicit timezone context for temporal data analysis
+
+3. **Code style improvement:**
+   - Added trailing comma after `Start` parameter
+   - Improves code maintainability
+
+## Full Details
+
+For a complete analysis including:
+- Full diff output
+- Commit history between versions
+- Detailed line-by-line changes
+- Impact analysis
+
+Please see: **[COMPARISON_REPORT.md](./COMPARISON_REPORT.md)**
+
+---
+
+## Commands Used
+
+To reproduce this comparison, you can use:
+
+```bash
+# Compare files changed
+git diff --name-status c71974c..HEAD
+
+# See the full diff
+git diff c71974c..HEAD
+
+# See statistics
+git diff --stat c71974c..HEAD
+```
+
+## Conclusion
+
+The differences between c71974c and the current version are minimal and focused on code quality improvements in the vignette documentation. The main functional change is the addition of an explicit timezone parameter.
