### 2025-12-16 4cb744a
Fix TZ handling robustness and wrap sapply examples in dontrun
Co-authored-by: cwy20030 <35785225+cwy20030@users.noreply.github.com>
