### 2025-12-16 0ace424
Add timezone parameter to BriefSum function
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
