+++ vignettes/Circadian-Analysis.Rmd
@@ -31,7 +31,8 @@ BdfList <-
     BriefSum (
         df = FlyEast,
         SR = 1 / 60,
-        Start = "2017-10-24 13:45:00"
+        Start = "2017-10-24 13:45:00",
+        TZ = "America/New_York"
     )
 
 
