+++ .github/workflows/pkgcheck.yaml
@@ -10,8 +10,6 @@ on:
   workflow_dispatch:
   # Run on every push to main
   push:
-    branches:
-      - Test
 
 jobs:
   pkgcheck:
