+++ .github/workflows/R-CMD-check.yaml
@@ -10,8 +10,9 @@ concurrency:
 # check-standard.yaml is likely a better choice.
 # usethis::use_github_action("check-standard") will install it.
 on:
+  # Manually trigger the Action under Actions/pkgcheck
+  workflow_dispatch:
   push:
-    branches: [Test]
   pull_request:
 
 name: R-CMD-check.yaml
