### 2025-12-16 3f6b976
Add timezone parameter to BriefSum function calls
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
