+++ vignettes/Time-Shift.Rmd
@@ -87,7 +87,8 @@ BdfList <-
   BriefSum(
     df = FlyEast,
     SR = 1 / 60,
-    Start = "2017-10-24 13:45:00"
+    Start = "2017-10-24 13:45:00",
+    TZ = "America/New_York"
   )
 
 
@@ -101,7 +102,8 @@ c(Bdf, df) %<-%
   BriefSum(
     df = FlyEast,
     SR = 1 / 60,
-    Start = "2017-10-24 13:45:00"
+    Start = "2017-10-24 13:45:00",
+    TZ = "America/New_York"
   )
 ```
 
