### 2025-12-16 aa44167
Enable manual trigger for R-CMD-check workflow
Added manual trigger for the GitHub Action.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
