### 2025-12-16 6bee8b3
Enable manual trigger for R-CMD-check workflow
Add manual trigger for workflow dispatch.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
