### 2025-12-16 d5ef607
Remove custom CSS from Time-Shift.Rmd
Removed custom CSS for scrolling and image sizing.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
