### 2025-12-16 #732 @cwy20030
Hi @jhollist,

[![R-CMD-check](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml)
[![pkgcheck](https://github.com/cwy20030/ActiGlobe/actions/workflows/pkgcheck.yaml/badge.svg)](https://github.com/cwy20030/ActiGlobe/actions/workflows/pkgcheck.yaml)
[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)

Thank you for all of your help.

We updated the package as suggested, specifically the testsuites. Because these edits do not affect vignette rendering for the website, we did not increment the version series. We also addressed several `goodpractice` suggestions and greatly reduced the cyclomatic complexity of ggCosinorM (from 72 to 3).

covr output
══════════════════════════════════════════════════════════════════════

# ActiGlobe Coverage – 86.76%

| File              | Lines | Relevant | Covered | Missed | Hits / Line | Coverage |
|-------------------|-------|----------|---------|--------|-------------|----------|
| R/write.cosinor.R | 311   | 95       | 63      | 32     | 7           | 66.32%   |
| R/write.act.R     | 167   | 26       | 20      | 6      | 1           | 76.92%   |
| R/Act2Daily.R     | 327   | 105      | 81      | 24     | 10          | 77.14%   |
| R/ggCosinorM.R    | 742   | 313      | 243     | 70     | 13          | 77.64%   |
| R/GuessTZ.R       | 156   | 37       | 29      | 8      | 332         | 78.38%   |
| R/C2T.R           | 133   | 33       | 26      | 7      | 59929       | 78.79%   |
| R/DateFormat.R    | 160   | 45       | 36      | 9      | 53648       | 80.00%   |
| R/Date2TotalT.R   | 106   | 28       | 23      | 5      | 495         | 82.14%   |
| R/ggActiGlobe.R   | 230   | 74       | 61      | 13     | 1           | 82.43%   |
| R/UTCwDST.R       | 133   | 29       | 24      | 5      | 6           | 82.76%   |
| R/UTC2Num.R       | 64    | 14       | 12      | 2      | 24          | 85.71%   |
| R/CosinorM.KDE.R  | 517   | 150      | 130     | 20     | 77          | 86.67%   |
| R/TravelLog.R     | 72    | 15       | 13      | 2      | 2           | 86.67%   |
| R/Demand.R        | 77    | 15       | 13      | 2      | 3           | 86.67%   |
| R/Rad2Hr.R        | 57    | 9        | 8       | 1      | 9           | 88.89%   |
| R/R2P.R           | 128   | 28       | 25      | 3      | 3           | 89.29%   |
| R/CosinorM.R      | 408   | 97       | 88      | 9      | 73          | 90.72%   |
| R/TZ2UTC.R        | 95    | 22       | 21      | 1      | 1           | 95.45%   |
| R/boot.seci.R     | 216   | 73       | 70      | 3      | 34          | 95.89%   |
| R/TAdjust.R       | 347   | 129      | 125     | 4      | 1           | 96.90%   |
| R/BriefSum.R      | 246   | 67       | 66      | 1      | 16          | 98.51%   |
| R/mIANA.R         | 200   | 121      | 120     | 1      | 15          | 99.17%   |
| R/Prob.Inact.R    | 189   | 45       | 45      | 0      | 18          | 100.00%  |
| R/se.CosinorM.R   | 224   | 43       | 43      | 0      | 22          | 100.00%  |
| R/TimeFormat.R    | 173   | 39       | 39      | 0      | 36589       | 100.00%  |
| R/UnitFactor.R    | 105   | 19       | 19      | 0      | 44          | 100.00%  |
| R/ValInput.R      | 93    | 18       | 18      | 0      | 5078        | 100.00%  |
| R/DST2GL.R        | 84    | 13       | 13      | 0      | 23          | 100.00%  |
| R/DST.R           | 85    | 13       | 13      | 0      | 13          | 100.00%  |
| R/Num2UTC.R       | 66    | 10       | 10      | 0      | 9           | 100.00%  |

══════════════════════════════════════════════════════════════════════

pkgcheck
══════════════════════════════════════════════════════════════════════

 ── ActiGlobe 0.2.1 ─────────────────────────────────────────────────────────
  ✔ Package name is available
  ✔ has a 'codemeta.json' file.
  ✔ has a 'contributing' file.
  ✔ uses 'roxygen2'.
  ✔ 'DESCRIPTION' has a URL field.
  ✔ 'DESCRIPTION' has a BugReports field.
  ✔ Package has at least one HTML vignette
  ✔ All functions have examples.
  ✔ Package has continuous integration checks.
  ✔ Package coverage is 86.7%.
  ✔ R CMD check found no errors.
  ✔ R CMD check found no warnings.
  ℹ Some goodpractice linters failed.
  ℹ Examples should not use `\dontrun` unless really necessary.
  ℹ Current status:
  ✔ GEE - top-notch code.
  ℹ 'pkgcheck' version: 0.1.2.253
  ── git ──
  • HEAD: 1d63f774
  • Default branch: main
  • Number of commits: 181
  • First commit: 03-10-2025
  • Number of authors: 2
  ── Package Structure ──
  ℹ Package uses the following languages:
  • R: 100%
  ℹ Package has
  • 4 authors.
  • 4 vignettes.
  • 4 internal data files.
  • 11 imported packages.
  • 21 exported functions (median 45 lines of code).
  • 83 non-exported functions (median 28 lines of code).
  • 3 parameters per function (median).

══════════════════════════════════════════════════════════════════════

P.S. On an unrelated note, we noticed that an .Rmd file containing inline CSS causes the pkgcheck CI to fail.

Here is the CSS code used.
══════════════════════════════════════════════════════════════════════

# CSS

<style>

.fullsize-scroll {
  width: 700px;
  overflow: auto;
  -webkit-overflow-scrolling: touch;
}

.fullsize-scroll img {
  width: 1200px !important;
  max-width: none !important;
  height: auto !important;
  display: block;
}

</style>

══════════════════════════════════════════════════════════════════════
