+++ README.Rmd
@@ -1,29 +1,29 @@
 ---
 output: github_document
+resource_files:
+  - .vignettes/images/Unadjusted_Overview.png
+  - .vignettes/images/Adjusted_Overview.png
 ---
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-```{r}
-#| echo: false
-knitr::opts_chunk$set (
-    collapse = TRUE,
-    comment = "#>",
-    fig.path = "man/figures/"
-)
-```
-
 # ActiGlobe <a href="https://github.com/cwy20030/ActiGlobe"><img src="man/figures/logo.png" align="right" height="138" alt="ActiGlobe website" /></a>
 
 <!-- badges: start -->
-![version](https://img.shields.io/badge/version-0.1.9-green)
-[![R-CMD-check](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml/badge.svg?branch=Test)](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml)
-[![pkgcheck](https://github.com/cwy20030/ActiGlobe/actions/workflows/pkgcheck.yaml/badge.svg?branch=Test)](https://github.com/cwy20030/ActiGlobe/actions/workflows/pkgcheck.yaml)
+
+![version](https://img.shields.io/badge/version-0.2.1-green)
+[![R-CMD-check](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/cwy20030/ActiGlobe/actions/workflows/R-CMD-check.yaml)
+[![pkgcheck](https://github.com/cwy20030/ActiGlobe/actions/workflows/pkgcheck.yaml/badge.svg)](https://github.com/cwy20030/ActiGlobe/actions/workflows/pkgcheck.yaml)
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 <!-- badges: end -->
 
-ActiGlobe is designed to make it easy to pre-process longitudinal actigraphy recordings, especially for the recordings affected by time shift due to daylight saving changes and or cross-continental travel. It streamlines the process from end-to-end, to simplify the process of pre-processing, analyzing, and exporting daily actigraphy data and visual reports, making it an essential tool for researchers and professionals in the field. To learn how to use it, check the vignettes for a quick tutorial.
-
+ActiGlobe is designed to make it easy to pre-process longitudinal
+actigraphy recordings, especially for the recordings affected by time
+shift due to daylight saving changes and or cross-continental travel. It
+streamlines the process from end-to-end, to simplify the process of
+pre-processing, analyzing, and exporting daily actigraphy data and
+visual reports, making it an essential tool for researchers and
+professionals in the field.
 
 ## Installation
 
@@ -31,53 +31,67 @@ The pre-released version of ActiGlobe can installed from
 [GitHub](https://github.com/) with:
 
 #### ActiGlobe-dev
-```{r eval = FALSE}
-# If devtools is not available locally, please download it by removing the number symbol before the 'install.packages' code.
+
+```{r eval=FALSE}
+# If devtools is not available locally, please download it by removing the number symbol before the 'install.packages' code. 
 # install.packages("devtools")
[…]
