### 2025-12-17 2a8ebf3
Increase image width in README for figures
Updated image width for better visibility in figures.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
