### 2025-12-17 8818577
Add write permissions for contents in R-CMD-check
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
