+++ .github/workflows/R-CMD-check.yaml
@@ -22,9 +22,9 @@ permissions: read-all
 jobs:
   R-CMD-check:
     runs-on: ${{ matrix.config.os }}
-
     name: ${{ matrix.config.os }} (${{ matrix.config.r }})
-
+    permissions:
+      contents: write
     strategy:
       fail-fast: false
       matrix:
