### 2025-12-17 51a7c4b
Delete .github/workflows/stale.yml
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
