### 2025-12-17 e91720e
Update permissions in R-CMD-check workflow
Removed write permission for contents in R-CMD-check job.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
