### 2025-12-17 54bbd16
Fix badge formatting in README.md
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
