+++ articles/Time-Shift.md
@@ -10,7 +10,7 @@ actigraphy measures, please go to tutorial title: Graphic-Report.
 ## Load the Libraries
 
 ``` r
-library (ActiGlobe)
+library(ActiGlobe)
 
 ### Optional Library
 # library(zeallot)
@@ -22,9 +22,9 @@ them using the function \[“install.packages()”\].
 ## Load Example Data: FlyEast
 
 ``` r
-data ("FlyEast")
+data("FlyEast")
 
-head (FlyEast) ### Only the first few lines
+head(FlyEast) ### Only the first few lines
 ```
 
 | Activity | X2   | Marker |
@@ -73,15 +73,15 @@ recording data in a `data.list`.
 
 ``` r
 BdfList <-
-    BriefSum (
-        df = FlyEast,
-        SR = 1 / 60,
-        Start = "2017-10-24 13:45:00",
-        TZ = "America/New_York"
-    )
+  BriefSum(
+    df = FlyEast,
+    SR = 1 / 60,
+    Start = "2017-10-24 13:45:00",
+    TZ = "America/New_York"
+  )
 
-### An overview of the output structure from BriefSum()
-str (BdfList, max.level = 1)
+
+str(BdfList, max.level = 1) ### An overview of the output structure from BriefSum()
 #> List of 2
 #>  $ Bdf:Classes 'ActiGlobe' and 'data.frame': 35 obs. of  13 variables:
 #>  $ df :Classes 'ActiGlobe' and 'data.frame': 48847 obs. of  9 variables:
@@ -91,12 +91,13 @@ We can also take the advantage of `zeallot` to store multiple outputs at
 once just like in `Matlab` and `Python`.
 
 ``` r
-c (Bdf, df) %<-%
-    BriefSum (
-        df = FlyEast,
-        SR = 1 / 60,
-        Start = "2017-10-24 13:45:00"
-    )
+c(Bdf, df) %<-%
+  BriefSum(
+    df = FlyEast,
+    SR = 1 / 60,
+    Start = "2017-10-24 13:45:00",
+    TZ = "America/New_York"
+  )
 ```
 
 In the brief summary of daily recording, we would have thirteen
@@ -107,12 +108,10 @@ recording is less than 24 hours and the presence of daylight-saving time
 
 ``` r
 Bdf <- BdfList$Bdf
-head (Bdf)
+head(Bdf)
 ```
 
-![Table 2 The Header of Bdf](images/Bdf_head.png)
-
-Table 2 The Header of Bdf
+##### Table 2 The Header of Bdf ![Table 2 The Header of Bdf](images/Bdf_head.png)
 
 In the enriched data - `df`, we have both the original data stored in
 `FlyEast` and some new information, such as the time stamp of each data
@@ -121,9 +120,7 @@ point.
 ``` r
 df <- BdfList$df
 
-head (df)
-### This should give us the same first few lines of FlyEast dataset
-### with a few new columns created by [BrifSum()].
+head(df) ### This should give us the same first few lines of FlyEast dataset with a few new columns created by [BrifSum()].
 ```
 
 | Activity | X2 | Marker | DateTime | Date | Time | UTC | DaylightSaving | nPoint |
@@ -164,9 +161,9 @@ or
 in the R console.
 
 ``` r
-data (TLog)
+data(TLog)
 
-head (TLog)
+head(TLog)
 ```
 
 | ID       | UTC_Offset | Country_with_Daylight_Saving | date_Start | date_End |
@@ -197,9 +194,9 @@ timetable in ActiGlobe. We can simply use `View(IANA)` to pull up the
[…]
