+++ pkgdown.yml
@@ -1,4 +1,4 @@
-pandoc: '3.2'
+pandoc: 3.6.3
 pkgdown: 2.2.0
 pkgdown_sha: ~
 articles:
@@ -6,7 +6,7 @@ articles:
   Graphic-Report: Graphic-Report.html
   Introduction: Introduction.html
   Time-Shift: Time-Shift.html
-last_built: 2025-12-16T06:41Z
+last_built: 2025-12-17T20:23Z
 urls:
   reference: https://cwy20030.github.io/ActiGlobe/reference
   article: https://cwy20030.github.io/ActiGlobe/articles
