+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   Graphic-Report: Graphic-Report.html
   Introduction: Introduction.html
   Time-Shift: Time-Shift.html
-last_built: 2025-12-17T20:23Z
+last_built: 2025-12-17T20:41Z
 urls:
   reference: https://cwy20030.github.io/ActiGlobe/reference
   article: https://cwy20030.github.io/ActiGlobe/articles
