### 2025-12-17 7175398
Enable detailed output for pkgcheck action
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
