+++ .github/workflows/pkgcheck.yaml
@@ -18,3 +18,5 @@ jobs:
       issues: write
     steps:
       - uses: ropensci-review-tools/pkgcheck-action@main
+        with:
+          summary-only: false
