### 2025-12-17 a508181
Remove write permission for issues in pkgcheck job
Removed write permission for issues in pkgcheck job.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
