+++ .github/workflows/pkgcheck.yaml
@@ -14,8 +14,6 @@ on:
 jobs:
   pkgcheck:
     runs-on: ubuntu-latest
-    permissions:
-      issues: write
     steps:
       - uses: ropensci-review-tools/pkgcheck-action@main
         with:
