### 2025-12-17 d64b2bb
Add permissions to pkgcheck workflow
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
