+++ .github/workflows/pkgcheck.yaml
@@ -14,6 +14,9 @@ on:
 jobs:
   pkgcheck:
     runs-on: ubuntu-latest
+    permissions:
+      contents: read
+      issues: read 
     steps:
       - uses: ropensci-review-tools/pkgcheck-action@main
         with:
