### 2025-12-27 6fa997a
Add GitHub downloads badge to README
Added GitHub downloads badge to README.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
