### 2025-12-27 bd8c3db
Enhance README with project status and metrics badges
Updated project status badge and added GitHub metrics badges.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
