### 2025-12-27 ee4b1c4
Enhance README with more GitHub badges
Added additional badges for GitHub metrics and downloads.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
