### 2025-12-29 a94cea4
Update badge styles in README.md
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
