### 2025-12-29 8c6ed2d
Revise citation for ActiGlobe package version
Updated citation information for the ActiGlobe package to version 0.2.1.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
