+++ index.md
@@ -17,19 +17,19 @@ The pre-released version of ActiGlobe can installed from
 
 ``` r
 # If devtools is not available locally, please download it by removing the number symbol before the 'install.packages' code. 
-# install.packages("devtools")
-devtools::install_github("cwy20030/ActiGlobe")
+# install.packages ("devtools")
+devtools::install_github ("cwy20030/ActiGlobe")
 
 
 # To properly install tutorial, please, use the following code
-devtools::install_github("cwy20030/ActiGlobe",
-                         build_vignettes = TRUE)
+devtools::install_github ("cwy20030/ActiGlobe",
+                          build_vignettes = TRUE)
 ```
 
 Coming soon… ActiGlobe-release
 
 ``` r
-install.packages("ActiGlobe")
+install.packages ("ActiGlobe")
 ```
 
   
@@ -40,22 +40,23 @@ install.packages("ActiGlobe")
 
 ``` r
 citation("ActiGlobe")
-#> To cite package 'ActiGlobe' in publications use:
-#> 
-#>   Yao C, Varesco G, Simonelli G (2025). _ActiGlobe: Wearable Recording
-#>   Processor for Cross-Continental Travel_. R package version 0.1.9,
-#>   <https://github.com/cwy20030/ActiGlobe>.
-#> 
-#> A BibTeX entry for LaTeX users is
-#> 
-#>   @Manual{,
-#>     title = {ActiGlobe: Wearable Recording Processor for Cross-Continental Travel},
-#>     author = {C. William Yao and Giorgio Varesco and Guido Simonelli},
-#>     year = {2025},
-#>     note = {R package version 0.1.9},
-#>     url = {https://github.com/cwy20030/ActiGlobe},
-#>   }
 ```
+
+#### To cite package ‘ActiGlobe’ in publications use:
+
+Yao C, Varesco G, Simonelli G (2025). *ActiGlobe: Wearable Recording
+Processor for Cross-Continental Travel*. R package version 0.2.1,
+<https://github.com/cwy20030/ActiGlobe>.
+
+#### A BibTeX entry for LaTeX users is
+
+      @Manual{,
+        title = {ActiGlobe: Wearable Recording Processor for Cross-Continental Travel},
+         author = {C. William Yao and Giorgio Varesco and Guido Simonelli},
+         year = {2025},
+         note = {R package version 0.2.1},
+         url = {https://github.com/cwy20030/ActiGlobe},
+       }
 
 To convert to an EndNote-compatible format, paste the BibTeX entry in
 [online bibtex-converter](https://asouqi.github.io/bibtex-converter)
@@ -64,28 +65,28 @@ To convert to an EndNote-compatible format, paste the BibTeX entry in
 
 ``` r
 ### Load ActiGlobe into R
-library(ActiGlobe)
+library (ActiGlobe)
 ```
 
 ``` r
 ### Replace FlyEast with the dataset and specify sampling rate in SR and the start of the recording. 
-BdfList = 
-BriefSum(df = FlyEast,
-         SR = 1/60,
-         Start = "2017-10-24 13:45:00")
+BdfList <- 
+BriefSum (df    = FlyEast,
+          SR    = 1/60,
+          Start = "2017-10-24 13:45:00")
 
 
 ### Extract the summary report and the enriched data
 Bdf <- BdfList$Bdf
-df <- BdfList$df
+df  <- BdfList$df
 ```
 
 ``` r
 ### Quick overview of the original recording
-ggActiGlobe(df = df, 
-            Bdf = Bdf,
-            VAct = "Activity",
-            VDT = "DateTime")
+ggActiGlobe (df   = df, 
[…]
