+++ articles/Time-Shift.md
@@ -10,10 +10,10 @@ actigraphy measures, please go to tutorial title: Graphic-Report.
 ## Load the Libraries
 
 ``` r
-library(ActiGlobe)
+library (ActiGlobe)
 
 ### Optional Library
-# library(zeallot)
+# library (zeallot)
 ```
 
 If any of the packages has yet been installed, we can always install
@@ -24,7 +24,7 @@ them using the function \[“install.packages()”\].
 ``` r
 data("FlyEast")
 
-head(FlyEast) ### Only the first few lines
+head (FlyEast) ### Only the first few lines
 ```
 
 | Activity | X2   | Marker |
@@ -41,10 +41,10 @@ Table 1. An Overview of the FlyEast Dataset
 For more information, type
 [`?FlyEast`](https://cwy20030.github.io/ActiGlobe/reference/FlyEast.md)
 or
-[`help(FlyEast)`](https://cwy20030.github.io/ActiGlobe/reference/FlyEast.md)
+[`help (FlyEast)`](https://cwy20030.github.io/ActiGlobe/reference/FlyEast.md)
 in your R console.
 
-\< Tip for Beginner to R\> 1. Type: View(FlyEast) in the R console will
+\< Tip for Beginner to R\> 1. Type: View (FlyEast) in the R console will
 allow users to review the data structure of FlyEast. 2. Use **Tab**
 button to help finish the code. Try typing: “Fly” in the Console (i.e.,
 the window where we can see a `>`), then press the button **Tab** on the
@@ -76,12 +76,11 @@ BdfList <-
   BriefSum(
     df = FlyEast,
     SR = 1 / 60,
-    Start = "2017-10-24 13:45:00",
-    TZ = "America/New_York"
+    Start = "2017-10-24 13:45:00"
   )
 
 
-str(BdfList, max.level = 1) ### An overview of the output structure from BriefSum()
+str (BdfList, max.level = 1) ### An overview of the output structure from BriefSum()
 #> List of 2
 #>  $ Bdf:Classes 'ActiGlobe' and 'data.frame': 35 obs. of  13 variables:
 #>  $ df :Classes 'ActiGlobe' and 'data.frame': 48847 obs. of  9 variables:
@@ -91,12 +90,11 @@ We can also take the advantage of `zeallot` to store multiple outputs at
 once just like in `Matlab` and `Python`.
 
 ``` r
-c(Bdf, df) %<-%
+c (Bdf, df) %<-%
   BriefSum(
     df = FlyEast,
     SR = 1 / 60,
-    Start = "2017-10-24 13:45:00",
-    TZ = "America/New_York"
+    Start = "2017-10-24 13:45:00"
   )
 ```
 
@@ -108,10 +106,19 @@ recording is less than 24 hours and the presence of daylight-saving time
 
 ``` r
 Bdf <- BdfList$Bdf
-head(Bdf)
+head (Bdf)
 ```
 
-##### Table 2 The Header of Bdf ![Table 2 The Header of Bdf](images/Bdf_head.png)
+| Date | Epoch | UTC | TZ_code | Daylight_Saving | Recording_Start | Recording_End | GL_Offset | nDataPoints | Cumulative_Start_Second | Cumulative_End_Second | Excluded | Warning |
+|:---|---:|:---|:---|:---|:---|:---|---:|---:|---:|---:|:---|:---|
[…]
