### 2025-12-29 30f8183
Fix formatting of article links in index.md
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
