+++ articles/index.md
@@ -2,7 +2,7 @@
 
 ### All vignettes
 
-- [Circadian-Analysis](https://cwy20030.github.io/ActiGlobe/articles/Circadian-Analysis.md):
-- [Graphic-Report](https://cwy20030.github.io/ActiGlobe/articles/Graphic-Report.md):
 - [Introduction](https://cwy20030.github.io/ActiGlobe/articles/Introduction.md):
 - [Time-Shift](https://cwy20030.github.io/ActiGlobe/articles/Time-Shift.md):
+- [Circadian-Analysis](https://cwy20030.github.io/ActiGlobe/articles/Circadian-Analysis.md):
+- [Graphic-Report](https://cwy20030.github.io/ActiGlobe/articles/Graphic-Report.md):
