### 2025-12-29 6e3193d
Revise version headings in NEWS.md
Updated version headings in NEWS.md for clarity.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
