+++ articles/Introduction.md
@@ -9,7 +9,7 @@ the rest of the tutorials in this package.
 ## Load the Libraries
 
 ``` r
-library(ActiGlobe)
+library (ActiGlobe)
 
 ### Optional Library
 # library(zeallot)
@@ -21,7 +21,7 @@ them using the function
 
 ## Introduction
 
-### What is Actigraphy?
+## What is Actigraphy?
 
 Actigraphy is a non-invasive method used for measuring sleep/wake cycles
 by using a small device, usually in the form of a wrist-worn watch. It
@@ -188,6 +188,235 @@ ActiGlobe package. The following vignettes will guide you through:
 We encourage you to work through each tutorial to become familiar with
 the ActiGlobe workflow and discover how this tool can enhance your
 actigraphy research.
+
+------------------------------------------------------------------------
+
+## Getting Started with R: Essential Skills and Pro Tips
+
+R is a powerful, specialized language for data analysis, statistics, and
+visualization. Whether you are a new user or an experienced analyst,
+mastering R’s ecosystem will streamline your workflow—especially when
+working with packages like ActiGlobe.
+
+### Quick Start: Running R
+
+- **Install R:** Download from [CRAN](https://cran.r-project.org/).
+- **Choose an IDE:** [RStudio](https://posit.co/) is the gold standard
+  for R, providing an all-in-one interface for scripting, output,
+  packages, and visualization.
+  - Alternatives: [Visual Studio
+    Code](https://code.visualstudio.com/docs/languages/r) with R
+    extensions, or even Google Colab for running R code via Jupyter
+    interface.
+
+**Try out a basic calculation:**
+
+``` r
+2 + 2
+#> [1] 4
+```
+
+**To reuse the value later, assign it to a variable:**
+
+``` r
+x <- 
+  2 + 2
+
+## We can verify it by printing the stored value in x:
+print(x)
+#> [1] 4
+```
+
+### Why R? Robust and Reliable
+
+R’s CRAN committee reviews every package rigorously for compatibility
+and correctness on all major operating systems. This strict review
+process in unique among common programming languages (such as Python’s
+PyPI and MATLAB) in the world of research. In addition, CRAN
+periodically recheck existing packages compatibility to ensure they
+continue to meet standards as R evolves. While these checks may
+introduce some delays in package updates, they provide a crucial layer
+of quality assurance and avoid version-bounded issues (a common issue
+among python libraries). For this reason, the primary development of
+ActiGlobe is and will be in R to ensure long-term stability and
+reliability for users.
+
+### Essential Keyboard Shortcuts (RStudio and General Use)
+
+Speed up coding with these shortcuts (useful in RStudio):
+
+| Description           | Windows_Linux    | MacOS            |
+|:----------------------|:-----------------|:-----------------|
+| Autocomplete (TAB)    | TAB              | TAB              |
+| Run line/selection    | Ctrl + Enter     | Cmd + Enter      |
+| Pipe `%>%`            | Ctrl + Shift + M | Cmd + Shift + M  |
+| Comment/uncomment     | Ctrl + Shift + C | Cmd + Shift + C  |
+| Find/replace          | Ctrl + F         | Cmd + F          |
+| Select all            | Ctrl + A         | Cmd + A          |
+| Save script           | Ctrl + S         | Cmd + S          |
[…]
