### 2025-12-29 b212375
Reorder articles in pkgdown.yml
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
