+++ pkgdown.yml
@@ -2,10 +2,10 @@ pandoc: 3.6.3
 pkgdown: 2.2.0
 pkgdown_sha: ~
 articles:
-  Circadian-Analysis: Circadian-Analysis.html
-  Graphic-Report: Graphic-Report.html
   Introduction: Introduction.html
   Time-Shift: Time-Shift.html
+  Circadian-Analysis: Circadian-Analysis.html
+  Graphic-Report: Graphic-Report.html
 last_built: 2025-12-29T22:33Z
 urls:
   reference: https://cwy20030.github.io/ActiGlobe/reference
