### 2025-12-30 cc96be2
Update traffic-to-badge.yml
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
