+++ README.md
@@ -1,0 +1,65 @@
+# ⚡️ Traffic badge
+
+**All data is automatically generated by [Traffic to Badge - GitHub Action](https://github.com/marketplace/actions/traffic-to-badge).**
+
+**Example usage repository: [yi-Xu-0100/traffic2badge](https://github.com/yi-Xu-0100/traffic2badge).**
+
+## 📚 [total-traffic-data-badge](https://github.com/cwy20030/ActiGlobe/tree/traffic#readme)
+
+### 📷 Badges without link of [total-traffic-data-badge](https://github.com/cwy20030/ActiGlobe/tree/traffic#readme)
+
+![total views](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_views.svg)
+![total views per week](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_views_per_week.svg)
+![total clones](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_clones.svg)
+![total clones per week](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_clones_per_week.svg)
+
+```markdown
+![total views](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_views.svg)
+![total views per week](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_views_per_week.svg)
+![total clones](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_clones.svg)
+![total clones per week](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_clones_per_week.svg)
+```
+
+### 🔗 Badges with link of [total-traffic-data-badge](https://github.com/cwy20030/ActiGlobe/tree/traffic#readme)
+
+[![total views](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_views.svg)](https://github.com/cwy20030/ActiGlobe/tree/traffic#-total-traffic-data-badge)
+[![total views per week](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_views_per_week.svg)](https://github.com/cwy20030/ActiGlobe/tree/traffic#-total-traffic-data-badge)
+[![total clones](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_clones.svg)](https://github.com/cwy20030/ActiGlobe/tree/traffic#-total-traffic-data-badge)
+[![total clones per week](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_clones_per_week.svg)](https://github.com/cwy20030/ActiGlobe/tree/traffic#-total-traffic-data-badge)
+
+```markdown
+[![total views](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_views.svg)](https://github.com/cwy20030/ActiGlobe/tree/traffic#-total-traffic-data-badge)
+[![total views per week](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_views_per_week.svg)](https://github.com/cwy20030/ActiGlobe/tree/traffic#-total-traffic-data-badge)
+[![total clones](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/total_clones.svg)](https://github.com/cwy20030/ActiGlobe/tree/traffic#-total-traffic-data-badge)
[…]
