### 2025-12-30 2862a85
Change cron schedule to daily at midnight UTC
Updated cron schedule to run daily at midnight UTC.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
