### 2025-12-30 1d6acf3
Change badge colors in traffic-to-badge.yml
Updated color settings for traffic badge workflow.

Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
