+++ README.md
@@ -15,7 +15,7 @@
 <!-- badges: start -->
 ![views](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/traffic-ActiGlobe/views.svg)
 ![GitHub code search count](https://img.shields.io/github/search?query=ActiGlobe)
-![clones per week](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/traffic-ActiGlobe/clones_per_week.svg)
+![clones](https://raw.githubusercontent.com/cwy20030/ActiGlobe/traffic/traffic-ActiGlobe/clones.svg)
 <!-- badges: end -->
 
 ActiGlobe is designed to make it easy to pre-process longitudinal
