### 2026-01-15 04c2286
Fix typo in section header for FGLS model
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
