+++ vignettes/Circadian-Analysis.Rmd
@@ -16,6 +16,9 @@ knitr::opts_chunk$set (
 )
 ```
 
+What will be covered in this tutorial?
+This specific tutorial is designed to show users how to use ActiGlobe to perform circadian analysis via cosinor models or Gaussian kernel density estimation (KDE). For data pre-processing or generation of graphic/excel report of daily actigraphy measures, please go to tutorials titled: Time-Shift and Graphic-Report.
+
 ```{r setup, include = FALSE}
 library (ActiGlobe)
 ```
@@ -77,7 +80,7 @@ If we process `FlyEast` using the `cosinor::cosinor.lm`, we will notice that the
 
 
 
-#### Ecnometry-modified Cosinor Model with Feasible General Least Sqaure (FGLS)
+#### Econometry-modified Cosinor Model with Feasible General Least Square (FGLS)
 Besides, the traditional OLS-based cosinor model,`CosinorM` also allows users to estimate the model using the FGLS-based method. In ActiGlobe, the FGLS was implemented via a weighted residual least square approximation. This two-step process often produces a better representation of the "border" between sleep and wakefulness than traditional OLS. Nevertheless, neither method can realistically capture our activity patterns, given that they are simply single-phase cosinor models.
 ```{r Model 1 FGLS, message=FALSE, warning=FALSE, fig.height=5, fig.width=7, fig.alt=  "Figure 2: Fitting of FGLS-based Cosinor Model"}
 fit.fgls <-
