### 2026-01-20 #732 @robitalec
Hi @cwy20030 - I'm looking forward to helping with this review of ActiGlobe!
Thank you for submitting it for review at rOpenSci.

Thanks to you and @jhollist for your work so far on the editors checks including  increasing the test coverage, adding return values and goodpractice linters.

While I am working through my editor checks, I wanted to check with you @cwy20030 about some things I've noticed so far:

- We see the recommendation in the [R-CMD-check.yaml](https://github.com/cwy20030/ActiGlobe/blob/main/.github/workflows/R-CMD-check.yaml#L9-L11) that the current workflow is likely overkill for most R packages. Would you consider using the standard check action with `usethis::use_github_action('check-standard')`?
- The [README.md](https://github.com/cwy20030/ActiGlobe/blob/main/README.md?plain=1#L2) seems to indicate that it is generated from a README.Rmd, but I'm not seeing one in the repository or in the .gitignore
- In the goodpractice notes, I'm seeing three uses of `sapply()` in the function [GuessTZ](https://github.com/cwy20030/ActiGlobe/blob/main/R/GuessTZ.R). The recommendation to avoid `sapply()` is covered [here](https://adv-r.hadley.nz/functionals.html?q=sapply#map-atomic).
- Is this [.gitignore](https://github.com/cwy20030/ActiGlobe/blob/main/vignettes/.gitignore) intending to ignore [these](https://github.com/cwy20030/ActiGlobe/tree/main/vignettes/images) images?

If possible, to help me and the reviewers, can you try to link any commits or pull requests that address these and future comments on ActiGlobe?

Thanks!
