+++ vignettes/Graphic-Report.Rmd
@@ -58,7 +58,9 @@ data (TLog)
 
 
 ```{r TAdjust, eval=FALSE, message=FALSE, warning=FALSE, include=FALSE}
-Bdf.adj <- TAdjust (Bdf, TLog)
+Bdf.adj <- TAdjust (Bdf = Bdf, 
+                    TLog = TLog,
+                    )
 ```
 
 ## Step 0: Pre-requisits
@@ -107,7 +109,7 @@ for (i in names (dfList$Daily_df) [11:12]) {
         VAct = "Activity",
         VDT = "DateTime"
     ) +
-        scale_color_grey () #### Change the graph colour
+        scale_colour_grey () #### Change the graph colour
 }
 ```
 
