+++ vignettes/Time-Shift.Rmd
@@ -148,7 +148,9 @@ knitr::kable (head (IANA), format = "html", table.attr = "style='width:450px;'")
 Before `ActiGlobe` directly modify information stored in the enriched longitudinal recording - `df`, we will need to adjust the brief summary stored in `Bdf`. This design allows us to quickly scan through the summary file, just in case any adjustment is not properly addressed.
 
 ```{r TAdjust, message=FALSE, warning=FALSE}
-Bdf.adj <- TAdjust (Bdf, TLog)
+Bdf.adj <- TAdjust (Bdf  = Bdf, 
+                    TLog = TLog,
+                    TZ   = "America/New_York")
 ```
 
 

[…]
