### 2026-02-04 28a2897
Bump version, refactor KDE cosinor & docs
Release version 0.2.2 with multiple improvements: refactor CosinorM.KDE to remove dense-grid evaluation (reduce memory use) and switch to observation-angle integrals; update boot.seci and ggCosinorM to no longer depend on dense-grid outputs. Tighten CI matrix and change concurrency grouping in GitHub Actions; add README.Rmd to .gitignore. Change UTCwDST signature to accept an optional DT and pass DT/All/fork to GuessTZ, and improve GuessTZ return formats and matching logic. Apply widespread roxygen/Rd formatting and documentation fixes (ValInput, Date2
[…]
