+++ R/ggCosinorM.R
@@ -35,21 +35,21 @@
 #'
 #' @param object A fitted model of class \code{\link{CosinorM}} or
 #'   \code{\link{CosinorM.KDE}}.
-#' @param labels Logical; Default `TRUE` places repelled labels on
+#' @param labels Logical; Default: TRUE, which places repelled labels on
 #'   the plot with MESOR, amplitude(s), and acrophase(s).
-#' @param ci Logical; Default `TRUE` computes and draws pointwise
+#' @param ci Logical; Default: TRUE, which computes and draws pointwise
 #'   parametric confidence bands for the fitted cosinor curve using
 #'   the model covariance.
-#' @param ci_level Confidence level for the pointwise bands, expressed
-#'   in numeric value between 0 and 1 (default: `0.95`).
-#' @param n Integer; number of points on the fine prediction grid used
-#'   to draw the fitted cosinor and confidence ribbon (default: `400`).
-#' @param point_size Numeric; plotting size for observed points
-#'   (default: `0.5`).
+#' @param ci_level Numeric scaler. The range of the confidence interval,
+#' 	 expressed in numeric value between 0 and 1. Default: 0.95
+#' @param n Integer scaler. Number of points on the fine prediction grid used
+#'   to draw the fitted cosinor and confidence ribbon. Default: 400
+#' @param point_size Numeric scaler. Plotting size for observed points.
+#'   Default: 0.5
 #' @param title_extra Optional character string appended to the plot
 #'   title for extra context.
 #' @param legend.position Position of the legend on the plot; default
-#'   is `"right"`. Other options include `"top"`, `"bottom"`, `"left"`,
+#'   is "right". Other options include "top", "bottom", "left",
 #'   or a numeric vector of length two specifying x and y coordinates.
 #' @param ... Additional arguments (currently ignored) kept for future
 #'   update.
@@ -373,13 +373,7 @@ GetKDEFit <- function (object, t_obs, tau) {
     SeFit <- if (!is.null (Kdf$fitted.se))
       Kdf$fitted.se else rep (NA_real_, length (FitPred))
     list (newt = NewT, fit_pred = FitPred, se_fit = SeFit, uses_kdf = TRUE)
-  } else if (!is.null (object$grid)) {
-    Gtheta <- object$grid$theta
-    NewT <- (Gtheta * tau) / (2 * pi)
-    FitPred <- object$grid$fitted.values
-    SeFit <- object$grid$fitted.se
-    list (newt = NewT, fit_pred = FitPred, se_fit = SeFit, uses_kdf = TRUE)
-  } else {
+   } else {
     NULL
   }
 }
