### 2026-02-04 61610f5
Bump badge to 0.2.2 and clarify NEWS
Update README badge to show version 0.2.2. Clarify NEWS.md entry by specifying CosinorM.KDE (instead of generic KDE) for the Pdf/kIF dense-grid removal, aligning the note with subsequent changes to boot.seci and ggCosinorM.
