+++ NEWS.md
@@ -20,7 +20,7 @@ Revision history for the R/ActiGlobe package
     - remove the use of sapply in GuessTZ to improve functional stability
     - minor Roxygen documentation edits in DST, Date2TotalT, Rad2Hr, TZ2UTC, UTC2Num, ValInput, ggActiGlobe, ggCosinorM, TAdjust
     - improve UTCwDST functional flexibility by allowing users to specify date input through DT (defaults to 2021-01-01)
-    - simplify KDE internals (Pdf/kIF computation) by removing dense-grid evaluation to reduce memory consumption
+    - simplify CosinorM.KDE internals (Pdf/kIF computation) by removing dense-grid evaluation to reduce memory consumption
         - update boot.seci and ggCosinorM to remove support of dense-grid evaluation output from CosinorM.KDE
 
 ## Version 0.2.1, 2025-11-30
