### 2026-02-04 5afccc3
Update NEWS.md
