### 2026-02-04 3bad045
Update NEWS.md
