### 2026-02-16 4caec6b
Add rOpenSci Software Peer Review badge to README
Signed-off-by: C. William Yao <35785225+cwy20030@users.noreply.github.com>
