### 2026-03-02 #732 @vincentvanhees
Hi @cwy20030, I am trying to run the example in the documentation (README and https://cwy20030.github.io/ActiGlobe/).

As `TLog` is derived from a non-existing file path, I assumed I could use the `TLog` object that is part of the package but get `Error in xj[i] : only 0's may be mixed with negative subscripts`. So, I guess it is not compatible with the FlyEast example data.

Could you please clarify whether the demo in the README is supposed to work out of the box?
If not, would it be possible to facilitate this by either adding the real _TravelLog_Template.csv_ to the repository or by updating the `TLog` package data object to match the FlyEast example?
