### 2026-04-15 #732 @cmaimone
I've focused so far on the documentation and a little on the structure of the functions, but I haven't completed a full review of all elements in the template, in part because I ran into some errors along the way.  Sharing what I have so far to help move package development and updates along.  It looks like I ran into some of the same issues as Vincent.

Will the remotes package work for the standard installation?  It has fewer dependencies than devtools for everyday users.

**Quick Start**

The Quick Start guide includes the following lines of code that extract components from a results object.  Is this necessary, or could the longer expressions be used in the example plotting function?  It's typically not recommended to have multiple copies of the same data in the environment, and this code results in the data being saved in BdfList as well as in the Bdf and df objects themselves.

```
### Extract the summary report and the enriched data
Bdf <- BdfList$Bdf
df  <- BdfList$df
```

The Quick Start guide has a figure with caption "Figure 1. Overview of the raw recording with clear day‑to‑day epoching misalignment" - while I haven't worked with this type of data before, I did not find the plot to be "clear" in terms of demonstrating the misalignment.  It could help other users to say a bit more here on what visually in the plot indicates the misalignment.  It would also help to briefly situate what this plot is aimed to do -- is it primarily a diagnostic plot?

I got stuck trying to work through the Quick Start when I got to `TLog <- read.csv ("WHERE/YOU/STOREd/THE/TRAVEL/LOG/TEMPLATE/TravelLog_Template.csv")` because I don't have such a dataset.  Can this be demonstrated with one of the included data files?  Or are there public files somewhere you could point people to?  For example, if you want to demonstrate a CSV file instead of using a built-in dataset, you could include a CSV in a repo and people could read directly from the URL.

This line "Take a coffee break if needed because ActiGlobe will adjust time shift and anonymize the travel destination to keep participants’ privacy." made me think that one of the functions was going to take a long time to run - I'm not sure if that's accurate or not because I wasn't able to run the surrounding code yet, but if that's not what you're trying to indicate, it may be something to reconsider.

Throughout the Quick Start, it would help to have a bit more text (just a sentence or two) introducing what you're doing before showing the code.  The current comments in the code blocks aren't quite enough to provide the narrative of the analysis workflow that is presented.

**Introduction**

If the introduction vignette is intended for someone like me who hasn't worked with this type of data before, it may help to start a step sooner and show and explain what the data looks like - ideally starting with the raw data, and then the sampled/filtered data.  I had a hard time following the Signal Processing section without knowing what the data looks like.

I would consider removing "Getting Started with R: Essential Skills and Pro Tips" - there are other beginner tutorials you could point people to, or if you think this info is sufficient to get someone started (my experience is that it won't be), you could put it elsewhere yourself and point to it.  It's not core to the package.

The advice about .libPaths() is not something I've encountered elsewhere (and I've been using R for a long time).  This might work for your workflow, but it's likely to cause issues for many other users.  People who have used R some may also encounter this and be confused as to whether this is some special requirement for your package.

**Time Shift**

What is the optional package zeallot used for?  I think it appears on other pages too.  I see this line later: "We can also take the advantage of zeallot to store multiple outputs at once just like in Matlab and Python." but it's not demonstrated, and I'm not sure what the goal is of doing this.  Is zeallot enabling the `%<-%` operator?

It looks like `%<-%` is used in the vignette, which means that zeallot shouldn't be commented out at the top of the vignette.  If you truly want to make it optional, then mark the code block that uses this as optional.

Use inline code formatting for R code for things like: "1. Type: View (FlyEast) in the R console" so that it shows clearly as code - it makes it easier for people to identify where the commands start and end.

I didn't follow how the anecdote at the beginning of Step 1 (While users born after 2YK may not be familiar with the concept of a travel diary...) really related to the format of the TLog data.  I also didn't follow the relationship between the included TLog data, the TravelLog() function, and the previous FlyEast data or the Bdf/df output objects.  I get that TravelLog() can make data frames structured like TLog, but I didn't follow why TLog is included directly instead of calculated from other data, or why/how I'd use the TravelLog() function myself.

I understand what the IANA data is, but it's not shown how to actually use it in the workflow.

It would be helpful to give people an idea of how long the code to create Bdf.adj might take to run.  I also was not able to execute that command -- I got an error.

There are 3 loops at the bottom of the vignette that are for making plots, but the output isn't included (I'd expect to see it). I think there might be some syntax errors there that are preventing the output from being included, as I ran into some errors with that code, and there are also some code highlighting inconsistencies.

I found the mixed style of the function names confusing.  Some (write.act, write.cosinor, gg..., boot.seci) use the lower case style with periods, while other function names are capitalized.  Included data sets are also capitalized, and at least one uses an underscore in the name instead of a period (FlyEast_adj).  This inconsistency extends to function argument names as well, with some in all lower case with underscores, others in capitals, others in mixed case, etc.  I'd strongly suggest choosing and applying a standard R style guide throughout the package to ensure uniform decisions and styling.

I'd suggest using more descriptive argument names for some of the functions.  For example, the BriefSum and ggActiGlobe functions in the Quick Start guide have argument names like df, SR, Bdf, VAct, etc.  Unless these come from some standard in the field that I'm not familiar with, it will be easier to read the documentation and then successfully use the code with slightly longer, descriptive names.

#### Additional Documentation Comments

This is largely a style choice, but I'd encourage you to choose a different main font for the documentation site.  The current font is difficult to read in places.  Perhaps keep the unique font for titles/headers and pair it with a more standard one for the text and code examples?

There are multiple places in the documentation (and readme in the repo, and examples in the documentation pages, and in the source code) where there is a space between the function name and the opening parentheses in R code.  This is not standard R code formatting, and it can be confusing for users.

In the readme, there is a code block that includes the code below.  This is an assignment, but it may not be clear to many R users because the second line lacks indentation.  It initially appears like the right hand side of the first line is missing, and that the second line starts a new command.  While spacing does not affect execution, it can help with humans reading code.

```
BdfList <-
BriefSum (df    = FlyEast,
          SR    = 1/60,
          Start = "2017-10-24 13:45:00")
```

In the documentation site, ensure that there aren't comments or lines of code that extend wider than the code blocks so that readers don't have to scroll horizontally.  The second block of the quick start guide is an example to look at, as is one in a section called "Adjust Travel-induced Time Shift".

Small things:

* The FlyEast documentation page doesn't reference the FlyEast_adj page,
* Typo: FlyEast, FlyEast_adj documentation pages: 40 thousands plus
* There's a header in the Introduction vignette not rendering correctly: ### Data Wrangling, Visualization, and Reproducibility Tutorials

---

Estimated hours spent reviewing: 3-4

#### Functionality Comments

Why does BriefSum() generate a datetime column and separate date and time columns in the df output?  These duplicate a lot of information.  Are all of these formats needed regularly in subsequent analyses?  Could other functions extract the needed components from datetime instead of storing duplicates of these values?

BriefSum() seems to conflate two steps, which is why there's a list of objects as output.  I'd consider separating the functionality of creating summary information and of creating the the analysis data frame.  These don't seem to need to go together, and combining them makes the workflow more complicated rather than less.

I'm concerned about the fork argument to TAdjust.  This gets passed to GuessTZ, which isn't an exposed function (or at least doesn't have a documentation page), which says "Parallel computation (`fork = TRUE`) uses all but two cores by default." - but it's not a default if there isn't a way for the user to functionally change this value (it's just what it is).  Using parallel computation on shared systems where you aren't controlling the number of cores used directly is dangerous and can result in violations of system policies.

I think there may be more effective ways than the plot generated by ggActiGlobe() to visualize where there are misalignments between the data and the dates.  The current plot relies on looking at a lot of points with fuzzy boundaries.  Have you explored alternatives that focus on helping people see potential misalignment over plotting all of the values?  One minor tweak that could help the plot scale to longer data streams and be a bit more flexible in terms of the plot size would be to run the dates down the y-axis and time across the y.  Then the plot can get longer (instead of wider) with more days.  Something like a ggridges plot or even a per-day boxplot or violin plot of the values might help someone visualize the shift across days more directly.

---

I appreciate that this package is addressing a very concrete need that you've encountered in working with this data.
