### 2026-04-15 #732 @robitalec
Thank you @cmaimone for your review! I really appreciate your comments on accessibility and providing sufficient description of {ActiGlobe}'s functions for users.
Can you please edit your review to include the number of estimated hours you spent reviewing so far?

There is plenty here for @cwy20030 and coauthors to start working on improvements and an initial response the reviews.
I think we can hold any additional comments from either reviewers until after @cwy20030 have had a chance to address the current reviews. If, however, you have any pressing comments, please feel free to include them here in the thread.

Thanks everyone!
